Huge simplification of library, merge all year values ino single tax scales
