+++ src/Adjustments/July2024/MedicareLevyReduction.php
@@ -7,14 +7,11 @@ namespace ManageIt\PaygTax\Adjustments\July2024;
 use ManageIt\PaygTax\Entities\Payer;
 use ManageIt\PaygTax\Entities\Payee;
 use ManageIt\PaygTax\Entities\Earning;
-use ManageIt\PaygTax\Entities\TaxAdjustment;
 use ManageIt\PaygTax\Entities\TaxScale;
-use ManageIt\PaygTax\TaxScales\July2024\Nat1004Scale2;
-use ManageIt\PaygTax\TaxScales\July2024\Nat1004Scale6;
 use ManageIt\PaygTax\Traits\WeeklyConversion;
 use ManageIt\PaygTax\Utilities\Math;
 
-class MedicareLevyReduction implements TaxAdjustment
+class MedicareLevyReduction
 {
     use WeeklyConversion;
 
@@ -34,30 +31,6 @@ class MedicareLevyReduction implements TaxAdjustment
         $this->children = $children;
     }
 
-    public function isEligible(Payer $payer, Payee $payee, TaxScale $taxScale, Earning $earning): bool
-    {
-        // The payee must be claiming the tax free threshold
-        if (!$payee->claimsTaxFreeThreshold()) {
-            return false;
-        }
-
-        // The payee must be claiming to have a spouse or children
-        if ($this->spouse === false && $this->children === 0) {
-            return false;
-        }
-
-        // If claiming a half Medicare levy exemption, the payee must be claiming children - the
-        // reduction is not applicable to payees who only have a spouse
-        if (
-            $payee->getMedicareLevyExemption() === \ManageIt\PaygTax\Entities\Payee::MEDICARE_LEVY_EXEMPTION_HALF
-            && $this->children === 0
-        ) {
-            return false;
-        }
-
-        return true;
-    }
-
     public function getAdjustmentAmount(Payer $payer, Payee $payee, TaxScale $taxScale, Earning $earning): float
     {
         $weeklyGross = $this->getWeeklyGross(
@@ -65,9 +38,9 @@ class MedicareLevyReduction implements TaxAdjustment
             $earning->getGrossAmount(),
         );
 
-        if ($taxScale instanceof Nat1004Scale2) {
+        if ($payee->getMedicareLevyExemption() === Payee::MEDICARE_LEVY_EXEMPTION_NONE) {
             // Weekly gross must be $500 or higher, but less than the shading out point
-            if ($weeklyGross < 500 || $weeklyGross >= $this->getShadingOutPoint($taxScale)) {
+            if ($weeklyGross < 500 || $weeklyGross >= $this->getShadingOutPoint($payee)) {
                 return 0;
             }
 
@@ -96,9 +69,9 @@ class MedicareLevyReduction implements TaxAdjustment
                     )
                 ) * -1;
             }
-        } elseif ($taxScale instanceof Nat1004Scale6) {
+        } elseif ($payee->getMedicareLevyExemption() === Payee::MEDICARE_LEVY_EXEMPTION_HALF) {
             // Weekly gross must be $843 or higher, but less than the shading out point
-            if ($weeklyGross < 843 || $weeklyGross >= $this->getShadingOutPoint($taxScale)) {
+            if ($weeklyGross < 843 || $weeklyGross >= $this->getShadingOutPoint($payee)) {
                 return 0;
             }
 
@@ -151,9 +124,9 @@ class MedicareLevyReduction implements TaxAdjustment
     /**
      * Calculate the shading out point where the reduction is no longer applicable.
      */
-    protected function getShadingOutPoint(TaxScale $taxScale): float
+    protected function getShadingOutPoint(Payee $payee): float
     {
-        if ($taxScale instanceof Nat1004Scale2) {
+        if ($payee->getMedicareLevyExemption() === Payee::MEDICARE_LEVY_EXEMPTION_NONE) {
             return floor(($this->getWeeklyFamilyThreshold() * 0.1) / 0.08);
         }
