+++ src/Classifiers/BaseClassifier.php
@@ -10,6 +10,10 @@ use ManageIt\PaygTax\Entities\Payee;
 use ManageIt\PaygTax\Entities\Earning;
 use ManageIt\PaygTax\Entities\TaxScale;
 use ManageIt\PaygTax\Exceptions\NoTaxScalesException;
+use ManageIt\PaygTax\TaxScales\Nat1004;
+use ManageIt\PaygTax\TaxScales\Nat3539;
+use ManageIt\PaygTax\TaxScales\Nat4466;
+use ManageIt\PaygTax\TaxScales\Nat75331;
 
 /**
  * Base classifier.
@@ -21,82 +25,26 @@ class BaseClassifier implements Classifier
 {
     public function getTaxScale(Payer $payer, Payee $payee, Earning $earning): TaxScale
     {
-        $scales = $this->availableTaxScales();
-        $applicable = [];
-
-        foreach ($scales as $scale) {
-            if ($scale->isEligible($payer, $payee, $earning)) {
-                $applicable[] = $scale;
-            }
+        // A person without a TFN must always use Nat1004.
+        if ($payee->hasTfnNumber() === false) {
+            return new Nat1004();
         }
 
-        if (count($applicable) === 0) {
-            throw new NoTaxScalesException('Unable to find a tax scale that is applicable for the given scenario.');
+        // Working Holiday Makers.
+        if ($payee->getResidencyStatus() === Payee::WORKING_HOLIDAY_MAKER) {
+            return new Nat75331();
         }
 
-        if (count($applicable) > 1) {
-            throw new NoTaxScalesException('More than one tax scale is applicable for the given scenario.');
+        // People with STSL debts.
+        if ($payee->hasSTSLDebt()) {
+            return new Nat3539();
         }
 
-        return $applicable[0];
-    }
-
-    /**
-     * Gets all available tax scales in the library.
-     *
-     * Returns a set of instances of each tax scale.
-     *
-     * @return TaxScale[]
-     */
-    public function availableTaxScales(): array
-    {
-        $taxScaleDir = realpath(__DIR__ . '/../TaxScales/');
-
-        if ($taxScaleDir === false) {
-            return [];
+        // Australian Residents claiming the Seniors and Pensioners Tax Offset.
+        if ($payee->getResidencyStatus() === Payee::RESIDENT && $payee->getSeniorsOffset() !== Payee::SENIORS_OFFSET_NONE) {
+            return new Nat4466();
         }
 
-        $taxScales = [];
-
-        $files = new \RecursiveIteratorIterator(
-            new \RecursiveDirectoryIterator($taxScaleDir),
-            \RecursiveIteratorIterator::LEAVES_ONLY
-        );
-
-        /** @var \SplFileObject $file */
-        foreach ($files as $file) {
-            if ($file->isFile() && $file->getRealPath() !== false) {
-                $className = $this->convertPathToClassName($file->getRealPath());
-
-                if (is_null($className)) {
-                    continue;
-                }
-
-                try {
-                    $instance = new $className();
-                    if (!$instance instanceof TaxScale) {
-                        continue;
-                    }
-                } catch (\Throwable $e) {
-                    continue;
-                }
-
-                $taxScales[] = $instance;
-            }
-        }
-
-        return $taxScales;
-    }
-
-    protected function convertPathToClassName(string $path): ?string
-    {
-        $taxScaleDir = realpath(__DIR__ . '/../TaxScales/');
-
-        if ($taxScaleDir === false) {
-            return null;
-        }
-
-        return 'ManageIt\\PaygTax\\TaxScales'
-            . str_replace('/', '\\', str_replace('.php', '', str_replace($taxScaleDir, '', $path)));
+        return new Nat1004();
     }
 }
