+++ src/TaxScales/Nat75331.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace ManageIt\PaygTax\TaxScales\October2020;
+namespace ManageIt\PaygTax\TaxScales;
 
 use ManageIt\PaygTax\Entities\Earning;
 use ManageIt\PaygTax\Entities\Payee;
@@ -14,8 +14,6 @@ use ManageIt\PaygTax\Utilities\Math;
 /**
  * Tax scale for Working Holiday Makers.
  *
- * This tax scale has come into effect from 13 October 2020.
- *
  * Note that this scale requires a YTD gross to be provided to correctly calculate tax, as tax percentages are based on
  * the running value.
  */
@@ -24,31 +22,21 @@ class Nat75331 implements TaxScale
     /**
      * {@inheritDoc}
      */
-    public function isEligible(Payer $payer, Payee $payee, Earning $earning): bool
-    {
-        // Only applies after 13 October 2020 and before 1 July 2024.
-        if (!Date::between($earning->getPayDate(), '2020-10-13', '2024-06-30')) {
-            return false;
-        }
-
-        // Only applies to payees whose residency is as a Working Holiday Maker
-        if ($payee->getResidencyStatus() !== \ManageIt\PaygTax\Entities\Payee::WORKING_HOLIDAY_MAKER) {
-            return false;
-        }
-
-        // Only applies if the payer is registered as a Working Holiday Maker employer.
-        if (!$payer->isRegisteredWhmEmployer()) {
-            return false;
-        }
-
-        return true;
-    }
-
-    /**
-     * {@inheritDoc}
-     */
     public function getTaxWithheldAmount(Payer $payer, Payee $payee, Earning $earning): float
     {
+        // WHM taxation changes from 1st July 2024.
+        if (Date::from($earning->getPayDate(), '2024-07-01')) {
+            if ($payee->getYtdGross() <= 45000 && $payee->hasTfnNumber()) {
+                return Math::round($earning->getGrossAmount() * 0.15);
+            }
+            if ($payee->getYtdGross() <= 135000 && $payee->hasTfnNumber()) {
+                return Math::round($earning->getGrossAmount() * 0.3);
+            }
+            if ($payee->getYtdGross() <= 190000 && $payee->hasTfnNumber()) {
+                return Math::round($earning->getGrossAmount() * 0.37);
+            }
+        }
+
         if ($payee->getYtdGross() <= 45000 && $payee->hasTfnNumber()) {
             return Math::round($earning->getGrossAmount() * 0.15);
         }
