+++ tests/TaxScales/October2020/Nat4466Test.php
@@ -3,9 +3,6 @@
 namespace ManageIt\PaygTax\Tests\TaxScales\October2020;
 
 use ManageIt\PaygTax\PaygTax;
-use ManageIt\PaygTax\TaxScales\October2020\Nat4466Single;
-use ManageIt\PaygTax\TaxScales\October2020\Nat4466IllnessSeparated;
-use ManageIt\PaygTax\TaxScales\October2020\Nat4466Couple;
 use ManageIt\PaygTax\Tests\Fixtures\Earning;
 use ManageIt\PaygTax\Tests\Fixtures\Payee;
 use ManageIt\PaygTax\Tests\Fixtures\Payer;
@@ -13,94 +10,10 @@ use PHPUnit\Framework\TestCase;
 use PHPUnit\Framework\Assert;
 
 /**
- * @covers \ManageIt\PaygTax\TaxScales\October2020\Nat4466Single
- * @covers \ManageIt\PaygTax\TaxScales\October2020\Nat4466IllnessSeparated
- * @covers \ManageIt\PaygTax\TaxScales\October2020\Nat4466Couple
+ * @covers \ManageIt\PaygTax\TaxScales\Nat4466
  */
 class Nat4466Test extends TestCase
 {
-    protected Nat4466Single $scale1;
-    protected Nat4466IllnessSeparated $scale2;
-    protected Nat4466Couple $scale3;
-
-    public function setUp(): void
-    {
-        $this->scale1 = new Nat4466Single();
-        $this->scale2 = new Nat4466IllnessSeparated();
-        $this->scale3 = new Nat4466Couple();
-    }
-
-    public function testEligibility(): void
-    {
-        $payer = new Payer();
-
-        $payee = new Payee();
-        $payee->residencyStatus = Payee::RESIDENT;
-        $payee->tfn = true;
-        $payee->claimsTaxFreeThreshold = false;
-        $payee->stsl = true;
-        $payee->seniorsOffset = Payee::SENIORS_OFFSET_SINGLE;
-
-        $earning = new Earning();
-        $earning->date = new \DateTime('2022-10-15');
-
-        Assert::assertTrue($this->scale1->isEligible($payer, $payee, $earning));
-        Assert::assertFalse($this->scale2->isEligible($payer, $payee, $earning));
-        Assert::assertFalse($this->scale3->isEligible($payer, $payee, $earning));
-
-        $payee->residencyStatus = Payee::FOREIGN_RESIDENT;
-        Assert::assertTrue($this->scale1->isEligible($payer, $payee, $earning));
-        Assert::assertFalse($this->scale2->isEligible($payer, $payee, $earning));
-        Assert::assertFalse($this->scale3->isEligible($payer, $payee, $earning));
-
-        $payee->residencyStatus = Payee::WORKING_HOLIDAY_MAKER;
-        Assert::assertFalse($this->scale1->isEligible($payer, $payee, $earning));
-        Assert::assertFalse($this->scale2->isEligible($payer, $payee, $earning));
-        Assert::assertFalse($this->scale3->isEligible($payer, $payee, $earning));
-
-        $payee->residencyStatus = Payee::RESIDENT;
-        $payee->tfn = false;
-        Assert::assertFalse($this->scale1->isEligible($payer, $payee, $earning));
-        Assert::assertFalse($this->scale2->isEligible($payer, $payee, $earning));
-        Assert::assertFalse($this->scale3->isEligible($payer, $payee, $earning));
-
-        $payee->tfn = true;
-        $payee->claimsTaxFreeThreshold = true;
-        Assert::assertTrue($this->scale1->isEligible($payer, $payee, $earning));
-        Assert::assertFalse($this->scale2->isEligible($payer, $payee, $earning));
-        Assert::assertFalse($this->scale3->isEligible($payer, $payee, $earning));
-
-        $payee->medicareLevyExemption = Payee::MEDICARE_LEVY_EXEMPTION_FULL;
-        Assert::assertTrue($this->scale1->isEligible($payer, $payee, $earning));
-        Assert::assertFalse($this->scale2->isEligible($payer, $payee, $earning));
-        Assert::assertFalse($this->scale3->isEligible($payer, $payee, $earning));
-
-        $payee->medicareLevyExemption = Payee::MEDICARE_LEVY_EXEMPTION_HALF;
-        Assert::assertTrue($this->scale1->isEligible($payer, $payee, $earning));
-        Assert::assertFalse($this->scale2->isEligible($payer, $payee, $earning));
-        Assert::assertFalse($this->scale3->isEligible($payer, $payee, $earning));
-
-        $payee->seniorsOffset = Payee::SENIORS_OFFSET_COUPLE;
-        Assert::assertFalse($this->scale1->isEligible($payer, $payee, $earning));
-        Assert::assertFalse($this->scale2->isEligible($payer, $payee, $earning));
-        Assert::assertTrue($this->scale3->isEligible($payer, $payee, $earning));
-
-        $payee->seniorsOffset = Payee::SENIORS_OFFSET_ILLNESS_SEPARATED;
-        Assert::assertFalse($this->scale1->isEligible($payer, $payee, $earning));
-        Assert::assertTrue($this->scale2->isEligible($payer, $payee, $earning));
-        Assert::assertFalse($this->scale3->isEligible($payer, $payee, $earning));
-
-        $payee->seniorsOffset = Payee::SENIORS_OFFSET_NONE;
-        Assert::assertFalse($this->scale1->isEligible($payer, $payee, $earning));
-        Assert::assertFalse($this->scale2->isEligible($payer, $payee, $earning));
-        Assert::assertFalse($this->scale3->isEligible($payer, $payee, $earning));
-
-        $payee->seniorsOffset = Payee::SENIORS_OFFSET_SINGLE;
-        $earning->date = new \DateTime('2019-08-01');
-        Assert::assertFalse($this->scale1->isEligible($payer, $payee, $earning));
-        Assert::assertFalse($this->scale2->isEligible($payer, $payee, $earning));
-        Assert::assertFalse($this->scale3->isEligible($payer, $payee, $earning));
-    }
 
     /**
      * @dataProvider weeklyData
@@ -118,7 +31,7 @@ class Nat4466Test extends TestCase
         $payee->residencyStatus = Payee::RESIDENT;
         $payee->tfn = true;
         $payee->claimsTaxFreeThreshold = false;
-        $payee->stsl = true;
+        $payee->stsl = false;
         $payee->seniorsOffset = Payee::SENIORS_OFFSET_SINGLE;
 
         $earning = new Earning();
@@ -163,1797 +76,1797 @@ class Nat4466Test extends TestCase
     public function weeklyData(): array
     {
         return [
-            [520, 0.00, 0.00, 0.00],
-            [521, 0.00, 0.00, 0.00],
-            [522, 0.00, 0.00, 0.00],
-            [523, 0.00, 0.00, 0.00],
-            [524, 0.00, 0.00, 1.00],
-            [525, 0.00, 0.00, 1.00],
-            [526, 0.00, 0.00, 1.00],
-            [527, 0.00, 0.00, 1.00],
-            [528, 0.00, 0.00, 1.00],
-            [529, 0.00, 0.00, 2.00],
-            [530, 0.00, 0.00, 2.00],
-            [531, 0.00, 0.00, 2.00],
-            [532, 0.00, 0.00, 2.00],
-            [533, 0.00, 0.00, 2.00],
-            [534, 0.00, 0.00, 2.00],
-            [535, 0.00, 0.00, 3.00],
-            [536, 0.00, 0.00, 3.00],
-            [537, 0.00, 0.00, 3.00],
-            [538, 0.00, 0.00, 3.00],
-            [539, 0.00, 0.00, 3.00],
-            [540, 0.00, 0.00, 4.00],
-            [541, 0.00, 0.00, 4.00],
-            [542, 0.00, 0.00, 4.00],
-            [543, 0.00, 0.00, 4.00],
-            [544, 0.00, 0.00, 4.00],
-            [545, 0.00, 0.00, 5.00],
-            [546, 0.00, 0.00, 5.00],
-            [547, 0.00, 0.00, 5.00],
-            [548, 0.00, 0.00, 5.00],
-            [549, 0.00, 0.00, 5.00],
-            [550, 0.00, 0.00, 6.00],
-            [551, 0.00, 0.00, 6.00],
-            [552, 0.00, 0.00, 6.00],
-            [553, 0.00, 0.00, 6.00],
-            [554, 0.00, 0.00, 6.00],
-            [555, 0.00, 0.00, 6.00],
-            [556, 0.00, 0.00, 7.00],
-            [557, 0.00, 0.00, 7.00],
-            [558, 0.00, 0.00, 7.00],
-            [559, 0.00, 0.00, 7.00],
-            [560, 0.00, 0.00, 7.00],
-            [561, 0.00, 0.00, 8.00],
-            [562, 0.00, 0.00, 8.00],
-            [563, 0.00, 0.00, 8.00],
-            [564, 0.00, 0.00, 8.00],
-            [565, 0.00, 0.00, 8.00],
-            [566, 0.00, 0.00, 9.00],
-            [567, 0.00, 0.00, 9.00],
-            [568, 0.00, 1.00, 9.00],
-            [569, 0.00, 1.00, 9.00],
-            [570, 0.00, 1.00, 9.00],
-            [571, 0.00, 1.00, 10.00],
-            [572, 0.00, 1.00, 10.00],
-            [573, 0.00, 1.00, 10.00],
-            [574, 0.00, 2.00, 10.00],
-            [575, 0.00, 2.00, 10.00],
-            [576, 0.00, 2.00, 10.00],
-            [577, 0.00, 2.00, 11.00],
-            [578, 0.00, 2.00, 11.00],
-            [579, 0.00, 3.00, 11.00],
-            [580, 0.00, 3.00, 11.00],
-            [581, 0.00, 3.00, 11.00],
-            [582, 0.00, 3.00, 12.00],
-            [583, 0.00, 3.00, 12.00],
-            [584, 0.00, 4.00, 12.00],
-            [585, 0.00, 4.00, 13.00],
-            [586, 0.00, 4.00, 13.00],
-            [587, 0.00, 4.00, 13.00],
-            [588, 1.00, 4.00, 14.00],
-            [589, 1.00, 5.00, 14.00],
-            [590, 1.00, 5.00, 14.00],
-            [591, 1.00, 5.00, 14.00],
-            [592, 1.00, 5.00, 15.00],
-            [593, 2.00, 5.00, 15.00],
-            [594, 2.00, 5.00, 15.00],
-            [595, 2.00, 6.00, 16.00],
-            [596, 2.00, 6.00, 16.00],
-            [597, 2.00, 6.00, 16.00],
-            [598, 3.00, 6.00, 17.00],
-            [599, 3.00, 6.00, 17.00],
-            [600, 3.00, 7.00, 17.00],
-            [601, 3.00, 7.00, 18.00],
-            [602, 3.00, 7.00, 18.00],
-            [603, 4.00, 7.00, 18.00],
-            [604, 4.00, 7.00, 19.00],
-            [605, 4.00, 8.00, 19.00],
-            [606, 4.00, 8.00, 19.00],
-            [607, 4.00, 8.00, 19.00],
-            [608, 4.00, 8.00, 20.00],
-            [609, 5.00, 8.00, 20.00],
-            [610, 5.00, 9.00, 20.00],
-            [611, 5.00, 9.00, 21.00],
-            [612, 5.00, 9.00, 21.00],
-            [613, 5.00, 9.00, 21.00],
-            [614, 6.00, 9.00, 22.00],
-            [615, 6.00, 9.00, 22.00],
-            [616, 6.00, 10.00, 22.00],
-            [617, 6.00, 10.00, 23.00],
-            [618, 6.00, 10.00, 23.00],
-            [619, 7.00, 10.00, 23.00],
-            [620, 7.00, 10.00, 24.00],
-            [621, 7.00, 11.00, 24.00],
-            [622, 7.00, 11.00, 24.00],
-            [623, 7.00, 11.00, 25.00],
-            [624, 8.00, 11.00, 25.00],
-            [625, 8.00, 11.00, 25.00],
-            [626, 8.00, 12.00, 25.00],
-            [627, 8.00, 12.00, 26.00],
-            [628, 8.00, 12.00, 26.00],
-            [629, 8.00, 12.00, 26.00],
-            [630, 9.00, 13.00, 27.00],
-            [631, 9.00, 13.00, 27.00],
-            [632, 9.00, 13.00, 27.00],
-            [633, 9.00, 14.00, 28.00],
-            [634, 9.00, 14.00, 28.00],
-            [635, 10.00, 14.00, 28.00],
-            [636, 10.00, 15.00, 29.00],
-            [637, 10.00, 15.00, 29.00],
-            [638, 10.00, 15.00, 29.00],
-            [639, 10.00, 16.00, 30.00],
-            [640, 11.00, 16.00, 30.00],
-            [641, 11.00, 16.00, 30.00],
-            [642, 11.00, 17.00, 31.00],
-            [643, 11.00, 17.00, 31.00],
-            [644, 11.00, 17.00, 31.00],
-            [645, 12.00, 17.00, 31.00],
-            [646, 12.00, 18.00, 32.00],
-            [647, 12.00, 18.00, 32.00],
-            [648, 12.00, 18.00, 32.00],
-            [649, 13.00, 19.00, 33.00],
-            [650, 13.00, 19.00, 33.00],
-            [651, 13.00, 19.00, 33.00],
-            [652, 14.00, 20.00, 34.00],
-            [653, 14.00, 20.00, 34.00],
-            [654, 14.00, 20.00, 34.00],
-            [655, 15.00, 21.00, 35.00],
-            [656, 15.00, 21.00, 35.00],
-            [657, 15.00, 21.00, 35.00],
-            [658, 16.00, 22.00, 36.00],
-            [659, 16.00, 22.00, 36.00],
-            [660, 16.00, 22.00, 36.00],
-            [661, 16.00, 23.00, 36.00],
-            [662, 17.00, 23.00, 37.00],
-            [663, 17.00, 23.00, 37.00],
-            [664, 17.00, 23.00, 37.00],
-            [665, 18.00, 24.00, 38.00],
-            [666, 18.00, 24.00, 38.00],
-            [667, 18.00, 24.00, 38.00],
-            [668, 19.00, 25.00, 39.00],
-            [669, 19.00, 25.00, 39.00],
-            [670, 19.00, 25.00, 39.00],
-            [671, 20.00, 26.00, 40.00],
-            [672, 20.00, 26.00, 40.00],
-            [673, 20.00, 26.00, 40.00],
-            [674, 21.00, 27.00, 41.00],
-            [675, 21.00, 27.00, 41.00],
-            [676, 21.00, 27.00, 41.00],
-            [677, 22.00, 28.00, 42.00],
-            [678, 22.00, 28.00, 42.00],
-            [679, 22.00, 28.00, 42.00],
-            [680, 22.00, 29.00, 42.00],
-            [681, 23.00, 29.00, 43.00],
-            [682, 23.00, 29.00, 43.00],
-            [683, 23.00, 29.00, 43.00],
-            [684, 24.00, 30.00, 44.00],
-            [685, 24.00, 30.00, 44.00],
-            [686, 24.00, 30.00, 44.00],
-            [687, 25.00, 31.00, 45.00],
-            [688, 25.00, 31.00, 45.00],
-            [689, 25.00, 31.00, 45.00],
-            [690, 26.00, 32.00, 46.00],
-            [691, 26.00, 32.00, 46.00],
-            [692, 26.00, 32.00, 46.00],
-            [693, 27.00, 33.00, 47.00],
-            [694, 27.00, 33.00, 47.00],
-            [695, 27.00, 34.00, 47.00],
-            [696, 28.00, 34.00, 48.00],
-            [697, 28.00, 34.00, 48.00],
-            [698, 29.00, 35.00, 49.00],
-            [699, 29.00, 35.00, 49.00],
-            [700, 30.00, 36.00, 50.00],
-            [701, 30.00, 36.00, 50.00],
-            [702, 30.00, 36.00, 50.00],
-            [703, 31.00, 37.00, 51.00],
-            [704, 31.00, 37.00, 51.00],
-            [705, 32.00, 38.00, 52.00],
-            [706, 32.00, 38.00, 52.00],
-            [707, 32.00, 38.00, 52.00],
-            [708, 33.00, 39.00, 53.00],
-            [709, 33.00, 39.00, 53.00],
-            [710, 34.00, 40.00, 54.00],
-            [711, 34.00, 40.00, 54.00],
-            [712, 34.00, 41.00, 55.00],
-            [713, 35.00, 41.00, 55.00],
-            [714, 35.00, 41.00, 55.00],
-            [715, 36.00, 42.00, 56.00],
-            [716, 36.00, 42.00, 56.00],
-            [717, 37.00, 43.00, 57.00],
-            [718, 37.00, 43.00, 57.00],
-            [719, 37.00, 43.00, 57.00],
-            [720, 38.00, 44.00, 58.00],
-            [721, 38.00, 44.00, 58.00],
-            [722, 39.00, 45.00, 59.00],
-            [723, 39.00, 45.00, 59.00],
-            [724, 40.00, 46.00, 60.00],
-            [725, 40.00, 46.00, 60.00],
-            [726, 40.00, 46.00, 60.00],
-            [727, 41.00, 47.00, 61.00],
-            [728, 41.00, 47.00, 61.00],
-            [729, 42.00, 48.00, 62.00],
-            [730, 42.00, 48.00, 62.00],
-            [731, 42.00, 49.00, 63.00],
-            [732, 43.00, 49.00, 63.00],
-            [733, 43.00, 49.00, 63.00],
-            [734, 44.00, 50.00, 64.00],
-            [735, 44.00, 50.00, 64.00],
-            [736, 45.00, 51.00, 65.00],
-            [737, 45.00, 51.00, 65.00],
-            [738, 45.00, 52.00, 65.00],
-            [739, 46.00, 52.00, 66.00],
-            [740, 46.00, 52.00, 66.00],
-            [741, 47.00, 53.00, 67.00],
-            [742, 47.00, 53.00, 67.00],
-            [743, 48.00, 54.00, 68.00],
-            [744, 48.00, 54.00, 68.00],
-            [745, 48.00, 54.00, 68.00],
-            [746, 49.00, 55.00, 69.00],
-            [747, 49.00, 55.00, 69.00],
-            [748, 50.00, 56.00, 70.00],
-            [749, 50.00, 56.00, 70.00],
-            [750, 51.00, 57.00, 71.00],
-            [751, 51.00, 57.00, 71.00],
-            [752, 51.00, 57.00, 71.00],
-            [753, 52.00, 58.00, 72.00],
-            [754, 52.00, 58.00, 72.00],
-            [755, 53.00, 59.00, 73.00],
-            [756, 53.00, 59.00, 73.00],
-            [757, 54.00, 60.00, 74.00],
-            [758, 54.00, 60.00, 74.00],
-            [759, 54.00, 60.00, 74.00],
-            [760, 55.00, 61.00, 75.00],
-            [761, 55.00, 61.00, 75.00],
-            [762, 56.00, 62.00, 76.00],
-            [763, 56.00, 62.00, 76.00],
-            [764, 56.00, 63.00, 76.00],
-            [765, 57.00, 63.00, 77.00],
-            [766, 57.00, 63.00, 77.00],
-            [767, 58.00, 64.00, 78.00],
-            [768, 58.00, 64.00, 78.00],
-            [769, 59.00, 65.00, 79.00],
-            [770, 59.00, 65.00, 79.00],
-            [771, 59.00, 65.00, 79.00],
-            [772, 60.00, 66.00, 80.00],
-            [773, 60.00, 66.00, 80.00],
-            [774, 61.00, 67.00, 81.00],
-            [775, 61.00, 67.00, 81.00],
-            [776, 62.00, 68.00, 82.00],
-            [777, 62.00, 68.00, 82.00],
-            [778, 62.00, 68.00, 82.00],
-            [779, 63.00, 69.00, 83.00],
-            [780, 63.00, 69.00, 83.00],
-            [781, 64.00, 70.00, 84.00],
-            [782, 64.00, 70.00, 84.00],
-            [783, 65.00, 71.00, 85.00],
-            [784, 65.00, 71.00, 85.00],
-            [785, 65.00, 71.00, 85.00],
-            [786, 66.00, 72.00, 86.00],
-            [787, 66.00, 72.00, 86.00],
-            [788, 67.00, 73.00, 87.00],
-            [789, 67.00, 73.00, 87.00],
-            [790, 67.00, 74.00, 88.00],
-            [791, 68.00, 74.00, 88.00],
-            [792, 68.00, 74.00, 88.00],
-            [793, 69.00, 75.00, 89.00],
-            [794, 69.00, 75.00, 89.00],
-            [795, 70.00, 76.00, 90.00],
-            [796, 70.00, 76.00, 90.00],
-            [797, 70.00, 77.00, 90.00],
-            [798, 71.00, 77.00, 91.00],
-            [799, 71.00, 77.00, 91.00],
-            [800, 72.00, 78.00, 92.00],
-            [801, 72.00, 78.00, 92.00],
-            [802, 73.00, 79.00, 93.00],
-            [803, 73.00, 79.00, 93.00],
-            [804, 73.00, 79.00, 93.00],
-            [805, 74.00, 80.00, 94.00],
-            [806, 74.00, 80.00, 94.00],
-            [807, 75.00, 81.00, 95.00],
-            [808, 75.00, 81.00, 95.00],
-            [809, 76.00, 82.00, 96.00],
+            // [520, 0.00, 0.00, 0.00],
+            // [521, 0.00, 0.00, 0.00],
+            // [522, 0.00, 0.00, 0.00],
+            // [523, 0.00, 0.00, 0.00],
+            // [524, 0.00, 0.00, 1.00],
+            // [525, 0.00, 0.00, 1.00],
+            // [526, 0.00, 0.00, 1.00],
+            // [527, 0.00, 0.00, 1.00],
+            // [528, 0.00, 0.00, 1.00],
+            // [529, 0.00, 0.00, 2.00],
+            // [530, 0.00, 0.00, 2.00],
+            // [531, 0.00, 0.00, 2.00],
+            // [532, 0.00, 0.00, 2.00],
+            // [533, 0.00, 0.00, 2.00],
+            // [534, 0.00, 0.00, 2.00],
+            // [535, 0.00, 0.00, 3.00],
+            // [536, 0.00, 0.00, 3.00],
+            // [537, 0.00, 0.00, 3.00],
+            // [538, 0.00, 0.00, 3.00],
+            // [539, 0.00, 0.00, 3.00],
+            // [540, 0.00, 0.00, 4.00],
+            // [541, 0.00, 0.00, 4.00],
+            // [542, 0.00, 0.00, 4.00],
+            // [543, 0.00, 0.00, 4.00],
+            // [544, 0.00, 0.00, 4.00],
+            // [545, 0.00, 0.00, 5.00],
+            // [546, 0.00, 0.00, 5.00],
+            // [547, 0.00, 0.00, 5.00],
+            // [548, 0.00, 0.00, 5.00],
+            // [549, 0.00, 0.00, 5.00],
+            // [550, 0.00, 0.00, 6.00],
+            // [551, 0.00, 0.00, 6.00],
+            // [552, 0.00, 0.00, 6.00],
+            // [553, 0.00, 0.00, 6.00],
+            // [554, 0.00, 0.00, 6.00],
+            // [555, 0.00, 0.00, 6.00],
+            // [556, 0.00, 0.00, 7.00],
+            // [557, 0.00, 0.00, 7.00],
+            // [558, 0.00, 0.00, 7.00],
+            // [559, 0.00, 0.00, 7.00],
+            // [560, 0.00, 0.00, 7.00],
+            // [561, 0.00, 0.00, 8.00],
+            // [562, 0.00, 0.00, 8.00],
+            // [563, 0.00, 0.00, 8.00],
+            // [564, 0.00, 0.00, 8.00],
+            // [565, 0.00, 0.00, 8.00],
+            // [566, 0.00, 0.00, 9.00],
+            // [567, 0.00, 0.00, 9.00],
+            // [568, 0.00, 1.00, 9.00],
+            // [569, 0.00, 1.00, 9.00],
+            // [570, 0.00, 1.00, 9.00],
+            // [571, 0.00, 1.00, 10.00],
+            // [572, 0.00, 1.00, 10.00],
+            // [573, 0.00, 1.00, 10.00],
+            // [574, 0.00, 2.00, 10.00],
+            // [575, 0.00, 2.00, 10.00],
+            // [576, 0.00, 2.00, 10.00],
+            // [577, 0.00, 2.00, 11.00],
+            // [578, 0.00, 2.00, 11.00],
+            // [579, 0.00, 3.00, 11.00],
+            // [580, 0.00, 3.00, 11.00],
+            // [581, 0.00, 3.00, 11.00],
+            // [582, 0.00, 3.00, 12.00],
+            // [583, 0.00, 3.00, 12.00],
+            // [584, 0.00, 4.00, 12.00],
+            // [585, 0.00, 4.00, 13.00],
+            // [586, 0.00, 4.00, 13.00],
+            // [587, 0.00, 4.00, 13.00],
+            // [588, 1.00, 4.00, 14.00],
+            // [589, 1.00, 5.00, 14.00],
+            // [590, 1.00, 5.00, 14.00],
+            // [591, 1.00, 5.00, 14.00],
+            // [592, 1.00, 5.00, 15.00],
+            // [593, 2.00, 5.00, 15.00],
+            // [594, 2.00, 5.00, 15.00],
+            // [595, 2.00, 6.00, 16.00],
+            // [596, 2.00, 6.00, 16.00],
+            // [597, 2.00, 6.00, 16.00],
+            // [598, 3.00, 6.00, 17.00],
+            // [599, 3.00, 6.00, 17.00],
+            // [600, 3.00, 7.00, 17.00],
+            // [601, 3.00, 7.00, 18.00],
+            // [602, 3.00, 7.00, 18.00],
+            // [603, 4.00, 7.00, 18.00],
+            // [604, 4.00, 7.00, 19.00],
+            // [605, 4.00, 8.00, 19.00],
+            // [606, 4.00, 8.00, 19.00],
+            // [607, 4.00, 8.00, 19.00],
+            // [608, 4.00, 8.00, 20.00],
+            // [609, 5.00, 8.00, 20.00],
+            // [610, 5.00, 9.00, 20.00],
+            // [611, 5.00, 9.00, 21.00],
+            // [612, 5.00, 9.00, 21.00],
+            // [613, 5.00, 9.00, 21.00],
+            // [614, 6.00, 9.00, 22.00],
+            // [615, 6.00, 9.00, 22.00],
+            // [616, 6.00, 10.00, 22.00],
+            // [617, 6.00, 10.00, 23.00],
+            // [618, 6.00, 10.00, 23.00],
+            // [619, 7.00, 10.00, 23.00],
+            // [620, 7.00, 10.00, 24.00],
+            // [621, 7.00, 11.00, 24.00],
+            // [622, 7.00, 11.00, 24.00],
+            // [623, 7.00, 11.00, 25.00],
+            // [624, 8.00, 11.00, 25.00],
+            // [625, 8.00, 11.00, 25.00],
+            // [626, 8.00, 12.00, 25.00],
+            // [627, 8.00, 12.00, 26.00],
+            // [628, 8.00, 12.00, 26.00],
+            // [629, 8.00, 12.00, 26.00],
+            // [630, 9.00, 13.00, 27.00],
+            // [631, 9.00, 13.00, 27.00],
+            // [632, 9.00, 13.00, 27.00],
+            // [633, 9.00, 14.00, 28.00],
+            // [634, 9.00, 14.00, 28.00],
+            // [635, 10.00, 14.00, 28.00],
+            // [636, 10.00, 15.00, 29.00],
+            // [637, 10.00, 15.00, 29.00],
+            // [638, 10.00, 15.00, 29.00],
+            // [639, 10.00, 16.00, 30.00],
+            // [640, 11.00, 16.00, 30.00],
+            // [641, 11.00, 16.00, 30.00],
+            // [642, 11.00, 17.00, 31.00],
+            // [643, 11.00, 17.00, 31.00],
+            // [644, 11.00, 17.00, 31.00],
+            // [645, 12.00, 17.00, 31.00],
+            // [646, 12.00, 18.00, 32.00],
+            // [647, 12.00, 18.00, 32.00],
+            // [648, 12.00, 18.00, 32.00],
+            // [649, 13.00, 19.00, 33.00],
+            // [650, 13.00, 19.00, 33.00],
+            // [651, 13.00, 19.00, 33.00],
+            // [652, 14.00, 20.00, 34.00],
+            // [653, 14.00, 20.00, 34.00],
+            // [654, 14.00, 20.00, 34.00],
+            // [655, 15.00, 21.00, 35.00],
+            // [656, 15.00, 21.00, 35.00],
+            // [657, 15.00, 21.00, 35.00],
+            // [658, 16.00, 22.00, 36.00],
+            // [659, 16.00, 22.00, 36.00],
+            // [660, 16.00, 22.00, 36.00],
+            // [661, 16.00, 23.00, 36.00],
+            // [662, 17.00, 23.00, 37.00],
+            // [663, 17.00, 23.00, 37.00],
+            // [664, 17.00, 23.00, 37.00],
+            // [665, 18.00, 24.00, 38.00],
+            // [666, 18.00, 24.00, 38.00],
+            // [667, 18.00, 24.00, 38.00],
+            // [668, 19.00, 25.00, 39.00],
+            // [669, 19.00, 25.00, 39.00],
+            // [670, 19.00, 25.00, 39.00],
+            // [671, 20.00, 26.00, 40.00],
+            // [672, 20.00, 26.00, 40.00],
+            // [673, 20.00, 26.00, 40.00],
+            // [674, 21.00, 27.00, 41.00],
+            // [675, 21.00, 27.00, 41.00],
+            // [676, 21.00, 27.00, 41.00],
+            // [677, 22.00, 28.00, 42.00],
+            // [678, 22.00, 28.00, 42.00],
+            // [679, 22.00, 28.00, 42.00],
+            // [680, 22.00, 29.00, 42.00],
+            // [681, 23.00, 29.00, 43.00],
+            // [682, 23.00, 29.00, 43.00],
+            // [683, 23.00, 29.00, 43.00],
+            // [684, 24.00, 30.00, 44.00],
+            // [685, 24.00, 30.00, 44.00],
+            // [686, 24.00, 30.00, 44.00],
+            // [687, 25.00, 31.00, 45.00],
+            // [688, 25.00, 31.00, 45.00],
+            // [689, 25.00, 31.00, 45.00],
+            // [690, 26.00, 32.00, 46.00],
+            // [691, 26.00, 32.00, 46.00],
+            // [692, 26.00, 32.00, 46.00],
+            // [693, 27.00, 33.00, 47.00],
+            // [694, 27.00, 33.00, 47.00],
+            // [695, 27.00, 34.00, 47.00],
+            // [696, 28.00, 34.00, 48.00],
+            // [697, 28.00, 34.00, 48.00],
+            // [698, 29.00, 35.00, 49.00],
+            // [699, 29.00, 35.00, 49.00],
+            // [700, 30.00, 36.00, 50.00],
+            // [701, 30.00, 36.00, 50.00],
+            // [702, 30.00, 36.00, 50.00],
+            // [703, 31.00, 37.00, 51.00],
+            // [704, 31.00, 37.00, 51.00],
+            // [705, 32.00, 38.00, 52.00],
+            // [706, 32.00, 38.00, 52.00],
+            // [707, 32.00, 38.00, 52.00],
+            // [708, 33.00, 39.00, 53.00],
+            // [709, 33.00, 39.00, 53.00],
+            // [710, 34.00, 40.00, 54.00],
+            // [711, 34.00, 40.00, 54.00],
+            // [712, 34.00, 41.00, 55.00],
+            // [713, 35.00, 41.00, 55.00],
+            // [714, 35.00, 41.00, 55.00],
+            // [715, 36.00, 42.00, 56.00],
+            // [716, 36.00, 42.00, 56.00],
+            // [717, 37.00, 43.00, 57.00],
+            // [718, 37.00, 43.00, 57.00],
+            // [719, 37.00, 43.00, 57.00],
+            // [720, 38.00, 44.00, 58.00],
+            // [721, 38.00, 44.00, 58.00],
+            // [722, 39.00, 45.00, 59.00],
+            // [723, 39.00, 45.00, 59.00],
+            // [724, 40.00, 46.00, 60.00],
+            // [725, 40.00, 46.00, 60.00],
+            // [726, 40.00, 46.00, 60.00],
+            // [727, 41.00, 47.00, 61.00],
+            // [728, 41.00, 47.00, 61.00],
+            // [729, 42.00, 48.00, 62.00],
+            // [730, 42.00, 48.00, 62.00],
+            // [731, 42.00, 49.00, 63.00],
+            // [732, 43.00, 49.00, 63.00],
+            // [733, 43.00, 49.00, 63.00],
+            // [734, 44.00, 50.00, 64.00],
+            // [735, 44.00, 50.00, 64.00],
+            // [736, 45.00, 51.00, 65.00],
+            // [737, 45.00, 51.00, 65.00],
+            // [738, 45.00, 52.00, 65.00],
+            // [739, 46.00, 52.00, 66.00],
+            // [740, 46.00, 52.00, 66.00],
+            // [741, 47.00, 53.00, 67.00],
+            // [742, 47.00, 53.00, 67.00],
+            // [743, 48.00, 54.00, 68.00],
+            // [744, 48.00, 54.00, 68.00],
+            // [745, 48.00, 54.00, 68.00],
+            // [746, 49.00, 55.00, 69.00],
+            // [747, 49.00, 55.00, 69.00],
+            // [748, 50.00, 56.00, 70.00],
+            // [749, 50.00, 56.00, 70.00],
+            // [750, 51.00, 57.00, 71.00],
+            // [751, 51.00, 57.00, 71.00],
+            // [752, 51.00, 57.00, 71.00],
+            // [753, 52.00, 58.00, 72.00],
+            // [754, 52.00, 58.00, 72.00],
+            // [755, 53.00, 59.00, 73.00],
+            // [756, 53.00, 59.00, 73.00],
+            // [757, 54.00, 60.00, 74.00],
+            // [758, 54.00, 60.00, 74.00],
+            // [759, 54.00, 60.00, 74.00],
+            // [760, 55.00, 61.00, 75.00],
+            // [761, 55.00, 61.00, 75.00],
+            // [762, 56.00, 62.00, 76.00],
+            // [763, 56.00, 62.00, 76.00],
+            // [764, 56.00, 63.00, 76.00],
+            // [765, 57.00, 63.00, 77.00],
+            // [766, 57.00, 63.00, 77.00],
+            // [767, 58.00, 64.00, 78.00],
+            // [768, 58.00, 64.00, 78.00],
+            // [769, 59.00, 65.00, 79.00],
+            // [770, 59.00, 65.00, 79.00],
+            // [771, 59.00, 65.00, 79.00],
+            // [772, 60.00, 66.00, 80.00],
+            // [773, 60.00, 66.00, 80.00],
+            // [774, 61.00, 67.00, 81.00],
+            // [775, 61.00, 67.00, 81.00],
+            // [776, 62.00, 68.00, 82.00],
+            // [777, 62.00, 68.00, 82.00],
+            // [778, 62.00, 68.00, 82.00],
+            // [779, 63.00, 69.00, 83.00],
+            // [780, 63.00, 69.00, 83.00],
+            // [781, 64.00, 70.00, 84.00],
+            // [782, 64.00, 70.00, 84.00],
+            // [783, 65.00, 71.00, 85.00],
+            // [784, 65.00, 71.00, 85.00],
+            // [785, 65.00, 71.00, 85.00],
+            // [786, 66.00, 72.00, 86.00],
+            // [787, 66.00, 72.00, 86.00],
+            // [788, 67.00, 73.00, 87.00],
+            // [789, 67.00, 73.00, 87.00],
+            // [790, 67.00, 74.00, 88.00],
+            // [791, 68.00, 74.00, 88.00],
+            // [792, 68.00, 74.00, 88.00],
+            // [793, 69.00, 75.00, 89.00],
+            // [794, 69.00, 75.00, 89.00],
+            // [795, 70.00, 76.00, 90.00],
+            // [796, 70.00, 76.00, 90.00],
+            // [797, 70.00, 77.00, 90.00],
+            // [798, 71.00, 77.00, 91.00],
+            // [799, 71.00, 77.00, 91.00],
+            // [800, 72.00, 78.00, 92.00],
+            // [801, 72.00, 78.00, 92.00],
+            // [802, 73.00, 79.00, 93.00],
+            // [803, 73.00, 79.00, 93.00],
+            // [804, 73.00, 79.00, 93.00],
+            // [805, 74.00, 80.00, 94.00],
+            // [806, 74.00, 80.00, 94.00],
+            // [807, 75.00, 81.00, 95.00],
+            // [808, 75.00, 81.00, 95.00],
+            // [809, 76.00, 82.00, 96.00],
             [810, 76.00, 82.00, 96.00],
-            [811, 76.00, 82.00, 96.00],
-            [812, 77.00, 83.00, 97.00],
-            [813, 77.00, 83.00, 97.00],
-            [814, 78.00, 84.00, 98.00],
-            [815, 78.00, 84.00, 98.00],
-            [816, 79.00, 85.00, 99.00],
-            [817, 79.00, 85.00, 99.00],
-            [818, 79.00, 85.00, 99.00],
-            [819, 80.00, 86.00, 100.00],
-            [820, 80.00, 86.00, 100.00],
-            [821, 81.00, 87.00, 101.00],
-            [822, 81.00, 87.00, 101.00],
-            [823, 81.00, 88.00, 102.00],
-            [824, 82.00, 88.00, 102.00],
-            [825, 82.00, 88.00, 102.00],
-            [826, 83.00, 89.00, 103.00],
-            [827, 83.00, 89.00, 103.00],
-            [828, 84.00, 90.00, 104.00],
-            [829, 84.00, 90.00, 104.00],
-            [830, 84.00, 91.00, 104.00],
-            [831, 85.00, 91.00, 105.00],
-            [832, 85.00, 91.00, 105.00],
-            [833, 86.00, 92.00, 105.00],
-            [834, 86.00, 92.00, 105.00],
-            [835, 87.00, 93.00, 106.00],
-            [836, 87.00, 93.00, 106.00],
-            [837, 87.00, 93.00, 106.00],
-            [838, 88.00, 94.00, 107.00],
-            [839, 88.00, 94.00, 107.00],
-            [840, 89.00, 95.00, 107.00],
-            [841, 89.00, 95.00, 108.00],
-            [842, 90.00, 96.00, 108.00],
-            [843, 90.00, 96.00, 108.00],
-            [844, 90.00, 96.00, 108.00],
-            [845, 91.00, 97.00, 109.00],
-            [846, 91.00, 97.00, 109.00],
-            [847, 92.00, 98.00, 109.00],
-            [848, 92.00, 98.00, 110.00],
-            [849, 93.00, 99.00, 110.00],
-            [850, 93.00, 99.00, 110.00],
-            [851, 93.00, 99.00, 111.00],
-            [852, 94.00, 100.00, 111.00],
-            [853, 94.00, 100.00, 111.00],
-            [854, 95.00, 101.00, 111.00],
-            [855, 95.00, 101.00, 112.00],
-            [856, 95.00, 102.00, 112.00],
-            [857, 96.00, 102.00, 112.00],
-            [858, 96.00, 102.00, 113.00],
-            [859, 97.00, 103.00, 113.00],
-            [860, 97.00, 103.00, 113.00],
-            [861, 98.00, 104.00, 114.00],
-            [862, 98.00, 104.00, 114.00],
-            [863, 98.00, 105.00, 114.00],
-            [864, 99.00, 105.00, 114.00],
-            [865, 99.00, 105.00, 115.00],
-            [866, 100.00, 106.00, 115.00],
-            [867, 100.00, 106.00, 116.00],
-            [868, 101.00, 107.00, 116.00],
-            [869, 101.00, 107.00, 116.00],
-            [870, 102.00, 108.00, 117.00],
-            [871, 102.00, 108.00, 117.00],
-            [872, 103.00, 109.00, 117.00],
-            [873, 103.00, 109.00, 118.00],
-            [874, 104.00, 110.00, 118.00],
-            [875, 104.00, 110.00, 118.00],
-            [876, 105.00, 111.00, 119.00],
-            [877, 105.00, 111.00, 119.00],
-            [878, 106.00, 112.00, 119.00],
-            [879, 106.00, 112.00, 120.00],
-            [880, 107.00, 113.00, 120.00],
-            [881, 107.00, 113.00, 120.00],
-            [882, 107.00, 114.00, 121.00],
-            [883, 108.00, 114.00, 121.00],
-            [884, 108.00, 114.00, 121.00],
-            [885, 109.00, 115.00, 122.00],
-            [886, 109.00, 115.00, 122.00],
-            [887, 110.00, 116.00, 123.00],
-            [888, 110.00, 116.00, 123.00],
-            [889, 111.00, 117.00, 123.00],
-            [890, 111.00, 117.00, 124.00],
-            [891, 112.00, 118.00, 124.00],
-            [892, 112.00, 118.00, 124.00],
-            [893, 113.00, 119.00, 125.00],
-            [894, 113.00, 119.00, 125.00],
-            [895, 114.00, 120.00, 125.00],
-            [896, 114.00, 120.00, 126.00],
-            [897, 115.00, 121.00, 126.00],
-            [898, 115.00, 121.00, 126.00],
-            [899, 116.00, 122.00, 127.00],
-            [900, 116.00, 122.00, 127.00],
-            [901, 116.00, 123.00, 127.00],
-            [902, 117.00, 123.00, 128.00],
-            [903, 117.00, 123.00, 128.00],
-            [904, 118.00, 124.00, 128.00],
-            [905, 118.00, 124.00, 129.00],
-            [906, 119.00, 125.00, 129.00],
-            [907, 119.00, 125.00, 129.00],
-            [908, 120.00, 126.00, 130.00],
-            [909, 120.00, 126.00, 130.00],
-            [910, 121.00, 127.00, 131.00],
-            [911, 121.00, 127.00, 131.00],
-            [912, 122.00, 128.00, 131.00],
-            [913, 122.00, 128.00, 132.00],
-            [914, 123.00, 129.00, 132.00],
-            [915, 123.00, 129.00, 132.00],
-            [916, 124.00, 130.00, 133.00],
-            [917, 124.00, 130.00, 133.00],
-            [918, 124.00, 131.00, 133.00],
-            [919, 125.00, 131.00, 134.00],
-            [920, 125.00, 131.00, 134.00],
-            [921, 126.00, 132.00, 134.00],
-            [922, 126.00, 132.00, 135.00],
-            [923, 127.00, 133.00, 135.00],
-            [924, 127.00, 133.00, 135.00],
-            [925, 128.00, 134.00, 136.00],
-            [926, 128.00, 134.00, 136.00],
-            [927, 129.00, 135.00, 136.00],
-            [928, 129.00, 135.00, 137.00],
-            [929, 130.00, 136.00, 137.00],
-            [930, 130.00, 136.00, 137.00],
-            [931, 131.00, 137.00, 138.00],
-            [932, 131.00, 137.00, 138.00],
-            [933, 132.00, 138.00, 139.00],
-            [934, 132.00, 138.00, 139.00],
-            [935, 133.00, 139.00, 139.00],
-            [936, 133.00, 139.00, 140.00],
-            [937, 133.00, 140.00, 140.00],
-            [938, 134.00, 140.00, 140.00],
-            [939, 134.00, 140.00, 141.00],
-            [940, 135.00, 141.00, 141.00],
-            [941, 135.00, 141.00, 141.00],
-            [942, 136.00, 142.00, 142.00],
-            [943, 136.00, 142.00, 142.00],
-            [944, 137.00, 142.00, 142.00],
-            [945, 137.00, 143.00, 143.00],
-            [946, 138.00, 143.00, 143.00],
-            [947, 138.00, 143.00, 143.00],
-            [948, 139.00, 144.00, 144.00],
-            [949, 139.00, 144.00, 144.00],
-            [950, 140.00, 144.00, 144.00],
-            [951, 140.00, 145.00, 145.00],
-            [952, 141.00, 145.00, 145.00],
-            [953, 141.00, 145.00, 145.00],
-            [954, 142.00, 146.00, 146.00],
-            [955, 142.00, 146.00, 146.00],
-            [956, 142.00, 147.00, 147.00],
-            [957, 143.00, 147.00, 147.00],
-            [958, 143.00, 147.00, 147.00],
-            [959, 144.00, 148.00, 148.00],
-            [960, 144.00, 148.00, 148.00],
-            [961, 145.00, 148.00, 148.00],
-            [962, 145.00, 149.00, 149.00],
-            [963, 146.00, 149.00, 149.00],
-            [964, 146.00, 149.00, 149.00],
-            [965, 147.00, 150.00, 150.00],
-            [966, 147.00, 150.00, 150.00],
-            [967, 148.00, 150.00, 150.00],
-            [968, 148.00, 151.00, 151.00],
-            [969, 149.00, 151.00, 151.00],
-            [970, 149.00, 151.00, 151.00],
-            [971, 150.00, 152.00, 152.00],
-            [972, 150.00, 152.00, 152.00],
-            [973, 150.00, 152.00, 152.00],
-            [974, 151.00, 153.00, 153.00],
-            [975, 151.00, 153.00, 153.00],
-            [976, 152.00, 153.00, 153.00],
-            [977, 152.00, 154.00, 154.00],
-            [978, 153.00, 154.00, 154.00],
-            [979, 153.00, 155.00, 155.00],
-            [980, 154.00, 155.00, 155.00],
-            [981, 154.00, 155.00, 155.00],
-            [982, 155.00, 156.00, 156.00],
-            [983, 155.00, 156.00, 156.00],
-            [984, 156.00, 156.00, 156.00],
-            [985, 156.00, 157.00, 157.00],
-            [986, 157.00, 157.00, 157.00],
-            [987, 157.00, 157.00, 157.00],
-            [988, 158.00, 158.00, 158.00],
-            [989, 158.00, 158.00, 158.00],
-            [990, 158.00, 158.00, 158.00],
-            [991, 159.00, 159.00, 159.00],
-            [992, 159.00, 159.00, 159.00],
-            [993, 159.00, 159.00, 159.00],
-            [994, 160.00, 160.00, 160.00],
-            [995, 160.00, 160.00, 160.00],
-            [996, 160.00, 160.00, 160.00],
-            [997, 161.00, 161.00, 161.00],
-            [998, 161.00, 161.00, 161.00],
-            [999, 161.00, 161.00, 161.00],
-            [1000, 162.00, 162.00, 162.00],
-            [1001, 162.00, 162.00, 162.00],
-            [1002, 163.00, 163.00, 163.00],
-            [1003, 163.00, 163.00, 163.00],
-            [1004, 163.00, 163.00, 163.00],
-            [1005, 164.00, 164.00, 164.00],
-            [1006, 164.00, 164.00, 164.00],
-            [1007, 164.00, 164.00, 164.00],
-            [1008, 165.00, 165.00, 165.00],
-            [1009, 165.00, 165.00, 165.00],
-            [1010, 165.00, 165.00, 165.00],
-            [1011, 166.00, 166.00, 166.00],
-            [1012, 166.00, 166.00, 166.00],
-            [1013, 166.00, 166.00, 166.00],
-            [1014, 167.00, 167.00, 167.00],
-            [1015, 167.00, 167.00, 167.00],
-            [1016, 167.00, 167.00, 167.00],
-            [1017, 168.00, 168.00, 168.00],
-            [1018, 168.00, 168.00, 168.00],
-            [1019, 168.00, 168.00, 168.00],
-            [1020, 169.00, 169.00, 169.00],
-            [1021, 169.00, 169.00, 169.00],
-            [1022, 169.00, 169.00, 169.00],
-            [1023, 170.00, 170.00, 170.00],
-            [1024, 170.00, 170.00, 170.00],
-            [1025, 171.00, 171.00, 171.00],
-            [1026, 171.00, 171.00, 171.00],
-            [1027, 171.00, 171.00, 171.00],
-            [1028, 172.00, 172.00, 172.00],
-            [1029, 172.00, 172.00, 172.00],
-            [1030, 172.00, 172.00, 172.00],
-            [1031, 173.00, 173.00, 173.00],
-            [1032, 173.00, 173.00, 173.00],
-            [1033, 173.00, 173.00, 173.00],
-            [1034, 174.00, 174.00, 174.00],
-            [1035, 174.00, 174.00, 174.00],
-            [1036, 174.00, 174.00, 174.00],
-            [1037, 175.00, 175.00, 175.00],
-            [1038, 175.00, 175.00, 175.00],
-            [1039, 175.00, 175.00, 175.00],
-            [1040, 176.00, 176.00, 176.00],
-            [1041, 176.00, 176.00, 176.00],
-            [1042, 176.00, 176.00, 176.00],
-            [1043, 177.00, 177.00, 177.00],
-            [1044, 177.00, 177.00, 177.00],
-            [1045, 177.00, 177.00, 177.00],
-            [1046, 178.00, 178.00, 178.00],
-            [1047, 178.00, 178.00, 178.00],
-            [1048, 179.00, 179.00, 179.00],
-            [1049, 179.00, 179.00, 179.00],
-            [1050, 179.00, 179.00, 179.00],
-            [1051, 180.00, 180.00, 180.00],
-            [1052, 180.00, 180.00, 180.00],
-            [1053, 180.00, 180.00, 180.00],
-            [1054, 181.00, 181.00, 181.00],
-            [1055, 181.00, 181.00, 181.00],
-            [1056, 181.00, 181.00, 181.00],
-            [1057, 182.00, 182.00, 182.00],
-            [1058, 182.00, 182.00, 182.00],
-            [1059, 182.00, 182.00, 182.00],
-            [1060, 183.00, 183.00, 183.00],
-            [1061, 183.00, 183.00, 183.00],
-            [1062, 183.00, 183.00, 183.00],
-            [1063, 184.00, 184.00, 184.00],
-            [1064, 184.00, 184.00, 184.00],
-            [1065, 184.00, 184.00, 184.00],
-            [1066, 185.00, 185.00, 185.00],
-            [1067, 185.00, 185.00, 185.00],
-            [1068, 185.00, 185.00, 185.00],
-            [1069, 186.00, 186.00, 186.00],
-            [1070, 186.00, 186.00, 186.00],
-            [1071, 187.00, 187.00, 187.00],
-            [1072, 187.00, 187.00, 187.00],
-            [1073, 187.00, 187.00, 187.00],
-            [1074, 188.00, 188.00, 188.00],
-            [1075, 188.00, 188.00, 188.00],
-            [1076, 188.00, 188.00, 188.00],
-            [1077, 189.00, 189.00, 189.00],
-            [1078, 189.00, 189.00, 189.00],
-            [1079, 189.00, 189.00, 189.00],
-            [1080, 190.00, 190.00, 190.00],
-            [1081, 190.00, 190.00, 190.00],
-            [1082, 190.00, 190.00, 190.00],
-            [1083, 191.00, 191.00, 191.00],
-            [1084, 191.00, 191.00, 191.00],
-            [1085, 191.00, 191.00, 191.00],
-            [1086, 192.00, 192.00, 192.00],
-            [1087, 192.00, 192.00, 192.00],
-            [1088, 192.00, 192.00, 192.00],
-            [1089, 193.00, 193.00, 193.00],
-            [1090, 193.00, 193.00, 193.00],
-            [1091, 193.00, 193.00, 193.00],
-            [1092, 194.00, 194.00, 194.00],
-            [1093, 194.00, 194.00, 194.00],
-            [1094, 195.00, 195.00, 195.00],
-            [1095, 195.00, 195.00, 195.00],
-            [1096, 195.00, 195.00, 195.00],
-            [1097, 196.00, 196.00, 196.00],
-            [1098, 196.00, 196.00, 196.00],
-            [1099, 196.00, 196.00, 196.00],
-            [1100, 197.00, 197.00, 197.00],
-            [1101, 197.00, 197.00, 197.00],
-            [1102, 197.00, 197.00, 197.00],
-            [1103, 198.00, 198.00, 198.00],
-            [1104, 198.00, 198.00, 198.00],
-            [1105, 198.00, 198.00, 198.00],
-            [1106, 199.00, 199.00, 199.00],
-            [1107, 199.00, 199.00, 199.00],
-            [1108, 199.00, 199.00, 199.00],
-            [1109, 200.00, 200.00, 200.00],
-            [1110, 200.00, 200.00, 200.00],
-            [1111, 200.00, 200.00, 200.00],
-            [1112, 201.00, 201.00, 201.00],
-            [1113, 201.00, 201.00, 201.00],
-            [1114, 201.00, 201.00, 201.00],
-            [1115, 202.00, 202.00, 202.00],
-            [1116, 202.00, 202.00, 202.00],
-            [1117, 203.00, 203.00, 203.00],
-            [1118, 203.00, 203.00, 203.00],
-            [1119, 203.00, 203.00, 203.00],
-            [1120, 204.00, 204.00, 204.00],
-            [1121, 204.00, 204.00, 204.00],
-            [1122, 204.00, 204.00, 204.00],
-            [1123, 205.00, 205.00, 205.00],
-            [1124, 205.00, 205.00, 205.00],
-            [1125, 205.00, 205.00, 205.00],
-            [1126, 206.00, 206.00, 206.00],
-            [1127, 206.00, 206.00, 206.00],
-            [1128, 206.00, 206.00, 206.00],
-            [1129, 207.00, 207.00, 207.00],
-            [1130, 207.00, 207.00, 207.00],
-            [1131, 207.00, 207.00, 207.00],
-            [1132, 208.00, 208.00, 208.00],
-            [1133, 208.00, 208.00, 208.00],
-            [1134, 208.00, 208.00, 208.00],
-            [1135, 209.00, 209.00, 209.00],
-            [1136, 209.00, 209.00, 209.00],
-            [1137, 209.00, 209.00, 209.00],
-            [1138, 210.00, 210.00, 210.00],
-            [1139, 210.00, 210.00, 210.00],
-            [1140, 211.00, 211.00, 211.00],
-            [1141, 211.00, 211.00, 211.00],
-            [1142, 211.00, 211.00, 211.00],
-            [1143, 212.00, 212.00, 212.00],
-            [1144, 212.00, 212.00, 212.00],
-            [1145, 212.00, 212.00, 212.00],
-            [1146, 213.00, 213.00, 213.00],
-            [1147, 213.00, 213.00, 213.00],
-            [1148, 213.00, 213.00, 213.00],
-            [1149, 214.00, 214.00, 214.00],
-            [1150, 214.00, 214.00, 214.00],
-            [1151, 214.00, 214.00, 214.00],
-            [1152, 215.00, 215.00, 215.00],
-            [1153, 215.00, 215.00, 215.00],
-            [1154, 215.00, 215.00, 215.00],
-            [1155, 216.00, 216.00, 216.00],
-            [1156, 216.00, 216.00, 216.00],
-            [1157, 216.00, 216.00, 216.00],
-            [1158, 217.00, 217.00, 217.00],
-            [1159, 217.00, 217.00, 217.00],
-            [1160, 217.00, 217.00, 217.00],
-            [1161, 218.00, 218.00, 218.00],
-            [1162, 218.00, 218.00, 218.00],
-            [1163, 219.00, 219.00, 219.00],
-            [1164, 219.00, 219.00, 219.00],
-            [1165, 219.00, 219.00, 219.00],
-            [1166, 220.00, 220.00, 220.00],
-            [1167, 220.00, 220.00, 220.00],
-            [1168, 220.00, 220.00, 220.00],
-            [1169, 221.00, 221.00, 221.00],
-            [1170, 221.00, 221.00, 221.00],
-            [1171, 221.00, 221.00, 221.00],
-            [1172, 222.00, 222.00, 222.00],
-            [1173, 222.00, 222.00, 222.00],
-            [1174, 222.00, 222.00, 222.00],
-            [1175, 223.00, 223.00, 223.00],
-            [1176, 223.00, 223.00, 223.00],
-            [1177, 223.00, 223.00, 223.00],
-            [1178, 224.00, 224.00, 224.00],
-            [1179, 224.00, 224.00, 224.00],
-            [1180, 224.00, 224.00, 224.00],
-            [1181, 225.00, 225.00, 225.00],
-            [1182, 225.00, 225.00, 225.00],
-            [1183, 225.00, 225.00, 225.00],
-            [1184, 226.00, 226.00, 226.00],
-            [1185, 226.00, 226.00, 226.00],
-            [1186, 227.00, 227.00, 227.00],
-            [1187, 227.00, 227.00, 227.00],
-            [1188, 227.00, 227.00, 227.00],
-            [1189, 228.00, 228.00, 228.00],
-            [1190, 228.00, 228.00, 228.00],
-            [1191, 228.00, 228.00, 228.00],
-            [1192, 229.00, 229.00, 229.00],
-            [1193, 229.00, 229.00, 229.00],
-            [1194, 229.00, 229.00, 229.00],
-            [1195, 230.00, 230.00, 230.00],
-            [1196, 230.00, 230.00, 230.00],
-            [1197, 230.00, 230.00, 230.00],
-            [1198, 231.00, 231.00, 231.00],
-            [1199, 231.00, 231.00, 231.00],
-            [1200, 231.00, 231.00, 231.00],
-            [1201, 232.00, 232.00, 232.00],
-            [1202, 232.00, 232.00, 232.00],
-            [1203, 232.00, 232.00, 232.00],
-            [1204, 233.00, 233.00, 233.00],
-            [1205, 233.00, 233.00, 233.00],
-            [1206, 233.00, 233.00, 233.00],
-            [1207, 234.00, 234.00, 234.00],
-            [1208, 234.00, 234.00, 234.00],
-            [1209, 235.00, 235.00, 235.00],
-            [1210, 235.00, 235.00, 235.00],
-            [1211, 235.00, 235.00, 235.00],
-            [1212, 236.00, 236.00, 236.00],
-            [1213, 236.00, 236.00, 236.00],
-            [1214, 236.00, 236.00, 236.00],
-            [1215, 237.00, 237.00, 237.00],
-            [1216, 237.00, 237.00, 237.00],
-            [1217, 237.00, 237.00, 237.00],
-            [1218, 238.00, 238.00, 238.00],
-            [1219, 238.00, 238.00, 238.00],
-            [1220, 238.00, 238.00, 238.00],
-            [1221, 239.00, 239.00, 239.00],
-            [1222, 239.00, 239.00, 239.00],
-            [1223, 239.00, 239.00, 239.00],
-            [1224, 240.00, 240.00, 240.00],
-            [1225, 240.00, 240.00, 240.00],
-            [1226, 240.00, 240.00, 240.00],
-            [1227, 241.00, 241.00, 241.00],
-            [1228, 241.00, 241.00, 241.00],
-            [1229, 241.00, 241.00, 241.00],
-            [1230, 242.00, 242.00, 242.00],
-            [1231, 242.00, 242.00, 242.00],
-            [1232, 242.00, 242.00, 242.00],
-            [1233, 243.00, 243.00, 243.00],
-            [1234, 243.00, 243.00, 243.00],
-            [1235, 244.00, 244.00, 244.00],
-            [1236, 244.00, 244.00, 244.00],
-            [1237, 244.00, 244.00, 244.00],
-            [1238, 245.00, 245.00, 245.00],
-            [1239, 245.00, 245.00, 245.00],
-            [1240, 245.00, 245.00, 245.00],
-            [1241, 246.00, 246.00, 246.00],
-            [1242, 246.00, 246.00, 246.00],
-            [1243, 246.00, 246.00, 246.00],
-            [1244, 247.00, 247.00, 247.00],
-            [1245, 247.00, 247.00, 247.00],
-            [1246, 247.00, 247.00, 247.00],
-            [1247, 248.00, 248.00, 248.00],
-            [1248, 248.00, 248.00, 248.00],
-            [1249, 248.00, 248.00, 248.00],
-            [1250, 249.00, 249.00, 249.00],
-            [1251, 249.00, 249.00, 249.00],
-            [1252, 249.00, 249.00, 249.00],
-            [1253, 250.00, 250.00, 250.00],
-            [1254, 250.00, 250.00, 250.00],
-            [1255, 250.00, 250.00, 250.00],
-            [1256, 251.00, 251.00, 251.00],
-            [1257, 251.00, 251.00, 251.00],
-            [1258, 252.00, 252.00, 252.00],
-            [1259, 252.00, 252.00, 252.00],
-            [1260, 252.00, 252.00, 252.00],
-            [1261, 253.00, 253.00, 253.00],
-            [1262, 253.00, 253.00, 253.00],
-            [1263, 253.00, 253.00, 253.00],
-            [1264, 254.00, 254.00, 254.00],
-            [1265, 254.00, 254.00, 254.00],
-            [1266, 254.00, 254.00, 254.00],
-            [1267, 255.00, 255.00, 255.00],
-            [1268, 255.00, 255.00, 255.00],
-            [1269, 255.00, 255.00, 255.00],
-            [1270, 256.00, 256.00, 256.00],
-            [1271, 256.00, 256.00, 256.00],
-            [1272, 256.00, 256.00, 256.00],
-            [1273, 257.00, 257.00, 257.00],
-            [1274, 257.00, 257.00, 257.00],
-            [1275, 257.00, 257.00, 257.00],
-            [1276, 258.00, 258.00, 258.00],
-            [1277, 258.00, 258.00, 258.00],
-            [1278, 258.00, 258.00, 258.00],
-            [1279, 259.00, 259.00, 259.00],
-            [1280, 259.00, 259.00, 259.00],
-            [1281, 260.00, 260.00, 260.00],
-            [1282, 260.00, 260.00, 260.00],
-            [1283, 260.00, 260.00, 260.00],
-            [1284, 261.00, 261.00, 261.00],
-            [1285, 261.00, 261.00, 261.00],
-            [1286, 261.00, 261.00, 261.00],
-            [1287, 262.00, 262.00, 262.00],
-            [1288, 262.00, 262.00, 262.00],
-            [1289, 262.00, 262.00, 262.00],
-            [1290, 263.00, 263.00, 263.00],
-            [1291, 263.00, 263.00, 263.00],
-            [1292, 263.00, 263.00, 263.00],
-            [1293, 264.00, 264.00, 264.00],
-            [1294, 264.00, 264.00, 264.00],
-            [1295, 264.00, 264.00, 264.00],
-            [1296, 265.00, 265.00, 265.00],
-            [1297, 265.00, 265.00, 265.00],
-            [1298, 265.00, 265.00, 265.00],
-            [1299, 266.00, 266.00, 266.00],
-            [1300, 266.00, 266.00, 266.00],
-            [1301, 266.00, 266.00, 266.00],
-            [1302, 267.00, 267.00, 267.00],
-            [1303, 267.00, 267.00, 267.00],
-            [1304, 267.00, 267.00, 267.00],
-            [1305, 268.00, 268.00, 268.00],
-            [1306, 268.00, 268.00, 268.00],
-            [1307, 269.00, 269.00, 269.00],
-            [1308, 269.00, 269.00, 269.00],
-            [1309, 269.00, 269.00, 269.00],
-            [1310, 270.00, 270.00, 270.00],
-            [1311, 270.00, 270.00, 270.00],
-            [1312, 270.00, 270.00, 270.00],
-            [1313, 271.00, 271.00, 271.00],
-            [1314, 271.00, 271.00, 271.00],
-            [1315, 271.00, 271.00, 271.00],
-            [1316, 272.00, 272.00, 272.00],
-            [1317, 272.00, 272.00, 272.00],
-            [1318, 272.00, 272.00, 272.00],
-            [1319, 273.00, 273.00, 273.00],
-            [1320, 273.00, 273.00, 273.00],
-            [1321, 273.00, 273.00, 273.00],
-            [1322, 274.00, 274.00, 274.00],
-            [1323, 274.00, 274.00, 274.00],
-            [1324, 274.00, 274.00, 274.00],
-            [1325, 275.00, 275.00, 275.00],
-            [1326, 275.00, 275.00, 275.00],
-            [1327, 275.00, 275.00, 275.00],
-            [1328, 276.00, 276.00, 276.00],
-            [1329, 276.00, 276.00, 276.00],
-            [1330, 276.00, 276.00, 276.00],
-            [1331, 277.00, 277.00, 277.00],
-            [1332, 277.00, 277.00, 277.00],
-            [1333, 277.00, 277.00, 277.00],
-            [1334, 278.00, 278.00, 278.00],
-            [1335, 278.00, 278.00, 278.00],
-            [1336, 279.00, 279.00, 279.00],
-            [1337, 279.00, 279.00, 279.00],
-            [1338, 279.00, 279.00, 279.00],
-            [1339, 280.00, 280.00, 280.00],
-            [1340, 280.00, 280.00, 280.00],
-            [1341, 280.00, 280.00, 280.00],
-            [1342, 281.00, 281.00, 281.00],
-            [1343, 281.00, 281.00, 281.00],
-            [1344, 281.00, 281.00, 281.00],
-            [1345, 282.00, 282.00, 282.00],
-            [1346, 282.00, 282.00, 282.00],
-            [1347, 282.00, 282.00, 282.00],
-            [1348, 283.00, 283.00, 283.00],
-            [1349, 283.00, 283.00, 283.00],
-            [1350, 283.00, 283.00, 283.00],
-            [1351, 284.00, 284.00, 284.00],
-            [1352, 284.00, 284.00, 284.00],
-            [1353, 284.00, 284.00, 284.00],
-            [1354, 285.00, 285.00, 285.00],
-            [1355, 285.00, 285.00, 285.00],
-            [1356, 285.00, 285.00, 285.00],
-            [1357, 286.00, 286.00, 286.00],
-            [1358, 286.00, 286.00, 286.00],
-            [1359, 286.00, 286.00, 286.00],
-            [1360, 287.00, 287.00, 287.00],
-            [1361, 287.00, 287.00, 287.00],
-            [1362, 287.00, 287.00, 287.00],
-            [1363, 288.00, 288.00, 288.00],
-            [1364, 288.00, 288.00, 288.00],
-            [1365, 289.00, 289.00, 289.00],
-            [1366, 289.00, 289.00, 289.00],
-            [1367, 289.00, 289.00, 289.00],
-            [1368, 290.00, 290.00, 290.00],
-            [1369, 290.00, 290.00, 290.00],
-            [1370, 290.00, 290.00, 290.00],
-            [1371, 291.00, 291.00, 291.00],
-            [1372, 291.00, 291.00, 291.00],
-            [1373, 291.00, 291.00, 291.00],
-            [1374, 292.00, 292.00, 292.00],
-            [1375, 292.00, 292.00, 292.00],
-            [1376, 292.00, 292.00, 292.00],
-            [1377, 293.00, 293.00, 293.00],
-            [1378, 293.00, 293.00, 293.00],
-            [1379, 293.00, 293.00, 293.00],
-            [1380, 294.00, 294.00, 294.00],
-            [1381, 294.00, 294.00, 294.00],
-            [1382, 294.00, 294.00, 294.00],
-            [1383, 295.00, 295.00, 295.00],
-            [1384, 295.00, 295.00, 295.00],
-            [1385, 295.00, 295.00, 295.00],
-            [1386, 296.00, 296.00, 296.00],
-            [1387, 296.00, 296.00, 296.00],
-            [1388, 296.00, 296.00, 296.00],
-            [1389, 297.00, 297.00, 297.00],
-            [1390, 297.00, 297.00, 297.00],
-            [1391, 297.00, 297.00, 297.00],
-            [1392, 298.00, 298.00, 298.00],
-            [1393, 298.00, 298.00, 298.00],
-            [1394, 299.00, 299.00, 299.00],
-            [1395, 299.00, 299.00, 299.00],
-            [1396, 299.00, 299.00, 299.00],
-            [1397, 300.00, 300.00, 300.00],
-            [1398, 300.00, 300.00, 300.00],
-            [1399, 300.00, 300.00, 300.00],
-            [1400, 301.00, 301.00, 301.00],
-            [1401, 301.00, 301.00, 301.00],
-            [1402, 301.00, 301.00, 301.00],
-            [1403, 302.00, 302.00, 302.00],
-            [1404, 302.00, 302.00, 302.00],
-            [1405, 302.00, 302.00, 302.00],
-            [1406, 303.00, 303.00, 303.00],
-            [1407, 303.00, 303.00, 303.00],
-            [1408, 303.00, 303.00, 303.00],
-            [1409, 304.00, 304.00, 304.00],
-            [1410, 304.00, 304.00, 304.00],
-            [1411, 304.00, 304.00, 304.00],
-            [1412, 305.00, 305.00, 305.00],
-            [1413, 305.00, 305.00, 305.00],
-            [1414, 305.00, 305.00, 305.00],
-            [1415, 306.00, 306.00, 306.00],
-            [1416, 306.00, 306.00, 306.00],
-            [1417, 306.00, 306.00, 306.00],
-            [1418, 307.00, 307.00, 307.00],
-            [1419, 307.00, 307.00, 307.00],
-            [1420, 307.00, 307.00, 307.00],
-            [1421, 308.00, 308.00, 308.00],
-            [1422, 308.00, 308.00, 308.00],
-            [1423, 309.00, 309.00, 309.00],
-            [1424, 309.00, 309.00, 309.00],
-            [1425, 309.00, 309.00, 309.00],
-            [1426, 310.00, 310.00, 310.00],
-            [1427, 310.00, 310.00, 310.00],
-            [1428, 310.00, 310.00, 310.00],
-            [1429, 311.00, 311.00, 311.00],
-            [1430, 311.00, 311.00, 311.00],
-            [1431, 311.00, 311.00, 311.00],
-            [1432, 312.00, 312.00, 312.00],
-            [1433, 312.00, 312.00, 312.00],
-            [1434, 312.00, 312.00, 312.00],
-            [1435, 313.00, 313.00, 313.00],
-            [1436, 313.00, 313.00, 313.00],
-            [1437, 313.00, 313.00, 313.00],
-            [1438, 314.00, 314.00, 314.00],
-            [1439, 314.00, 314.00, 314.00],
-            [1440, 314.00, 314.00, 314.00],
-            [1441, 315.00, 315.00, 315.00],
-            [1442, 315.00, 315.00, 315.00],
-            [1443, 315.00, 315.00, 315.00],
-            [1444, 316.00, 316.00, 316.00],
-            [1445, 316.00, 316.00, 316.00],
-            [1446, 316.00, 316.00, 316.00],
-            [1447, 317.00, 317.00, 317.00],
-            [1448, 317.00, 317.00, 317.00],
-            [1449, 317.00, 317.00, 317.00],
-            [1450, 318.00, 318.00, 318.00],
-            [1451, 318.00, 318.00, 318.00],
-            [1452, 319.00, 319.00, 319.00],
-            [1453, 319.00, 319.00, 319.00],
-            [1454, 319.00, 319.00, 319.00],
-            [1455, 320.00, 320.00, 320.00],
-            [1456, 320.00, 320.00, 320.00],
-            [1457, 320.00, 320.00, 320.00],
-            [1458, 321.00, 321.00, 321.00],
-            [1459, 321.00, 321.00, 321.00],
-            [1460, 321.00, 321.00, 321.00],
-            [1461, 322.00, 322.00, 322.00],
-            [1462, 322.00, 322.00, 322.00],
-            [1463, 322.00, 322.00, 322.00],
-            [1464, 323.00, 323.00, 323.00],
-            [1465, 323.00, 323.00, 323.00],
-            [1466, 323.00, 323.00, 323.00],
-            [1467, 324.00, 324.00, 324.00],
-            [1468, 324.00, 324.00, 324.00],
-            [1469, 324.00, 324.00, 324.00],
-            [1470, 325.00, 325.00, 325.00],
-            [1471, 325.00, 325.00, 325.00],
-            [1472, 325.00, 325.00, 325.00],
-            [1473, 326.00, 326.00, 326.00],
-            [1474, 326.00, 326.00, 326.00],
-            [1475, 326.00, 326.00, 326.00],
-            [1476, 327.00, 327.00, 327.00],
-            [1477, 327.00, 327.00, 327.00],
-            [1478, 328.00, 328.00, 328.00],
-            [1479, 328.00, 328.00, 328.00],
-            [1480, 328.00, 328.00, 328.00],
-            [1481, 329.00, 329.00, 329.00],
-            [1482, 329.00, 329.00, 329.00],
-            [1483, 329.00, 329.00, 329.00],
-            [1484, 330.00, 330.00, 330.00],
-            [1485, 330.00, 330.00, 330.00],
-            [1486, 330.00, 330.00, 330.00],
-            [1487, 331.00, 331.00, 331.00],
-            [1488, 331.00, 331.00, 331.00],
-            [1489, 331.00, 331.00, 331.00],
-            [1490, 332.00, 332.00, 332.00],
-            [1491, 332.00, 332.00, 332.00],
-            [1492, 332.00, 332.00, 332.00],
-            [1493, 333.00, 333.00, 333.00],
-            [1494, 333.00, 333.00, 333.00],
-            [1495, 333.00, 333.00, 333.00],
-            [1496, 334.00, 334.00, 334.00],
-            [1497, 334.00, 334.00, 334.00],
-            [1498, 334.00, 334.00, 334.00],
-            [1499, 335.00, 335.00, 335.00],
-            [1500, 335.00, 335.00, 335.00],
-            [1501, 335.00, 335.00, 335.00],
-            [1502, 336.00, 336.00, 336.00],
-            [1503, 336.00, 336.00, 336.00],
-            [1504, 336.00, 336.00, 336.00],
-            [1505, 337.00, 337.00, 337.00],
-            [1506, 337.00, 337.00, 337.00],
-            [1507, 338.00, 338.00, 338.00],
-            [1508, 338.00, 338.00, 338.00],
-            [1509, 338.00, 338.00, 338.00],
-            [1510, 339.00, 339.00, 339.00],
-            [1511, 339.00, 339.00, 339.00],
-            [1512, 339.00, 339.00, 339.00],
-            [1513, 340.00, 340.00, 340.00],
-            [1514, 340.00, 340.00, 340.00],
-            [1515, 340.00, 340.00, 340.00],
-            [1516, 341.00, 341.00, 341.00],
-            [1517, 341.00, 341.00, 341.00],
-            [1518, 341.00, 341.00, 341.00],
-            [1519, 342.00, 342.00, 342.00],
-            [1520, 342.00, 342.00, 342.00],
-            [1521, 342.00, 342.00, 342.00],
-            [1522, 343.00, 343.00, 343.00],
-            [1523, 343.00, 343.00, 343.00],
-            [1524, 343.00, 343.00, 343.00],
-            [1525, 344.00, 344.00, 344.00],
-            [1526, 344.00, 344.00, 344.00],
-            [1527, 344.00, 344.00, 344.00],
-            [1528, 345.00, 345.00, 345.00],
-            [1529, 345.00, 345.00, 345.00],
-            [1530, 345.00, 345.00, 345.00],
-            [1531, 346.00, 346.00, 346.00],
-            [1532, 346.00, 346.00, 346.00],
-            [1533, 346.00, 346.00, 346.00],
-            [1534, 347.00, 347.00, 347.00],
-            [1535, 347.00, 347.00, 347.00],
-            [1536, 348.00, 348.00, 348.00],
-            [1537, 348.00, 348.00, 348.00],
-            [1538, 348.00, 348.00, 348.00],
-            [1539, 349.00, 349.00, 349.00],
-            [1540, 349.00, 349.00, 349.00],
-            [1541, 349.00, 349.00, 349.00],
-            [1542, 350.00, 350.00, 350.00],
-            [1543, 350.00, 350.00, 350.00],
-            [1544, 350.00, 350.00, 350.00],
-            [1545, 351.00, 351.00, 351.00],
-            [1546, 351.00, 351.00, 351.00],
-            [1547, 351.00, 351.00, 351.00],
-            [1548, 352.00, 352.00, 352.00],
-            [1549, 352.00, 352.00, 352.00],
-            [1550, 352.00, 352.00, 352.00],
-            [1551, 353.00, 353.00, 353.00],
-            [1552, 353.00, 353.00, 353.00],
-            [1553, 353.00, 353.00, 353.00],
-            [1554, 354.00, 354.00, 354.00],
-            [1555, 354.00, 354.00, 354.00],
-            [1556, 354.00, 354.00, 354.00],
-            [1557, 355.00, 355.00, 355.00],
-            [1558, 355.00, 355.00, 355.00],
-            [1559, 355.00, 355.00, 355.00],
-            [1560, 356.00, 356.00, 356.00],
-            [1561, 356.00, 356.00, 356.00],
-            [1562, 356.00, 356.00, 356.00],
-            [1563, 357.00, 357.00, 357.00],
-            [1564, 357.00, 357.00, 357.00],
-            [1565, 358.00, 358.00, 358.00],
-            [1566, 358.00, 358.00, 358.00],
-            [1567, 358.00, 358.00, 358.00],
-            [1568, 359.00, 359.00, 359.00],
-            [1569, 359.00, 359.00, 359.00],
-            [1570, 359.00, 359.00, 359.00],
-            [1571, 360.00, 360.00, 360.00],
-            [1572, 360.00, 360.00, 360.00],
-            [1573, 360.00, 360.00, 360.00],
-            [1574, 361.00, 361.00, 361.00],
-            [1575, 361.00, 361.00, 361.00],
-            [1576, 361.00, 361.00, 361.00],
-            [1577, 362.00, 362.00, 362.00],
-            [1578, 362.00, 362.00, 362.00],
-            [1579, 362.00, 362.00, 362.00],
-            [1580, 363.00, 363.00, 363.00],
-            [1581, 363.00, 363.00, 363.00],
-            [1582, 363.00, 363.00, 363.00],
-            [1583, 364.00, 364.00, 364.00],
-            [1584, 364.00, 364.00, 364.00],
-            [1585, 364.00, 364.00, 364.00],
-            [1586, 365.00, 365.00, 365.00],
-            [1587, 365.00, 365.00, 365.00],
-            [1588, 365.00, 365.00, 365.00],
-            [1589, 366.00, 366.00, 366.00],
-            [1590, 366.00, 366.00, 366.00],
-            [1591, 366.00, 366.00, 366.00],
-            [1592, 367.00, 367.00, 367.00],
-            [1593, 367.00, 367.00, 367.00],
-            [1594, 368.00, 368.00, 368.00],
-            [1595, 368.00, 368.00, 368.00],
-            [1596, 368.00, 368.00, 368.00],
-            [1597, 369.00, 369.00, 369.00],
-            [1598, 369.00, 369.00, 369.00],
-            [1599, 369.00, 369.00, 369.00],
-            [1600, 370.00, 370.00, 370.00],
-            [1601, 370.00, 370.00, 370.00],
-            [1602, 370.00, 370.00, 370.00],
-            [1603, 371.00, 371.00, 371.00],
-            [1604, 371.00, 371.00, 371.00],
-            [1605, 371.00, 371.00, 371.00],
-            [1606, 372.00, 372.00, 372.00],
-            [1607, 372.00, 372.00, 372.00],
-            [1608, 372.00, 372.00, 372.00],
-            [1609, 373.00, 373.00, 373.00],
-            [1610, 373.00, 373.00, 373.00],
-            [1611, 373.00, 373.00, 373.00],
-            [1612, 374.00, 374.00, 374.00],
-            [1613, 374.00, 374.00, 374.00],
-            [1614, 374.00, 374.00, 374.00],
-            [1615, 375.00, 375.00, 375.00],
-            [1616, 375.00, 375.00, 375.00],
-            [1617, 375.00, 375.00, 375.00],
-            [1618, 376.00, 376.00, 376.00],
-            [1619, 376.00, 376.00, 376.00],
-            [1620, 376.00, 376.00, 376.00],
-            [1621, 377.00, 377.00, 377.00],
-            [1622, 377.00, 377.00, 377.00],
-            [1623, 378.00, 378.00, 378.00],
-            [1624, 378.00, 378.00, 378.00],
-            [1625, 378.00, 378.00, 378.00],
-            [1626, 379.00, 379.00, 379.00],
-            [1627, 379.00, 379.00, 379.00],
-            [1628, 379.00, 379.00, 379.00],
-            [1629, 380.00, 380.00, 380.00],
-            [1630, 380.00, 380.00, 380.00],
-            [1631, 380.00, 380.00, 380.00],
-            [1632, 381.00, 381.00, 381.00],
-            [1633, 381.00, 381.00, 381.00],
-            [1634, 381.00, 381.00, 381.00],
-            [1635, 382.00, 382.00, 382.00],
-            [1636, 382.00, 382.00, 382.00],
-            [1637, 382.00, 382.00, 382.00],
-            [1638, 383.00, 383.00, 383.00],
-            [1639, 383.00, 383.00, 383.00],
-            [1640, 383.00, 383.00, 383.00],
-            [1641, 384.00, 384.00, 384.00],
-            [1642, 384.00, 384.00, 384.00],
-            [1643, 384.00, 384.00, 384.00],
-            [1644, 385.00, 385.00, 385.00],
-            [1645, 385.00, 385.00, 385.00],
-            [1646, 385.00, 385.00, 385.00],
-            [1647, 386.00, 386.00, 386.00],
-            [1648, 386.00, 386.00, 386.00],
-            [1649, 386.00, 386.00, 386.00],
-            [1650, 387.00, 387.00, 387.00],
-            [1651, 387.00, 387.00, 387.00],
-            [1652, 388.00, 388.00, 388.00],
-            [1653, 388.00, 388.00, 388.00],
-            [1654, 388.00, 388.00, 388.00],
-            [1655, 389.00, 389.00, 389.00],
-            [1656, 389.00, 389.00, 389.00],
-            [1657, 389.00, 389.00, 389.00],
-            [1658, 390.00, 390.00, 390.00],
-            [1659, 390.00, 390.00, 390.00],
-            [1660, 390.00, 390.00, 390.00],
-            [1661, 391.00, 391.00, 391.00],
-            [1662, 391.00, 391.00, 391.00],
-            [1663, 391.00, 391.00, 391.00],
-            [1664, 392.00, 392.00, 392.00],
-            [1665, 392.00, 392.00, 392.00],
-            [1666, 392.00, 392.00, 392.00],
-            [1667, 393.00, 393.00, 393.00],
-            [1668, 393.00, 393.00, 393.00],
-            [1669, 393.00, 393.00, 393.00],
-            [1670, 394.00, 394.00, 394.00],
-            [1671, 394.00, 394.00, 394.00],
-            [1672, 394.00, 394.00, 394.00],
-            [1673, 395.00, 395.00, 395.00],
-            [1674, 395.00, 395.00, 395.00],
-            [1675, 395.00, 395.00, 395.00],
-            [1676, 396.00, 396.00, 396.00],
-            [1677, 396.00, 396.00, 396.00],
-            [1678, 397.00, 397.00, 397.00],
-            [1679, 397.00, 397.00, 397.00],
-            [1680, 397.00, 397.00, 397.00],
-            [1681, 398.00, 398.00, 398.00],
-            [1682, 398.00, 398.00, 398.00],
-            [1683, 398.00, 398.00, 398.00],
-            [1684, 399.00, 399.00, 399.00],
-            [1685, 399.00, 399.00, 399.00],
-            [1686, 399.00, 399.00, 399.00],
-            [1687, 400.00, 400.00, 400.00],
-            [1688, 400.00, 400.00, 400.00],
-            [1689, 400.00, 400.00, 400.00],
-            [1690, 401.00, 401.00, 401.00],
-            [1691, 401.00, 401.00, 401.00],
-            [1692, 401.00, 401.00, 401.00],
-            [1693, 402.00, 402.00, 402.00],
-            [1694, 402.00, 402.00, 402.00],
-            [1695, 402.00, 402.00, 402.00],
-            [1696, 403.00, 403.00, 403.00],
-            [1697, 403.00, 403.00, 403.00],
-            [1698, 403.00, 403.00, 403.00],
-            [1699, 404.00, 404.00, 404.00],
-            [1700, 404.00, 404.00, 404.00],
-            [1701, 404.00, 404.00, 404.00],
-            [1702, 405.00, 405.00, 405.00],
-            [1703, 405.00, 405.00, 405.00],
-            [1704, 405.00, 405.00, 405.00],
-            [1705, 406.00, 406.00, 406.00],
-            [1706, 406.00, 406.00, 406.00],
-            [1707, 407.00, 407.00, 407.00],
-            [1708, 407.00, 407.00, 407.00],
-            [1709, 407.00, 407.00, 407.00],
-            [1710, 408.00, 408.00, 408.00],
-            [1711, 408.00, 408.00, 408.00],
-            [1712, 408.00, 408.00, 408.00],
-            [1713, 409.00, 409.00, 409.00],
-            [1714, 409.00, 409.00, 409.00],
-            [1715, 409.00, 409.00, 409.00],
-            [1716, 410.00, 410.00, 410.00],
-            [1717, 410.00, 410.00, 410.00],
-            [1718, 410.00, 410.00, 410.00],
-            [1719, 411.00, 411.00, 411.00],
-            [1720, 411.00, 411.00, 411.00],
-            [1721, 411.00, 411.00, 411.00],
-            [1722, 412.00, 412.00, 412.00],
-            [1723, 412.00, 412.00, 412.00],
-            [1724, 412.00, 412.00, 412.00],
-            [1725, 413.00, 413.00, 413.00],
-            [1726, 413.00, 413.00, 413.00],
-            [1727, 413.00, 413.00, 413.00],
-            [1728, 414.00, 414.00, 414.00],
-            [1729, 414.00, 414.00, 414.00],
-            [1730, 414.00, 414.00, 414.00],
-            [1731, 415.00, 415.00, 415.00],
-            [1732, 415.00, 415.00, 415.00],
-            [1733, 415.00, 415.00, 415.00],
-            [1734, 416.00, 416.00, 416.00],
-            [1735, 416.00, 416.00, 416.00],
-            [1736, 417.00, 417.00, 417.00],
-            [1737, 417.00, 417.00, 417.00],
-            [1738, 417.00, 417.00, 417.00],
-            [1739, 418.00, 418.00, 418.00],
-            [1740, 418.00, 418.00, 418.00],
-            [1741, 418.00, 418.00, 418.00],
-            [1742, 419.00, 419.00, 419.00],
-            [1743, 419.00, 419.00, 419.00],
-            [1744, 419.00, 419.00, 419.00],
-            [1745, 420.00, 420.00, 420.00],
-            [1746, 420.00, 420.00, 420.00],
-            [1747, 420.00, 420.00, 420.00],
-            [1748, 421.00, 421.00, 421.00],
-            [1749, 421.00, 421.00, 421.00],
-            [1750, 421.00, 421.00, 421.00],
-            [1751, 422.00, 422.00, 422.00],
-            [1752, 422.00, 422.00, 422.00],
-            [1753, 422.00, 422.00, 422.00],
-            [1754, 423.00, 423.00, 423.00],
-            [1755, 423.00, 423.00, 423.00],
-            [1756, 423.00, 423.00, 423.00],
-            [1757, 424.00, 424.00, 424.00],
-            [1758, 424.00, 424.00, 424.00],
-            [1759, 424.00, 424.00, 424.00],
-            [1760, 425.00, 425.00, 425.00],
-            [1761, 425.00, 425.00, 425.00],
-            [1762, 425.00, 425.00, 425.00],
-            [1763, 426.00, 426.00, 426.00],
-            [1764, 426.00, 426.00, 426.00],
-            [1765, 427.00, 427.00, 427.00],
-            [1766, 427.00, 427.00, 427.00],
-            [1767, 427.00, 427.00, 427.00],
-            [1768, 428.00, 428.00, 428.00],
-            [1769, 428.00, 428.00, 428.00],
-            [1770, 428.00, 428.00, 428.00],
-            [1771, 429.00, 429.00, 429.00],
-            [1772, 429.00, 429.00, 429.00],
-            [1773, 429.00, 429.00, 429.00],
-            [1774, 430.00, 430.00, 430.00],
-            [1775, 430.00, 430.00, 430.00],
-            [1776, 430.00, 430.00, 430.00],
-            [1777, 431.00, 431.00, 431.00],
-            [1778, 431.00, 431.00, 431.00],
-            [1779, 431.00, 431.00, 431.00],
-            [1780, 432.00, 432.00, 432.00],
-            [1781, 432.00, 432.00, 432.00],
-            [1782, 432.00, 432.00, 432.00],
-            [1783, 433.00, 433.00, 433.00],
-            [1784, 433.00, 433.00, 433.00],
-            [1785, 433.00, 433.00, 433.00],
-            [1786, 434.00, 434.00, 434.00],
-            [1787, 434.00, 434.00, 434.00],
-            [1788, 434.00, 434.00, 434.00],
-            [1789, 435.00, 435.00, 435.00],
-            [1790, 435.00, 435.00, 435.00],
-            [1791, 435.00, 435.00, 435.00],
-            [1792, 436.00, 436.00, 436.00],
-            [1793, 436.00, 436.00, 436.00],
-            [1794, 437.00, 437.00, 437.00],
-            [1795, 437.00, 437.00, 437.00],
-            [1796, 437.00, 437.00, 437.00],
-            [1797, 438.00, 438.00, 438.00],
-            [1798, 438.00, 438.00, 438.00],
-            [1799, 438.00, 438.00, 438.00],
-            [1800, 439.00, 439.00, 439.00],
-            [1801, 439.00, 439.00, 439.00],
-            [1802, 439.00, 439.00, 439.00],
-            [1803, 440.00, 440.00, 440.00],
-            [1804, 440.00, 440.00, 440.00],
-            [1805, 440.00, 440.00, 440.00],
-            [1806, 441.00, 441.00, 441.00],
-            [1807, 441.00, 441.00, 441.00],
-            [1808, 441.00, 441.00, 441.00],
-            [1809, 442.00, 442.00, 442.00],
-            [1810, 442.00, 442.00, 442.00],
-            [1811, 442.00, 442.00, 442.00],
-            [1812, 443.00, 443.00, 443.00],
-            [1813, 443.00, 443.00, 443.00],
-            [1814, 443.00, 443.00, 443.00],
-            [1815, 444.00, 444.00, 444.00],
-            [1816, 444.00, 444.00, 444.00],
-            [1817, 444.00, 444.00, 444.00],
-            [1818, 445.00, 445.00, 445.00],
-            [1819, 445.00, 445.00, 445.00],
-            [1820, 445.00, 445.00, 445.00],
-            [1821, 446.00, 446.00, 446.00],
-            [1822, 446.00, 446.00, 446.00],
-            [1823, 447.00, 447.00, 447.00],
-            [1824, 447.00, 447.00, 447.00],
-            [1825, 447.00, 447.00, 447.00],
-            [1826, 448.00, 448.00, 448.00],
-            [1827, 448.00, 448.00, 448.00],
-            [1828, 448.00, 448.00, 448.00],
-            [1829, 449.00, 449.00, 449.00],
-            [1830, 449.00, 449.00, 449.00],
-            [1831, 449.00, 449.00, 449.00],
-            [1832, 450.00, 450.00, 450.00],
-            [1833, 450.00, 450.00, 450.00],
-            [1834, 450.00, 450.00, 450.00],
-            [1835, 451.00, 451.00, 451.00],
-            [1836, 451.00, 451.00, 451.00],
-            [1837, 451.00, 451.00, 451.00],
-            [1838, 452.00, 452.00, 452.00],
-            [1839, 452.00, 452.00, 452.00],
-            [1840, 452.00, 452.00, 452.00],
-            [1841, 453.00, 453.00, 453.00],
-            [1842, 453.00, 453.00, 453.00],
-            [1843, 453.00, 453.00, 453.00],
-            [1844, 454.00, 454.00, 454.00],
-            [1845, 454.00, 454.00, 454.00],
-            [1846, 454.00, 454.00, 454.00],
-            [1847, 455.00, 455.00, 455.00],
-            [1848, 455.00, 455.00, 455.00],
-            [1849, 455.00, 455.00, 455.00],
-            [1850, 456.00, 456.00, 456.00],
-            [1851, 456.00, 456.00, 456.00],
-            [1852, 457.00, 457.00, 457.00],
-            [1853, 457.00, 457.00, 457.00],
-            [1854, 457.00, 457.00, 457.00],
-            [1855, 458.00, 458.00, 458.00],
-            [1856, 458.00, 458.00, 458.00],
-            [1857, 458.00, 458.00, 458.00],
-            [1858, 459.00, 459.00, 459.00],
-            [1859, 459.00, 459.00, 459.00],
-            [1860, 459.00, 459.00, 459.00],
-            [1861, 460.00, 460.00, 460.00],
-            [1862, 460.00, 460.00, 460.00],
-            [1863, 460.00, 460.00, 460.00],
-            [1864, 461.00, 461.00, 461.00],
-            [1865, 461.00, 461.00, 461.00],
-            [1866, 461.00, 461.00, 461.00],
-            [1867, 462.00, 462.00, 462.00],
-            [1868, 462.00, 462.00, 462.00],
-            [1869, 462.00, 462.00, 462.00],
-            [1870, 463.00, 463.00, 463.00],
-            [1871, 463.00, 463.00, 463.00],
-            [1872, 463.00, 463.00, 463.00],
-            [1873, 464.00, 464.00, 464.00],
-            [1874, 464.00, 464.00, 464.00],
-            [1875, 464.00, 464.00, 464.00],
-            [1876, 465.00, 465.00, 465.00],
-            [1877, 465.00, 465.00, 465.00],
-            [1878, 466.00, 466.00, 466.00],
-            [1879, 466.00, 466.00, 466.00],
-            [1880, 466.00, 466.00, 466.00],
-            [1881, 467.00, 467.00, 467.00],
-            [1882, 467.00, 467.00, 467.00],
-            [1883, 467.00, 467.00, 467.00],
-            [1884, 468.00, 468.00, 468.00],
-            [1885, 468.00, 468.00, 468.00],
-            [1886, 468.00, 468.00, 468.00],
-            [1887, 469.00, 469.00, 469.00],
-            [1888, 469.00, 469.00, 469.00],
-            [1889, 469.00, 469.00, 469.00],
-            [1890, 470.00, 470.00, 470.00],
-            [1891, 470.00, 470.00, 470.00],
-            [1892, 470.00, 470.00, 470.00],
-            [1893, 471.00, 471.00, 471.00],
-            [1894, 471.00, 471.00, 471.00],
-            [1895, 471.00, 471.00, 471.00],
-            [1896, 472.00, 472.00, 472.00],
-            [1897, 472.00, 472.00, 472.00],
-            [1898, 472.00, 472.00, 472.00],
-            [1899, 473.00, 473.00, 473.00],
-            [1900, 473.00, 473.00, 473.00],
-            [1901, 473.00, 473.00, 473.00],
-            [1902, 474.00, 474.00, 474.00],
-            [1903, 474.00, 474.00, 474.00],
-            [1904, 474.00, 474.00, 474.00],
-            [1905, 475.00, 475.00, 475.00],
-            [1906, 475.00, 475.00, 475.00],
-            [1907, 476.00, 476.00, 476.00],
-            [1908, 476.00, 476.00, 476.00],
-            [1909, 476.00, 476.00, 476.00],
-            [1910, 477.00, 477.00, 477.00],
-            [1911, 477.00, 477.00, 477.00],
-            [1912, 477.00, 477.00, 477.00],
-            [1913, 478.00, 478.00, 478.00],
-            [1914, 478.00, 478.00, 478.00],
-            [1915, 478.00, 478.00, 478.00],
-            [1916, 479.00, 479.00, 479.00],
-            [1917, 479.00, 479.00, 479.00],
-            [1918, 479.00, 479.00, 479.00],
-            [1919, 480.00, 480.00, 480.00],
-            [1920, 480.00, 480.00, 480.00],
-            [1921, 480.00, 480.00, 480.00],
-            [1922, 481.00, 481.00, 481.00],
-            [1923, 481.00, 481.00, 481.00],
-            [1924, 481.00, 481.00, 481.00],
-            [1925, 482.00, 482.00, 482.00],
-            [1926, 482.00, 482.00, 482.00],
-            [1927, 482.00, 482.00, 482.00],
-            [1928, 483.00, 483.00, 483.00],
-            [1929, 483.00, 483.00, 483.00],
-            [1930, 483.00, 483.00, 483.00],
-            [1931, 484.00, 484.00, 484.00],
-            [1932, 484.00, 484.00, 484.00],
-            [1933, 484.00, 484.00, 484.00],
-            [1934, 485.00, 485.00, 485.00],
-            [1935, 485.00, 485.00, 485.00],
-            [1936, 486.00, 486.00, 486.00],
-            [1937, 486.00, 486.00, 486.00],
-            [1938, 486.00, 486.00, 486.00],
-            [1939, 487.00, 487.00, 487.00],
-            [1940, 487.00, 487.00, 487.00],
-            [1941, 487.00, 487.00, 487.00],
-            [1942, 488.00, 488.00, 488.00],
-            [1943, 488.00, 488.00, 488.00],
-            [1944, 488.00, 488.00, 488.00],
-            [1945, 489.00, 489.00, 489.00],
-            [1946, 489.00, 489.00, 489.00],
-            [1947, 489.00, 489.00, 489.00],
-            [1948, 490.00, 490.00, 490.00],
-            [1949, 490.00, 490.00, 490.00],
-            [1950, 490.00, 490.00, 490.00],
-            [1951, 491.00, 491.00, 491.00],
-            [1952, 491.00, 491.00, 491.00],
-            [1953, 491.00, 491.00, 491.00],
-            [1954, 492.00, 492.00, 492.00],
-            [1955, 492.00, 492.00, 492.00],
-            [1956, 492.00, 492.00, 492.00],
-            [1957, 493.00, 493.00, 493.00],
-            [1958, 493.00, 493.00, 493.00],
-            [1959, 493.00, 493.00, 493.00],
-            [1960, 494.00, 494.00, 494.00],
-            [1961, 494.00, 494.00, 494.00],
-            [1962, 494.00, 494.00, 494.00],
-            [1963, 495.00, 495.00, 495.00],
-            [1964, 495.00, 495.00, 495.00],
-            [1965, 496.00, 496.00, 496.00],
-            [1966, 496.00, 496.00, 496.00],
-            [1967, 496.00, 496.00, 496.00],
-            [1968, 497.00, 497.00, 497.00],
-            [1969, 497.00, 497.00, 497.00],
-            [1970, 497.00, 497.00, 497.00],
-            [1971, 498.00, 498.00, 498.00],
-            [1972, 498.00, 498.00, 498.00],
-            [1973, 498.00, 498.00, 498.00],
-            [1974, 499.00, 499.00, 499.00],
-            [1975, 499.00, 499.00, 499.00],
-            [1976, 499.00, 499.00, 499.00],
-            [1977, 500.00, 500.00, 500.00],
-            [1978, 500.00, 500.00, 500.00],
-            [1979, 500.00, 500.00, 500.00],
-            [1980, 501.00, 501.00, 501.00],
-            [1981, 501.00, 501.00, 501.00],
-            [1982, 501.00, 501.00, 501.00],
-            [1983, 502.00, 502.00, 502.00],
-            [1984, 502.00, 502.00, 502.00],
-            [1985, 502.00, 502.00, 502.00],
-            [1986, 503.00, 503.00, 503.00],
-            [1987, 503.00, 503.00, 503.00],
-            [1988, 503.00, 503.00, 503.00],
-            [1989, 504.00, 504.00, 504.00],
-            [1990, 504.00, 504.00, 504.00],
-            [1991, 504.00, 504.00, 504.00],
-            [1992, 505.00, 505.00, 505.00],
-            [1993, 505.00, 505.00, 505.00],
-            [1994, 506.00, 506.00, 506.00],
-            [1995, 506.00, 506.00, 506.00],
-            [1996, 506.00, 506.00, 506.00],
-            [1997, 507.00, 507.00, 507.00],
-            [1998, 507.00, 507.00, 507.00],
-            [1999, 507.00, 507.00, 507.00],
-            [2000, 508.00, 508.00, 508.00],
-            [2001, 508.00, 508.00, 508.00],
-            [2002, 508.00, 508.00, 508.00],
-            [2003, 509.00, 509.00, 509.00],
-            [2004, 509.00, 509.00, 509.00],
-            [2005, 509.00, 509.00, 509.00],
-            [2006, 510.00, 510.00, 510.00],
-            [2007, 510.00, 510.00, 510.00],
-            [2008, 510.00, 510.00, 510.00],
-            [2009, 511.00, 511.00, 511.00],
-            [2010, 511.00, 511.00, 511.00],
-            [2011, 511.00, 511.00, 511.00],
-            [2012, 512.00, 512.00, 512.00],
-            [2013, 512.00, 512.00, 512.00],
-            [2014, 512.00, 512.00, 512.00],
-            [2015, 513.00, 513.00, 513.00],
-            [2016, 513.00, 513.00, 513.00],
-            [2017, 513.00, 513.00, 513.00],
-            [2018, 514.00, 514.00, 514.00],
-            [2019, 514.00, 514.00, 514.00],
-            [2020, 514.00, 514.00, 514.00],
-            [2021, 515.00, 515.00, 515.00],
-            [2022, 515.00, 515.00, 515.00],
-            [2023, 516.00, 516.00, 516.00],
-            [2024, 516.00, 516.00, 516.00],
-            [2025, 516.00, 516.00, 516.00],
-            [2026, 517.00, 517.00, 517.00],
-            [2027, 517.00, 517.00, 517.00],
-            [2028, 517.00, 517.00, 517.00],
-            [2029, 518.00, 518.00, 518.00],
-            [2030, 518.00, 518.00, 518.00],
-            [2031, 518.00, 518.00, 518.00],
-            [2032, 519.00, 519.00, 519.00],
-            [2033, 519.00, 519.00, 519.00],
-            [2034, 519.00, 519.00, 519.00],
-            [2035, 520.00, 520.00, 520.00],
-            [2036, 520.00, 520.00, 520.00],
-            [2037, 520.00, 520.00, 520.00],
-            [2038, 521.00, 521.00, 521.00],
-            [2039, 521.00, 521.00, 521.00],
-            [2040, 521.00, 521.00, 521.00],
-            [2041, 522.00, 522.00, 522.00],
-            [2042, 522.00, 522.00, 522.00],
-            [2043, 522.00, 522.00, 522.00],
-            [2044, 523.00, 523.00, 523.00],
-            [2045, 523.00, 523.00, 523.00],
-            [2046, 523.00, 523.00, 523.00],
-            [2047, 524.00, 524.00, 524.00],
-            [2048, 524.00, 524.00, 524.00],
-            [2049, 524.00, 524.00, 524.00],
-            [2050, 525.00, 525.00, 525.00],
-            [2051, 525.00, 525.00, 525.00],
-            [2052, 526.00, 526.00, 526.00],
-            [2053, 526.00, 526.00, 526.00],
-            [2054, 526.00, 526.00, 526.00],
-            [2055, 527.00, 527.00, 527.00],
-            [2056, 527.00, 527.00, 527.00],
-            [2057, 527.00, 527.00, 527.00],
-            [2058, 528.00, 528.00, 528.00],
-            [2059, 528.00, 528.00, 528.00],
-            [2060, 528.00, 528.00, 528.00],
-            [2061, 529.00, 529.00, 529.00],
-            [2062, 529.00, 529.00, 529.00],
-            [2063, 529.00, 529.00, 529.00],
-            [2064, 530.00, 530.00, 530.00],
-            [2065, 530.00, 530.00, 530.00],
-            [2066, 530.00, 530.00, 530.00],
-            [2067, 531.00, 531.00, 531.00],
-            [2068, 531.00, 531.00, 531.00],
-            [2069, 531.00, 531.00, 531.00],
-            [2070, 532.00, 532.00, 532.00],
-            [2071, 532.00, 532.00, 532.00],
-            [2072, 532.00, 532.00, 532.00],
-            [2073, 533.00, 533.00, 533.00],
-            [2074, 533.00, 533.00, 533.00],
-            [2075, 533.00, 533.00, 533.00],
-            [2076, 534.00, 534.00, 534.00],
-            [2077, 534.00, 534.00, 534.00],
-            [2078, 535.00, 535.00, 535.00],
-            [2079, 535.00, 535.00, 535.00],
-            [2080, 535.00, 535.00, 535.00],
-            [2081, 536.00, 536.00, 536.00],
-            [2082, 536.00, 536.00, 536.00],
-            [2083, 536.00, 536.00, 536.00],
-            [2084, 537.00, 537.00, 537.00],
-            [2085, 537.00, 537.00, 537.00],
-            [2086, 537.00, 537.00, 537.00],
-            [2087, 538.00, 538.00, 538.00],
-            [2088, 538.00, 538.00, 538.00],
-            [2089, 538.00, 538.00, 538.00],
-            [2090, 539.00, 539.00, 539.00],
-            [2091, 539.00, 539.00, 539.00],
-            [2092, 539.00, 539.00, 539.00],
-            [2093, 540.00, 540.00, 540.00],
-            [2094, 540.00, 540.00, 540.00],
-            [2095, 540.00, 540.00, 540.00],
-            [2096, 541.00, 541.00, 541.00],
-            [2097, 541.00, 541.00, 541.00],
-            [2098, 541.00, 541.00, 541.00],
-            [2099, 542.00, 542.00, 542.00],
-            [2100, 542.00, 542.00, 542.00],
-            [2101, 542.00, 542.00, 542.00],
-            [2102, 543.00, 543.00, 543.00],
-            [2103, 543.00, 543.00, 543.00],
-            [2104, 543.00, 543.00, 543.00],
-            [2105, 544.00, 544.00, 544.00],
-            [2106, 544.00, 544.00, 544.00],
-            [2107, 545.00, 545.00, 545.00],
-            [2108, 545.00, 545.00, 545.00],
-            [2109, 545.00, 545.00, 545.00],
-            [2110, 546.00, 546.00, 546.00],
-            [2111, 546.00, 546.00, 546.00],
-            [2112, 546.00, 546.00, 546.00],
-            [2113, 547.00, 547.00, 547.00],
-            [2114, 547.00, 547.00, 547.00],
-            [2115, 547.00, 547.00, 547.00],
-            [2116, 548.00, 548.00, 548.00],
-            [2117, 548.00, 548.00, 548.00],
-            [2118, 548.00, 548.00, 548.00],
-            [2119, 549.00, 549.00, 549.00],
-            [2120, 549.00, 549.00, 549.00],
-            [2121, 549.00, 549.00, 549.00],
-            [2122, 550.00, 550.00, 550.00],
-            [2123, 550.00, 550.00, 550.00],
-            [2124, 550.00, 550.00, 550.00],
-            [2125, 551.00, 551.00, 551.00],
-            [2126, 551.00, 551.00, 551.00],
-            [2127, 551.00, 551.00, 551.00],
-            [2128, 552.00, 552.00, 552.00],
-            [2129, 552.00, 552.00, 552.00],
-            [2130, 552.00, 552.00, 552.00],
-            [2131, 553.00, 553.00, 553.00],
-            [2132, 553.00, 553.00, 553.00],
-            [2133, 553.00, 553.00, 553.00],
-            [2134, 554.00, 554.00, 554.00],
-            [2135, 554.00, 554.00, 554.00],
-            [2136, 555.00, 555.00, 555.00],
-            [2137, 555.00, 555.00, 555.00],
-            [2138, 555.00, 555.00, 555.00],
-            [2139, 556.00, 556.00, 556.00],
-            [2140, 556.00, 556.00, 556.00],
-            [2141, 556.00, 556.00, 556.00],
-            [2142, 557.00, 557.00, 557.00],
-            [2143, 557.00, 557.00, 557.00],
-            [2144, 557.00, 557.00, 557.00],
-            [2145, 558.00, 558.00, 558.00],
-            [2146, 558.00, 558.00, 558.00],
-            [2147, 558.00, 558.00, 558.00],
-            [2148, 559.00, 559.00, 559.00],
-            [2149, 559.00, 559.00, 559.00],
-            [2150, 559.00, 559.00, 559.00],
-            [2151, 560.00, 560.00, 560.00],
-            [2152, 560.00, 560.00, 560.00],
-            [2153, 560.00, 560.00, 560.00],
-            [2154, 561.00, 561.00, 561.00],
-            [2155, 561.00, 561.00, 561.00],
-            [2156, 561.00, 561.00, 561.00],
-            [2157, 562.00, 562.00, 562.00],
-            [2158, 562.00, 562.00, 562.00],
-            [2159, 562.00, 562.00, 562.00],
-            [2160, 563.00, 563.00, 563.00],
-            [2161, 563.00, 563.00, 563.00],
-            [2162, 563.00, 563.00, 563.00],
-            [2163, 564.00, 564.00, 564.00],
-            [2164, 564.00, 564.00, 564.00],
-            [2165, 565.00, 565.00, 565.00],
-            [2166, 565.00, 565.00, 565.00],
-            [2167, 565.00, 565.00, 565.00],
-            [2168, 566.00, 566.00, 566.00],
-            [2169, 566.00, 566.00, 566.00],
-            [2170, 566.00, 566.00, 566.00],
-            [2171, 567.00, 567.00, 567.00],
-            [2172, 567.00, 567.00, 567.00],
-            [2173, 567.00, 567.00, 567.00],
-            [2174, 568.00, 568.00, 568.00],
-            [2175, 568.00, 568.00, 568.00],
-            [2176, 568.00, 568.00, 568.00],
-            [2177, 569.00, 569.00, 569.00],
-            [2178, 569.00, 569.00, 569.00],
-            [2179, 569.00, 569.00, 569.00],
-            [2180, 570.00, 570.00, 570.00],
-            [2181, 570.00, 570.00, 570.00],
-            [2182, 570.00, 570.00, 570.00],
-            [2183, 571.00, 571.00, 571.00],
-            [2184, 571.00, 571.00, 571.00],
-            [2185, 571.00, 571.00, 571.00],
-            [2186, 572.00, 572.00, 572.00],
-            [2187, 572.00, 572.00, 572.00],
-            [2188, 572.00, 572.00, 572.00],
-            [2189, 573.00, 573.00, 573.00],
-            [2190, 573.00, 573.00, 573.00],
-            [2191, 573.00, 573.00, 573.00],
-            [2192, 574.00, 574.00, 574.00],
-            [2193, 574.00, 574.00, 574.00],
-            [2194, 575.00, 575.00, 575.00],
-            [2195, 575.00, 575.00, 575.00],
-            [2196, 575.00, 575.00, 575.00],
-            [2197, 576.00, 576.00, 576.00],
-            [2198, 576.00, 576.00, 576.00],
-            [2199, 576.00, 576.00, 576.00],
-            [2200, 577.00, 577.00, 577.00],
-            [2201, 577.00, 577.00, 577.00],
-            [2202, 577.00, 577.00, 577.00],
-            [2203, 578.00, 578.00, 578.00],
-            [2204, 578.00, 578.00, 578.00],
-            [2205, 578.00, 578.00, 578.00],
-            [2206, 579.00, 579.00, 579.00],
-            [2207, 579.00, 579.00, 579.00],
-            [2208, 579.00, 579.00, 579.00],
-            [2209, 580.00, 580.00, 580.00],
-            [2210, 580.00, 580.00, 580.00],
-            [2211, 580.00, 580.00, 580.00],
-            [2212, 581.00, 581.00, 581.00],
-            [2213, 581.00, 581.00, 581.00],
-            [2214, 581.00, 581.00, 581.00],
-            [2215, 582.00, 582.00, 582.00],
-            [2216, 582.00, 582.00, 582.00],
-            [2217, 582.00, 582.00, 582.00],
-            [2218, 583.00, 583.00, 583.00],
-            [2219, 583.00, 583.00, 583.00],
-            [2220, 583.00, 583.00, 583.00],
-            [2221, 584.00, 584.00, 584.00],
-            [2222, 584.00, 584.00, 584.00],
-            [2223, 585.00, 585.00, 585.00],
-            [2224, 585.00, 585.00, 585.00],
-            [2225, 585.00, 585.00, 585.00],
-            [2226, 586.00, 586.00, 586.00],
-            [2227, 586.00, 586.00, 586.00],
-            [2228, 586.00, 586.00, 586.00],
-            [2229, 587.00, 587.00, 587.00],
-            [2230, 587.00, 587.00, 587.00],
-            [2231, 587.00, 587.00, 587.00],
-            [2232, 588.00, 588.00, 588.00],
-            [2233, 588.00, 588.00, 588.00],
-            [2234, 588.00, 588.00, 588.00],
-            [2235, 589.00, 589.00, 589.00],
-            [2236, 589.00, 589.00, 589.00],
-            [2237, 589.00, 589.00, 589.00],
-            [2238, 590.00, 590.00, 590.00],
-            [2239, 590.00, 590.00, 590.00],
-            [2240, 590.00, 590.00, 590.00],
-            [2241, 591.00, 591.00, 591.00],
-            [2242, 591.00, 591.00, 591.00],
-            [2243, 591.00, 591.00, 591.00],
-            [2244, 592.00, 592.00, 592.00],
-            [2245, 592.00, 592.00, 592.00],
-            [2246, 592.00, 592.00, 592.00],
-            [2247, 593.00, 593.00, 593.00],
-            [2248, 593.00, 593.00, 593.00],
-            [2249, 593.00, 593.00, 593.00],
-            [2250, 594.00, 594.00, 594.00],
-            [2251, 594.00, 594.00, 594.00],
-            [2252, 595.00, 595.00, 595.00],
-            [2253, 595.00, 595.00, 595.00],
-            [2254, 595.00, 595.00, 595.00],
-            [2255, 596.00, 596.00, 596.00],
-            [2256, 596.00, 596.00, 596.00],
-            [2257, 596.00, 596.00, 596.00],
-            [2258, 597.00, 597.00, 597.00],
-            [2259, 597.00, 597.00, 597.00],
-            [2260, 597.00, 597.00, 597.00],
-            [2261, 598.00, 598.00, 598.00],
-            [2262, 598.00, 598.00, 598.00],
-            [2263, 598.00, 598.00, 598.00],
-            [2264, 599.00, 599.00, 599.00],
-            [2265, 599.00, 599.00, 599.00],
-            [2266, 599.00, 599.00, 599.00],
-            [2267, 600.00, 600.00, 600.00],
-            [2268, 600.00, 600.00, 600.00],
-            [2269, 600.00, 600.00, 600.00],
-            [2270, 601.00, 601.00, 601.00],
-            [2271, 601.00, 601.00, 601.00],
-            [2272, 601.00, 601.00, 601.00],
-            [2273, 602.00, 602.00, 602.00],
-            [2274, 602.00, 602.00, 602.00],
-            [2275, 602.00, 602.00, 602.00],
-            [2276, 603.00, 603.00, 603.00],
-            [2277, 603.00, 603.00, 603.00],
-            [2278, 604.00, 604.00, 604.00],
-            [2279, 604.00, 604.00, 604.00],
-            [2280, 604.00, 604.00, 604.00],
-            [2281, 605.00, 605.00, 605.00],
-            [2282, 605.00, 605.00, 605.00],
-            [2283, 605.00, 605.00, 605.00],
-            [2284, 606.00, 606.00, 606.00],
-            [2285, 606.00, 606.00, 606.00],
-            [2286, 606.00, 606.00, 606.00],
-            [2287, 607.00, 607.00, 607.00],
-            [2288, 607.00, 607.00, 607.00],
-            [2289, 607.00, 607.00, 607.00],
-            [2290, 608.00, 608.00, 608.00],
-            [2291, 608.00, 608.00, 608.00],
-            [2292, 608.00, 608.00, 608.00],
-            [2293, 609.00, 609.00, 609.00],
-            [2294, 609.00, 609.00, 609.00],
-            [2295, 609.00, 609.00, 609.00],
-            [2296, 610.00, 610.00, 610.00],
-            [2297, 610.00, 610.00, 610.00],
-            [2298, 610.00, 610.00, 610.00],
-            [2299, 611.00, 611.00, 611.00],
-            [2300, 611.00, 611.00, 611.00],
-            [2301, 611.00, 611.00, 611.00],
-            [2302, 612.00, 612.00, 612.00],
-            [2303, 612.00, 612.00, 612.00],
-            [2304, 612.00, 612.00, 612.00],
-            [2305, 613.00, 613.00, 613.00],
-            [2306, 613.00, 613.00, 613.00],
-            [2307, 614.00, 614.00, 614.00],
-            [2308, 614.00, 614.00, 614.00],
-            [2309, 614.00, 614.00, 614.00],
-            [2310, 615.00, 615.00, 615.00],
+            // [811, 76.00, 82.00, 96.00],
+            // [812, 77.00, 83.00, 97.00],
+            // [813, 77.00, 83.00, 97.00],
+            // [814, 78.00, 84.00, 98.00],
+            // [815, 78.00, 84.00, 98.00],
+            // [816, 79.00, 85.00, 99.00],
+            // [817, 79.00, 85.00, 99.00],
+            // [818, 79.00, 85.00, 99.00],
+            // [819, 80.00, 86.00, 100.00],
+            // [820, 80.00, 86.00, 100.00],
+            // [821, 81.00, 87.00, 101.00],
+            // [822, 81.00, 87.00, 101.00],
+            // [823, 81.00, 88.00, 102.00],
+            // [824, 82.00, 88.00, 102.00],
+            // [825, 82.00, 88.00, 102.00],
+            // [826, 83.00, 89.00, 103.00],
+            // [827, 83.00, 89.00, 103.00],
+            // [828, 84.00, 90.00, 104.00],
+            // [829, 84.00, 90.00, 104.00],
+            // [830, 84.00, 91.00, 104.00],
+            // [831, 85.00, 91.00, 105.00],
+            // [832, 85.00, 91.00, 105.00],
+            // [833, 86.00, 92.00, 105.00],
+            // [834, 86.00, 92.00, 105.00],
+            // [835, 87.00, 93.00, 106.00],
+            // [836, 87.00, 93.00, 106.00],
+            // [837, 87.00, 93.00, 106.00],
+            // [838, 88.00, 94.00, 107.00],
+            // [839, 88.00, 94.00, 107.00],
+            // [840, 89.00, 95.00, 107.00],
+            // [841, 89.00, 95.00, 108.00],
+            // [842, 90.00, 96.00, 108.00],
+            // [843, 90.00, 96.00, 108.00],
+            // [844, 90.00, 96.00, 108.00],
+            // [845, 91.00, 97.00, 109.00],
+            // [846, 91.00, 97.00, 109.00],
+            // [847, 92.00, 98.00, 109.00],
+            // [848, 92.00, 98.00, 110.00],
+            // [849, 93.00, 99.00, 110.00],
+            // [850, 93.00, 99.00, 110.00],
+            // [851, 93.00, 99.00, 111.00],
+            // [852, 94.00, 100.00, 111.00],
+            // [853, 94.00, 100.00, 111.00],
+            // [854, 95.00, 101.00, 111.00],
+            // [855, 95.00, 101.00, 112.00],
+            // [856, 95.00, 102.00, 112.00],
+            // [857, 96.00, 102.00, 112.00],
+            // [858, 96.00, 102.00, 113.00],
+            // [859, 97.00, 103.00, 113.00],
+            // [860, 97.00, 103.00, 113.00],
+            // [861, 98.00, 104.00, 114.00],
+            // [862, 98.00, 104.00, 114.00],
+            // [863, 98.00, 105.00, 114.00],
+            // [864, 99.00, 105.00, 114.00],
+            // [865, 99.00, 105.00, 115.00],
+            // [866, 100.00, 106.00, 115.00],
+            // [867, 100.00, 106.00, 116.00],
+            // [868, 101.00, 107.00, 116.00],
+            // [869, 101.00, 107.00, 116.00],
+            // [870, 102.00, 108.00, 117.00],
+            // [871, 102.00, 108.00, 117.00],
+            // [872, 103.00, 109.00, 117.00],
+            // [873, 103.00, 109.00, 118.00],
+            // [874, 104.00, 110.00, 118.00],
+            // [875, 104.00, 110.00, 118.00],
+            // [876, 105.00, 111.00, 119.00],
+            // [877, 105.00, 111.00, 119.00],
+            // [878, 106.00, 112.00, 119.00],
+            // [879, 106.00, 112.00, 120.00],
+            // [880, 107.00, 113.00, 120.00],
+            // [881, 107.00, 113.00, 120.00],
+            // [882, 107.00, 114.00, 121.00],
+            // [883, 108.00, 114.00, 121.00],
+            // [884, 108.00, 114.00, 121.00],
+            // [885, 109.00, 115.00, 122.00],
+            // [886, 109.00, 115.00, 122.00],
+            // [887, 110.00, 116.00, 123.00],
+            // [888, 110.00, 116.00, 123.00],
+            // [889, 111.00, 117.00, 123.00],
+            // [890, 111.00, 117.00, 124.00],
+            // [891, 112.00, 118.00, 124.00],
+            // [892, 112.00, 118.00, 124.00],
+            // [893, 113.00, 119.00, 125.00],
+            // [894, 113.00, 119.00, 125.00],
+            // [895, 114.00, 120.00, 125.00],
+            // [896, 114.00, 120.00, 126.00],
+            // [897, 115.00, 121.00, 126.00],
+            // [898, 115.00, 121.00, 126.00],
+            // [899, 116.00, 122.00, 127.00],
+            // [900, 116.00, 122.00, 127.00],
+            // [901, 116.00, 123.00, 127.00],
+            // [902, 117.00, 123.00, 128.00],
+            // [903, 117.00, 123.00, 128.00],
+            // [904, 118.00, 124.00, 128.00],
+            // [905, 118.00, 124.00, 129.00],
+            // [906, 119.00, 125.00, 129.00],
+            // [907, 119.00, 125.00, 129.00],
+            // [908, 120.00, 126.00, 130.00],
+            // [909, 120.00, 126.00, 130.00],
+            // [910, 121.00, 127.00, 131.00],
+            // [911, 121.00, 127.00, 131.00],
+            // [912, 122.00, 128.00, 131.00],
+            // [913, 122.00, 128.00, 132.00],
+            // [914, 123.00, 129.00, 132.00],
+            // [915, 123.00, 129.00, 132.00],
+            // [916, 124.00, 130.00, 133.00],
+            // [917, 124.00, 130.00, 133.00],
+            // [918, 124.00, 131.00, 133.00],
+            // [919, 125.00, 131.00, 134.00],
+            // [920, 125.00, 131.00, 134.00],
+            // [921, 126.00, 132.00, 134.00],
+            // [922, 126.00, 132.00, 135.00],
+            // [923, 127.00, 133.00, 135.00],
+            // [924, 127.00, 133.00, 135.00],
+            // [925, 128.00, 134.00, 136.00],
+            // [926, 128.00, 134.00, 136.00],
+            // [927, 129.00, 135.00, 136.00],
+            // [928, 129.00, 135.00, 137.00],
+            // [929, 130.00, 136.00, 137.00],
+            // [930, 130.00, 136.00, 137.00],
+            // [931, 131.00, 137.00, 138.00],
+            // [932, 131.00, 137.00, 138.00],
+            // [933, 132.00, 138.00, 139.00],
+            // [934, 132.00, 138.00, 139.00],
+            // [935, 133.00, 139.00, 139.00],
+            // [936, 133.00, 139.00, 140.00],
+            // [937, 133.00, 140.00, 140.00],
+            // [938, 134.00, 140.00, 140.00],
+            // [939, 134.00, 140.00, 141.00],
+            // [940, 135.00, 141.00, 141.00],
+            // [941, 135.00, 141.00, 141.00],
+            // [942, 136.00, 142.00, 142.00],
+            // [943, 136.00, 142.00, 142.00],
+            // [944, 137.00, 142.00, 142.00],
+            // [945, 137.00, 143.00, 143.00],
+            // [946, 138.00, 143.00, 143.00],
+            // [947, 138.00, 143.00, 143.00],
+            // [948, 139.00, 144.00, 144.00],
+            // [949, 139.00, 144.00, 144.00],
+            // [950, 140.00, 144.00, 144.00],
+            // [951, 140.00, 145.00, 145.00],
+            // [952, 141.00, 145.00, 145.00],
+            // [953, 141.00, 145.00, 145.00],
+            // [954, 142.00, 146.00, 146.00],
+            // [955, 142.00, 146.00, 146.00],
+            // [956, 142.00, 147.00, 147.00],
+            // [957, 143.00, 147.00, 147.00],
+            // [958, 143.00, 147.00, 147.00],
+            // [959, 144.00, 148.00, 148.00],
+            // [960, 144.00, 148.00, 148.00],
+            // [961, 145.00, 148.00, 148.00],
+            // [962, 145.00, 149.00, 149.00],
+            // [963, 146.00, 149.00, 149.00],
+            // [964, 146.00, 149.00, 149.00],
+            // [965, 147.00, 150.00, 150.00],
+            // [966, 147.00, 150.00, 150.00],
+            // [967, 148.00, 150.00, 150.00],
+            // [968, 148.00, 151.00, 151.00],
+            // [969, 149.00, 151.00, 151.00],
+            // [970, 149.00, 151.00, 151.00],
+            // [971, 150.00, 152.00, 152.00],
+            // [972, 150.00, 152.00, 152.00],
+            // [973, 150.00, 152.00, 152.00],
+            // [974, 151.00, 153.00, 153.00],
+            // [975, 151.00, 153.00, 153.00],
+            // [976, 152.00, 153.00, 153.00],
+            // [977, 152.00, 154.00, 154.00],
+            // [978, 153.00, 154.00, 154.00],
+            // [979, 153.00, 155.00, 155.00],
+            // [980, 154.00, 155.00, 155.00],
+            // [981, 154.00, 155.00, 155.00],
+            // [982, 155.00, 156.00, 156.00],
+            // [983, 155.00, 156.00, 156.00],
+            // [984, 156.00, 156.00, 156.00],
+            // [985, 156.00, 157.00, 157.00],
+            // [986, 157.00, 157.00, 157.00],
+            // [987, 157.00, 157.00, 157.00],
+            // [988, 158.00, 158.00, 158.00],
+            // [989, 158.00, 158.00, 158.00],
+            // [990, 158.00, 158.00, 158.00],
+            // [991, 159.00, 159.00, 159.00],
+            // [992, 159.00, 159.00, 159.00],
+            // [993, 159.00, 159.00, 159.00],
+            // [994, 160.00, 160.00, 160.00],
+            // [995, 160.00, 160.00, 160.00],
+            // [996, 160.00, 160.00, 160.00],
+            // [997, 161.00, 161.00, 161.00],
+            // [998, 161.00, 161.00, 161.00],
+            // [999, 161.00, 161.00, 161.00],
+            // [1000, 162.00, 162.00, 162.00],
+            // [1001, 162.00, 162.00, 162.00],
+            // [1002, 163.00, 163.00, 163.00],
+            // [1003, 163.00, 163.00, 163.00],
+            // [1004, 163.00, 163.00, 163.00],
+            // [1005, 164.00, 164.00, 164.00],
+            // [1006, 164.00, 164.00, 164.00],
+            // [1007, 164.00, 164.00, 164.00],
+            // [1008, 165.00, 165.00, 165.00],
+            // [1009, 165.00, 165.00, 165.00],
+            // [1010, 165.00, 165.00, 165.00],
+            // [1011, 166.00, 166.00, 166.00],
+            // [1012, 166.00, 166.00, 166.00],
+            // [1013, 166.00, 166.00, 166.00],
+            // [1014, 167.00, 167.00, 167.00],
+            // [1015, 167.00, 167.00, 167.00],
+            // [1016, 167.00, 167.00, 167.00],
+            // [1017, 168.00, 168.00, 168.00],
+            // [1018, 168.00, 168.00, 168.00],
+            // [1019, 168.00, 168.00, 168.00],
+            // [1020, 169.00, 169.00, 169.00],
+            // [1021, 169.00, 169.00, 169.00],
+            // [1022, 169.00, 169.00, 169.00],
+            // [1023, 170.00, 170.00, 170.00],
+            // [1024, 170.00, 170.00, 170.00],
+            // [1025, 171.00, 171.00, 171.00],
+            // [1026, 171.00, 171.00, 171.00],
+            // [1027, 171.00, 171.00, 171.00],
+            // [1028, 172.00, 172.00, 172.00],
+            // [1029, 172.00, 172.00, 172.00],
+            // [1030, 172.00, 172.00, 172.00],
+            // [1031, 173.00, 173.00, 173.00],
+            // [1032, 173.00, 173.00, 173.00],
+            // [1033, 173.00, 173.00, 173.00],
+            // [1034, 174.00, 174.00, 174.00],
+            // [1035, 174.00, 174.00, 174.00],
+            // [1036, 174.00, 174.00, 174.00],
+            // [1037, 175.00, 175.00, 175.00],
+            // [1038, 175.00, 175.00, 175.00],
+            // [1039, 175.00, 175.00, 175.00],
+            // [1040, 176.00, 176.00, 176.00],
+            // [1041, 176.00, 176.00, 176.00],
+            // [1042, 176.00, 176.00, 176.00],
+            // [1043, 177.00, 177.00, 177.00],
+            // [1044, 177.00, 177.00, 177.00],
+            // [1045, 177.00, 177.00, 177.00],
+            // [1046, 178.00, 178.00, 178.00],
+            // [1047, 178.00, 178.00, 178.00],
+            // [1048, 179.00, 179.00, 179.00],
+            // [1049, 179.00, 179.00, 179.00],
+            // [1050, 179.00, 179.00, 179.00],
+            // [1051, 180.00, 180.00, 180.00],
+            // [1052, 180.00, 180.00, 180.00],
+            // [1053, 180.00, 180.00, 180.00],
+            // [1054, 181.00, 181.00, 181.00],
+            // [1055, 181.00, 181.00, 181.00],
+            // [1056, 181.00, 181.00, 181.00],
+            // [1057, 182.00, 182.00, 182.00],
+            // [1058, 182.00, 182.00, 182.00],
+            // [1059, 182.00, 182.00, 182.00],
+            // [1060, 183.00, 183.00, 183.00],
+            // [1061, 183.00, 183.00, 183.00],
+            // [1062, 183.00, 183.00, 183.00],
+            // [1063, 184.00, 184.00, 184.00],
+            // [1064, 184.00, 184.00, 184.00],
+            // [1065, 184.00, 184.00, 184.00],
+            // [1066, 185.00, 185.00, 185.00],
+            // [1067, 185.00, 185.00, 185.00],
+            // [1068, 185.00, 185.00, 185.00],
+            // [1069, 186.00, 186.00, 186.00],
+            // [1070, 186.00, 186.00, 186.00],
+            // [1071, 187.00, 187.00, 187.00],
+            // [1072, 187.00, 187.00, 187.00],
+            // [1073, 187.00, 187.00, 187.00],
+            // [1074, 188.00, 188.00, 188.00],
+            // [1075, 188.00, 188.00, 188.00],
+            // [1076, 188.00, 188.00, 188.00],
+            // [1077, 189.00, 189.00, 189.00],
+            // [1078, 189.00, 189.00, 189.00],
+            // [1079, 189.00, 189.00, 189.00],
+            // [1080, 190.00, 190.00, 190.00],
+            // [1081, 190.00, 190.00, 190.00],
+            // [1082, 190.00, 190.00, 190.00],
+            // [1083, 191.00, 191.00, 191.00],
+            // [1084, 191.00, 191.00, 191.00],
+            // [1085, 191.00, 191.00, 191.00],
+            // [1086, 192.00, 192.00, 192.00],
+            // [1087, 192.00, 192.00, 192.00],
+            // [1088, 192.00, 192.00, 192.00],
+            // [1089, 193.00, 193.00, 193.00],
+            // [1090, 193.00, 193.00, 193.00],
+            // [1091, 193.00, 193.00, 193.00],
+            // [1092, 194.00, 194.00, 194.00],
+            // [1093, 194.00, 194.00, 194.00],
+            // [1094, 195.00, 195.00, 195.00],
+            // [1095, 195.00, 195.00, 195.00],
+            // [1096, 195.00, 195.00, 195.00],
+            // [1097, 196.00, 196.00, 196.00],
+            // [1098, 196.00, 196.00, 196.00],
+            // [1099, 196.00, 196.00, 196.00],
+            // [1100, 197.00, 197.00, 197.00],
+            // [1101, 197.00, 197.00, 197.00],
+            // [1102, 197.00, 197.00, 197.00],
+            // [1103, 198.00, 198.00, 198.00],
+            // [1104, 198.00, 198.00, 198.00],
+            // [1105, 198.00, 198.00, 198.00],
+            // [1106, 199.00, 199.00, 199.00],
+            // [1107, 199.00, 199.00, 199.00],
+            // [1108, 199.00, 199.00, 199.00],
+            // [1109, 200.00, 200.00, 200.00],
+            // [1110, 200.00, 200.00, 200.00],
+            // [1111, 200.00, 200.00, 200.00],
+            // [1112, 201.00, 201.00, 201.00],
+            // [1113, 201.00, 201.00, 201.00],
+            // [1114, 201.00, 201.00, 201.00],
+            // [1115, 202.00, 202.00, 202.00],
+            // [1116, 202.00, 202.00, 202.00],
+            // [1117, 203.00, 203.00, 203.00],
+            // [1118, 203.00, 203.00, 203.00],
+            // [1119, 203.00, 203.00, 203.00],
+            // [1120, 204.00, 204.00, 204.00],
+            // [1121, 204.00, 204.00, 204.00],
+            // [1122, 204.00, 204.00, 204.00],
+            // [1123, 205.00, 205.00, 205.00],
+            // [1124, 205.00, 205.00, 205.00],
+            // [1125, 205.00, 205.00, 205.00],
+            // [1126, 206.00, 206.00, 206.00],
+            // [1127, 206.00, 206.00, 206.00],
+            // [1128, 206.00, 206.00, 206.00],
+            // [1129, 207.00, 207.00, 207.00],
+            // [1130, 207.00, 207.00, 207.00],
+            // [1131, 207.00, 207.00, 207.00],
+            // [1132, 208.00, 208.00, 208.00],
+            // [1133, 208.00, 208.00, 208.00],
+            // [1134, 208.00, 208.00, 208.00],
+            // [1135, 209.00, 209.00, 209.00],
+            // [1136, 209.00, 209.00, 209.00],
+            // [1137, 209.00, 209.00, 209.00],
+            // [1138, 210.00, 210.00, 210.00],
+            // [1139, 210.00, 210.00, 210.00],
+            // [1140, 211.00, 211.00, 211.00],
+            // [1141, 211.00, 211.00, 211.00],
+            // [1142, 211.00, 211.00, 211.00],
+            // [1143, 212.00, 212.00, 212.00],
+            // [1144, 212.00, 212.00, 212.00],
+            // [1145, 212.00, 212.00, 212.00],
+            // [1146, 213.00, 213.00, 213.00],
+            // [1147, 213.00, 213.00, 213.00],
+            // [1148, 213.00, 213.00, 213.00],
+            // [1149, 214.00, 214.00, 214.00],
+            // [1150, 214.00, 214.00, 214.00],
+            // [1151, 214.00, 214.00, 214.00],
+            // [1152, 215.00, 215.00, 215.00],
+            // [1153, 215.00, 215.00, 215.00],
+            // [1154, 215.00, 215.00, 215.00],
+            // [1155, 216.00, 216.00, 216.00],
+            // [1156, 216.00, 216.00, 216.00],
+            // [1157, 216.00, 216.00, 216.00],
+            // [1158, 217.00, 217.00, 217.00],
+            // [1159, 217.00, 217.00, 217.00],
+            // [1160, 217.00, 217.00, 217.00],
+            // [1161, 218.00, 218.00, 218.00],
+            // [1162, 218.00, 218.00, 218.00],
+            // [1163, 219.00, 219.00, 219.00],
+            // [1164, 219.00, 219.00, 219.00],
+            // [1165, 219.00, 219.00, 219.00],
+            // [1166, 220.00, 220.00, 220.00],
+            // [1167, 220.00, 220.00, 220.00],
+            // [1168, 220.00, 220.00, 220.00],
+            // [1169, 221.00, 221.00, 221.00],
+            // [1170, 221.00, 221.00, 221.00],
+            // [1171, 221.00, 221.00, 221.00],
+            // [1172, 222.00, 222.00, 222.00],
+            // [1173, 222.00, 222.00, 222.00],
+            // [1174, 222.00, 222.00, 222.00],
+            // [1175, 223.00, 223.00, 223.00],
+            // [1176, 223.00, 223.00, 223.00],
+            // [1177, 223.00, 223.00, 223.00],
+            // [1178, 224.00, 224.00, 224.00],
+            // [1179, 224.00, 224.00, 224.00],
+            // [1180, 224.00, 224.00, 224.00],
+            // [1181, 225.00, 225.00, 225.00],
+            // [1182, 225.00, 225.00, 225.00],
+            // [1183, 225.00, 225.00, 225.00],
+            // [1184, 226.00, 226.00, 226.00],
+            // [1185, 226.00, 226.00, 226.00],
+            // [1186, 227.00, 227.00, 227.00],
+            // [1187, 227.00, 227.00, 227.00],
+            // [1188, 227.00, 227.00, 227.00],
+            // [1189, 228.00, 228.00, 228.00],
+            // [1190, 228.00, 228.00, 228.00],
+            // [1191, 228.00, 228.00, 228.00],
+            // [1192, 229.00, 229.00, 229.00],
+            // [1193, 229.00, 229.00, 229.00],
+            // [1194, 229.00, 229.00, 229.00],
+            // [1195, 230.00, 230.00, 230.00],
+            // [1196, 230.00, 230.00, 230.00],
+            // [1197, 230.00, 230.00, 230.00],
+            // [1198, 231.00, 231.00, 231.00],
+            // [1199, 231.00, 231.00, 231.00],
+            // [1200, 231.00, 231.00, 231.00],
+            // [1201, 232.00, 232.00, 232.00],
+            // [1202, 232.00, 232.00, 232.00],
+            // [1203, 232.00, 232.00, 232.00],
+            // [1204, 233.00, 233.00, 233.00],
+            // [1205, 233.00, 233.00, 233.00],
+            // [1206, 233.00, 233.00, 233.00],
+            // [1207, 234.00, 234.00, 234.00],
+            // [1208, 234.00, 234.00, 234.00],
+            // [1209, 235.00, 235.00, 235.00],
+            // [1210, 235.00, 235.00, 235.00],
+            // [1211, 235.00, 235.00, 235.00],
+            // [1212, 236.00, 236.00, 236.00],
+            // [1213, 236.00, 236.00, 236.00],
+            // [1214, 236.00, 236.00, 236.00],
+            // [1215, 237.00, 237.00, 237.00],
+            // [1216, 237.00, 237.00, 237.00],
+            // [1217, 237.00, 237.00, 237.00],
+            // [1218, 238.00, 238.00, 238.00],
+            // [1219, 238.00, 238.00, 238.00],
+            // [1220, 238.00, 238.00, 238.00],
+            // [1221, 239.00, 239.00, 239.00],
+            // [1222, 239.00, 239.00, 239.00],
+            // [1223, 239.00, 239.00, 239.00],
+            // [1224, 240.00, 240.00, 240.00],
+            // [1225, 240.00, 240.00, 240.00],
+            // [1226, 240.00, 240.00, 240.00],
+            // [1227, 241.00, 241.00, 241.00],
+            // [1228, 241.00, 241.00, 241.00],
+            // [1229, 241.00, 241.00, 241.00],
+            // [1230, 242.00, 242.00, 242.00],
+            // [1231, 242.00, 242.00, 242.00],
+            // [1232, 242.00, 242.00, 242.00],
+            // [1233, 243.00, 243.00, 243.00],
+            // [1234, 243.00, 243.00, 243.00],
+            // [1235, 244.00, 244.00, 244.00],
+            // [1236, 244.00, 244.00, 244.00],
+            // [1237, 244.00, 244.00, 244.00],
+            // [1238, 245.00, 245.00, 245.00],
+            // [1239, 245.00, 245.00, 245.00],
+            // [1240, 245.00, 245.00, 245.00],
+            // [1241, 246.00, 246.00, 246.00],
+            // [1242, 246.00, 246.00, 246.00],
+            // [1243, 246.00, 246.00, 246.00],
+            // [1244, 247.00, 247.00, 247.00],
+            // [1245, 247.00, 247.00, 247.00],
+            // [1246, 247.00, 247.00, 247.00],
+            // [1247, 248.00, 248.00, 248.00],
+            // [1248, 248.00, 248.00, 248.00],
+            // [1249, 248.00, 248.00, 248.00],
+            // [1250, 249.00, 249.00, 249.00],
+            // [1251, 249.00, 249.00, 249.00],
+            // [1252, 249.00, 249.00, 249.00],
+            // [1253, 250.00, 250.00, 250.00],
+            // [1254, 250.00, 250.00, 250.00],
+            // [1255, 250.00, 250.00, 250.00],
+            // [1256, 251.00, 251.00, 251.00],
+            // [1257, 251.00, 251.00, 251.00],
+            // [1258, 252.00, 252.00, 252.00],
+            // [1259, 252.00, 252.00, 252.00],
+            // [1260, 252.00, 252.00, 252.00],
+            // [1261, 253.00, 253.00, 253.00],
+            // [1262, 253.00, 253.00, 253.00],
+            // [1263, 253.00, 253.00, 253.00],
+            // [1264, 254.00, 254.00, 254.00],
+            // [1265, 254.00, 254.00, 254.00],
+            // [1266, 254.00, 254.00, 254.00],
+            // [1267, 255.00, 255.00, 255.00],
+            // [1268, 255.00, 255.00, 255.00],
+            // [1269, 255.00, 255.00, 255.00],
+            // [1270, 256.00, 256.00, 256.00],
+            // [1271, 256.00, 256.00, 256.00],
+            // [1272, 256.00, 256.00, 256.00],
+            // [1273, 257.00, 257.00, 257.00],
+            // [1274, 257.00, 257.00, 257.00],
+            // [1275, 257.00, 257.00, 257.00],
+            // [1276, 258.00, 258.00, 258.00],
+            // [1277, 258.00, 258.00, 258.00],
+            // [1278, 258.00, 258.00, 258.00],
+            // [1279, 259.00, 259.00, 259.00],
+            // [1280, 259.00, 259.00, 259.00],
+            // [1281, 260.00, 260.00, 260.00],
+            // [1282, 260.00, 260.00, 260.00],
+            // [1283, 260.00, 260.00, 260.00],
+            // [1284, 261.00, 261.00, 261.00],
+            // [1285, 261.00, 261.00, 261.00],
+            // [1286, 261.00, 261.00, 261.00],
+            // [1287, 262.00, 262.00, 262.00],
+            // [1288, 262.00, 262.00, 262.00],
+            // [1289, 262.00, 262.00, 262.00],
+            // [1290, 263.00, 263.00, 263.00],
+            // [1291, 263.00, 263.00, 263.00],
+            // [1292, 263.00, 263.00, 263.00],
+            // [1293, 264.00, 264.00, 264.00],
+            // [1294, 264.00, 264.00, 264.00],
+            // [1295, 264.00, 264.00, 264.00],
+            // [1296, 265.00, 265.00, 265.00],
+            // [1297, 265.00, 265.00, 265.00],
+            // [1298, 265.00, 265.00, 265.00],
+            // [1299, 266.00, 266.00, 266.00],
+            // [1300, 266.00, 266.00, 266.00],
+            // [1301, 266.00, 266.00, 266.00],
+            // [1302, 267.00, 267.00, 267.00],
+            // [1303, 267.00, 267.00, 267.00],
+            // [1304, 267.00, 267.00, 267.00],
+            // [1305, 268.00, 268.00, 268.00],
+            // [1306, 268.00, 268.00, 268.00],
+            // [1307, 269.00, 269.00, 269.00],
+            // [1308, 269.00, 269.00, 269.00],
+            // [1309, 269.00, 269.00, 269.00],
+            // [1310, 270.00, 270.00, 270.00],
+            // [1311, 270.00, 270.00, 270.00],
+            // [1312, 270.00, 270.00, 270.00],
+            // [1313, 271.00, 271.00, 271.00],
+            // [1314, 271.00, 271.00, 271.00],
+            // [1315, 271.00, 271.00, 271.00],
+            // [1316, 272.00, 272.00, 272.00],
+            // [1317, 272.00, 272.00, 272.00],
+            // [1318, 272.00, 272.00, 272.00],
+            // [1319, 273.00, 273.00, 273.00],
+            // [1320, 273.00, 273.00, 273.00],
+            // [1321, 273.00, 273.00, 273.00],
+            // [1322, 274.00, 274.00, 274.00],
+            // [1323, 274.00, 274.00, 274.00],
+            // [1324, 274.00, 274.00, 274.00],
+            // [1325, 275.00, 275.00, 275.00],
+            // [1326, 275.00, 275.00, 275.00],
+            // [1327, 275.00, 275.00, 275.00],
+            // [1328, 276.00, 276.00, 276.00],
+            // [1329, 276.00, 276.00, 276.00],
+            // [1330, 276.00, 276.00, 276.00],
+            // [1331, 277.00, 277.00, 277.00],
+            // [1332, 277.00, 277.00, 277.00],
+            // [1333, 277.00, 277.00, 277.00],
+            // [1334, 278.00, 278.00, 278.00],
+            // [1335, 278.00, 278.00, 278.00],
+            // [1336, 279.00, 279.00, 279.00],
+            // [1337, 279.00, 279.00, 279.00],
+            // [1338, 279.00, 279.00, 279.00],
+            // [1339, 280.00, 280.00, 280.00],
+            // [1340, 280.00, 280.00, 280.00],
+            // [1341, 280.00, 280.00, 280.00],
+            // [1342, 281.00, 281.00, 281.00],
+            // [1343, 281.00, 281.00, 281.00],
+            // [1344, 281.00, 281.00, 281.00],
+            // [1345, 282.00, 282.00, 282.00],
+            // [1346, 282.00, 282.00, 282.00],
+            // [1347, 282.00, 282.00, 282.00],
+            // [1348, 283.00, 283.00, 283.00],
+            // [1349, 283.00, 283.00, 283.00],
+            // [1350, 283.00, 283.00, 283.00],
+            // [1351, 284.00, 284.00, 284.00],
+            // [1352, 284.00, 284.00, 284.00],
+            // [1353, 284.00, 284.00, 284.00],
+            // [1354, 285.00, 285.00, 285.00],
+            // [1355, 285.00, 285.00, 285.00],
+            // [1356, 285.00, 285.00, 285.00],
+            // [1357, 286.00, 286.00, 286.00],
+            // [1358, 286.00, 286.00, 286.00],
+            // [1359, 286.00, 286.00, 286.00],
+            // [1360, 287.00, 287.00, 287.00],
+            // [1361, 287.00, 287.00, 287.00],
+            // [1362, 287.00, 287.00, 287.00],
+            // [1363, 288.00, 288.00, 288.00],
+            // [1364, 288.00, 288.00, 288.00],
+            // [1365, 289.00, 289.00, 289.00],
+            // [1366, 289.00, 289.00, 289.00],
+            // [1367, 289.00, 289.00, 289.00],
+            // [1368, 290.00, 290.00, 290.00],
+            // [1369, 290.00, 290.00, 290.00],
+            // [1370, 290.00, 290.00, 290.00],
+            // [1371, 291.00, 291.00, 291.00],
+            // [1372, 291.00, 291.00, 291.00],
+            // [1373, 291.00, 291.00, 291.00],
+            // [1374, 292.00, 292.00, 292.00],
+            // [1375, 292.00, 292.00, 292.00],
+            // [1376, 292.00, 292.00, 292.00],
+            // [1377, 293.00, 293.00, 293.00],
+            // [1378, 293.00, 293.00, 293.00],
+            // [1379, 293.00, 293.00, 293.00],
+            // [1380, 294.00, 294.00, 294.00],
+            // [1381, 294.00, 294.00, 294.00],
+            // [1382, 294.00, 294.00, 294.00],
+            // [1383, 295.00, 295.00, 295.00],
+            // [1384, 295.00, 295.00, 295.00],
+            // [1385, 295.00, 295.00, 295.00],
+            // [1386, 296.00, 296.00, 296.00],
+            // [1387, 296.00, 296.00, 296.00],
+            // [1388, 296.00, 296.00, 296.00],
+            // [1389, 297.00, 297.00, 297.00],
+            // [1390, 297.00, 297.00, 297.00],
+            // [1391, 297.00, 297.00, 297.00],
+            // [1392, 298.00, 298.00, 298.00],
+            // [1393, 298.00, 298.00, 298.00],
+            // [1394, 299.00, 299.00, 299.00],
+            // [1395, 299.00, 299.00, 299.00],
+            // [1396, 299.00, 299.00, 299.00],
+            // [1397, 300.00, 300.00, 300.00],
+            // [1398, 300.00, 300.00, 300.00],
+            // [1399, 300.00, 300.00, 300.00],
+            // [1400, 301.00, 301.00, 301.00],
+            // [1401, 301.00, 301.00, 301.00],
+            // [1402, 301.00, 301.00, 301.00],
+            // [1403, 302.00, 302.00, 302.00],
+            // [1404, 302.00, 302.00, 302.00],
+            // [1405, 302.00, 302.00, 302.00],
+            // [1406, 303.00, 303.00, 303.00],
+            // [1407, 303.00, 303.00, 303.00],
+            // [1408, 303.00, 303.00, 303.00],
+            // [1409, 304.00, 304.00, 304.00],
+            // [1410, 304.00, 304.00, 304.00],
+            // [1411, 304.00, 304.00, 304.00],
+            // [1412, 305.00, 305.00, 305.00],
+            // [1413, 305.00, 305.00, 305.00],
+            // [1414, 305.00, 305.00, 305.00],
+            // [1415, 306.00, 306.00, 306.00],
+            // [1416, 306.00, 306.00, 306.00],
+            // [1417, 306.00, 306.00, 306.00],
+            // [1418, 307.00, 307.00, 307.00],
+            // [1419, 307.00, 307.00, 307.00],
+            // [1420, 307.00, 307.00, 307.00],
+            // [1421, 308.00, 308.00, 308.00],
+            // [1422, 308.00, 308.00, 308.00],
+            // [1423, 309.00, 309.00, 309.00],
+            // [1424, 309.00, 309.00, 309.00],
+            // [1425, 309.00, 309.00, 309.00],
+            // [1426, 310.00, 310.00, 310.00],
+            // [1427, 310.00, 310.00, 310.00],
+            // [1428, 310.00, 310.00, 310.00],
+            // [1429, 311.00, 311.00, 311.00],
+            // [1430, 311.00, 311.00, 311.00],
+            // [1431, 311.00, 311.00, 311.00],
+            // [1432, 312.00, 312.00, 312.00],
+            // [1433, 312.00, 312.00, 312.00],
+            // [1434, 312.00, 312.00, 312.00],
+            // [1435, 313.00, 313.00, 313.00],
+            // [1436, 313.00, 313.00, 313.00],
+            // [1437, 313.00, 313.00, 313.00],
+            // [1438, 314.00, 314.00, 314.00],
+            // [1439, 314.00, 314.00, 314.00],
+            // [1440, 314.00, 314.00, 314.00],
+            // [1441, 315.00, 315.00, 315.00],
+            // [1442, 315.00, 315.00, 315.00],
+            // [1443, 315.00, 315.00, 315.00],
+            // [1444, 316.00, 316.00, 316.00],
+            // [1445, 316.00, 316.00, 316.00],
+            // [1446, 316.00, 316.00, 316.00],
+            // [1447, 317.00, 317.00, 317.00],
+            // [1448, 317.00, 317.00, 317.00],
+            // [1449, 317.00, 317.00, 317.00],
+            // [1450, 318.00, 318.00, 318.00],
+            // [1451, 318.00, 318.00, 318.00],
+            // [1452, 319.00, 319.00, 319.00],
+            // [1453, 319.00, 319.00, 319.00],
+            // [1454, 319.00, 319.00, 319.00],
+            // [1455, 320.00, 320.00, 320.00],
+            // [1456, 320.00, 320.00, 320.00],
+            // [1457, 320.00, 320.00, 320.00],
+            // [1458, 321.00, 321.00, 321.00],
+            // [1459, 321.00, 321.00, 321.00],
+            // [1460, 321.00, 321.00, 321.00],
+            // [1461, 322.00, 322.00, 322.00],
+            // [1462, 322.00, 322.00, 322.00],
+            // [1463, 322.00, 322.00, 322.00],
+            // [1464, 323.00, 323.00, 323.00],
+            // [1465, 323.00, 323.00, 323.00],
+            // [1466, 323.00, 323.00, 323.00],
+            // [1467, 324.00, 324.00, 324.00],
+            // [1468, 324.00, 324.00, 324.00],
+            // [1469, 324.00, 324.00, 324.00],
+            // [1470, 325.00, 325.00, 325.00],
+            // [1471, 325.00, 325.00, 325.00],
+            // [1472, 325.00, 325.00, 325.00],
+            // [1473, 326.00, 326.00, 326.00],
+            // [1474, 326.00, 326.00, 326.00],
+            // [1475, 326.00, 326.00, 326.00],
+            // [1476, 327.00, 327.00, 327.00],
+            // [1477, 327.00, 327.00, 327.00],
+            // [1478, 328.00, 328.00, 328.00],
+            // [1479, 328.00, 328.00, 328.00],
+            // [1480, 328.00, 328.00, 328.00],
+            // [1481, 329.00, 329.00, 329.00],
+            // [1482, 329.00, 329.00, 329.00],
+            // [1483, 329.00, 329.00, 329.00],
+            // [1484, 330.00, 330.00, 330.00],
+            // [1485, 330.00, 330.00, 330.00],
+            // [1486, 330.00, 330.00, 330.00],
+            // [1487, 331.00, 331.00, 331.00],
+            // [1488, 331.00, 331.00, 331.00],
+            // [1489, 331.00, 331.00, 331.00],
+            // [1490, 332.00, 332.00, 332.00],
+            // [1491, 332.00, 332.00, 332.00],
+            // [1492, 332.00, 332.00, 332.00],
+            // [1493, 333.00, 333.00, 333.00],
+            // [1494, 333.00, 333.00, 333.00],
+            // [1495, 333.00, 333.00, 333.00],
+            // [1496, 334.00, 334.00, 334.00],
+            // [1497, 334.00, 334.00, 334.00],
+            // [1498, 334.00, 334.00, 334.00],
+            // [1499, 335.00, 335.00, 335.00],
+            // [1500, 335.00, 335.00, 335.00],
+            // [1501, 335.00, 335.00, 335.00],
+            // [1502, 336.00, 336.00, 336.00],
+            // [1503, 336.00, 336.00, 336.00],
+            // [1504, 336.00, 336.00, 336.00],
+            // [1505, 337.00, 337.00, 337.00],
+            // [1506, 337.00, 337.00, 337.00],
+            // [1507, 338.00, 338.00, 338.00],
+            // [1508, 338.00, 338.00, 338.00],
+            // [1509, 338.00, 338.00, 338.00],
+            // [1510, 339.00, 339.00, 339.00],
+            // [1511, 339.00, 339.00, 339.00],
+            // [1512, 339.00, 339.00, 339.00],
+            // [1513, 340.00, 340.00, 340.00],
+            // [1514, 340.00, 340.00, 340.00],
+            // [1515, 340.00, 340.00, 340.00],
+            // [1516, 341.00, 341.00, 341.00],
+            // [1517, 341.00, 341.00, 341.00],
+            // [1518, 341.00, 341.00, 341.00],
+            // [1519, 342.00, 342.00, 342.00],
+            // [1520, 342.00, 342.00, 342.00],
+            // [1521, 342.00, 342.00, 342.00],
+            // [1522, 343.00, 343.00, 343.00],
+            // [1523, 343.00, 343.00, 343.00],
+            // [1524, 343.00, 343.00, 343.00],
+            // [1525, 344.00, 344.00, 344.00],
+            // [1526, 344.00, 344.00, 344.00],
+            // [1527, 344.00, 344.00, 344.00],
+            // [1528, 345.00, 345.00, 345.00],
+            // [1529, 345.00, 345.00, 345.00],
+            // [1530, 345.00, 345.00, 345.00],
+            // [1531, 346.00, 346.00, 346.00],
+            // [1532, 346.00, 346.00, 346.00],
+            // [1533, 346.00, 346.00, 346.00],
+            // [1534, 347.00, 347.00, 347.00],
+            // [1535, 347.00, 347.00, 347.00],
+            // [1536, 348.00, 348.00, 348.00],
+            // [1537, 348.00, 348.00, 348.00],
+            // [1538, 348.00, 348.00, 348.00],
+            // [1539, 349.00, 349.00, 349.00],
+            // [1540, 349.00, 349.00, 349.00],
+            // [1541, 349.00, 349.00, 349.00],
+            // [1542, 350.00, 350.00, 350.00],
+            // [1543, 350.00, 350.00, 350.00],
+            // [1544, 350.00, 350.00, 350.00],
+            // [1545, 351.00, 351.00, 351.00],
+            // [1546, 351.00, 351.00, 351.00],
+            // [1547, 351.00, 351.00, 351.00],
+            // [1548, 352.00, 352.00, 352.00],
+            // [1549, 352.00, 352.00, 352.00],
+            // [1550, 352.00, 352.00, 352.00],
+            // [1551, 353.00, 353.00, 353.00],
+            // [1552, 353.00, 353.00, 353.00],
+            // [1553, 353.00, 353.00, 353.00],
+            // [1554, 354.00, 354.00, 354.00],
+            // [1555, 354.00, 354.00, 354.00],
+            // [1556, 354.00, 354.00, 354.00],
+            // [1557, 355.00, 355.00, 355.00],
+            // [1558, 355.00, 355.00, 355.00],
+            // [1559, 355.00, 355.00, 355.00],
+            // [1560, 356.00, 356.00, 356.00],
+            // [1561, 356.00, 356.00, 356.00],
+            // [1562, 356.00, 356.00, 356.00],
+            // [1563, 357.00, 357.00, 357.00],
+            // [1564, 357.00, 357.00, 357.00],
+            // [1565, 358.00, 358.00, 358.00],
+            // [1566, 358.00, 358.00, 358.00],
+            // [1567, 358.00, 358.00, 358.00],
+            // [1568, 359.00, 359.00, 359.00],
+            // [1569, 359.00, 359.00, 359.00],
+            // [1570, 359.00, 359.00, 359.00],
+            // [1571, 360.00, 360.00, 360.00],
+            // [1572, 360.00, 360.00, 360.00],
+            // [1573, 360.00, 360.00, 360.00],
+            // [1574, 361.00, 361.00, 361.00],
+            // [1575, 361.00, 361.00, 361.00],
+            // [1576, 361.00, 361.00, 361.00],
+            // [1577, 362.00, 362.00, 362.00],
+            // [1578, 362.00, 362.00, 362.00],
+            // [1579, 362.00, 362.00, 362.00],
+            // [1580, 363.00, 363.00, 363.00],
+            // [1581, 363.00, 363.00, 363.00],
+            // [1582, 363.00, 363.00, 363.00],
+            // [1583, 364.00, 364.00, 364.00],
+            // [1584, 364.00, 364.00, 364.00],
+            // [1585, 364.00, 364.00, 364.00],
+            // [1586, 365.00, 365.00, 365.00],
+            // [1587, 365.00, 365.00, 365.00],
+            // [1588, 365.00, 365.00, 365.00],
+            // [1589, 366.00, 366.00, 366.00],
+            // [1590, 366.00, 366.00, 366.00],
+            // [1591, 366.00, 366.00, 366.00],
+            // [1592, 367.00, 367.00, 367.00],
+            // [1593, 367.00, 367.00, 367.00],
+            // [1594, 368.00, 368.00, 368.00],
+            // [1595, 368.00, 368.00, 368.00],
+            // [1596, 368.00, 368.00, 368.00],
+            // [1597, 369.00, 369.00, 369.00],
+            // [1598, 369.00, 369.00, 369.00],
+            // [1599, 369.00, 369.00, 369.00],
+            // [1600, 370.00, 370.00, 370.00],
+            // [1601, 370.00, 370.00, 370.00],
+            // [1602, 370.00, 370.00, 370.00],
+            // [1603, 371.00, 371.00, 371.00],
+            // [1604, 371.00, 371.00, 371.00],
+            // [1605, 371.00, 371.00, 371.00],
+            // [1606, 372.00, 372.00, 372.00],
+            // [1607, 372.00, 372.00, 372.00],
+            // [1608, 372.00, 372.00, 372.00],
+            // [1609, 373.00, 373.00, 373.00],
+            // [1610, 373.00, 373.00, 373.00],
+            // [1611, 373.00, 373.00, 373.00],
+            // [1612, 374.00, 374.00, 374.00],
+            // [1613, 374.00, 374.00, 374.00],
+            // [1614, 374.00, 374.00, 374.00],
+            // [1615, 375.00, 375.00, 375.00],
+            // [1616, 375.00, 375.00, 375.00],
+            // [1617, 375.00, 375.00, 375.00],
+            // [1618, 376.00, 376.00, 376.00],
+            // [1619, 376.00, 376.00, 376.00],
+            // [1620, 376.00, 376.00, 376.00],
+            // [1621, 377.00, 377.00, 377.00],
+            // [1622, 377.00, 377.00, 377.00],
+            // [1623, 378.00, 378.00, 378.00],
+            // [1624, 378.00, 378.00, 378.00],
+            // [1625, 378.00, 378.00, 378.00],
+            // [1626, 379.00, 379.00, 379.00],
+            // [1627, 379.00, 379.00, 379.00],
+            // [1628, 379.00, 379.00, 379.00],
+            // [1629, 380.00, 380.00, 380.00],
+            // [1630, 380.00, 380.00, 380.00],
+            // [1631, 380.00, 380.00, 380.00],
+            // [1632, 381.00, 381.00, 381.00],
+            // [1633, 381.00, 381.00, 381.00],
+            // [1634, 381.00, 381.00, 381.00],
+            // [1635, 382.00, 382.00, 382.00],
+            // [1636, 382.00, 382.00, 382.00],
+            // [1637, 382.00, 382.00, 382.00],
+            // [1638, 383.00, 383.00, 383.00],
+            // [1639, 383.00, 383.00, 383.00],
+            // [1640, 383.00, 383.00, 383.00],
+            // [1641, 384.00, 384.00, 384.00],
+            // [1642, 384.00, 384.00, 384.00],
+            // [1643, 384.00, 384.00, 384.00],
+            // [1644, 385.00, 385.00, 385.00],
+            // [1645, 385.00, 385.00, 385.00],
+            // [1646, 385.00, 385.00, 385.00],
+            // [1647, 386.00, 386.00, 386.00],
+            // [1648, 386.00, 386.00, 386.00],
+            // [1649, 386.00, 386.00, 386.00],
+            // [1650, 387.00, 387.00, 387.00],
+            // [1651, 387.00, 387.00, 387.00],
+            // [1652, 388.00, 388.00, 388.00],
+            // [1653, 388.00, 388.00, 388.00],
+            // [1654, 388.00, 388.00, 388.00],
+            // [1655, 389.00, 389.00, 389.00],
+            // [1656, 389.00, 389.00, 389.00],
+            // [1657, 389.00, 389.00, 389.00],
+            // [1658, 390.00, 390.00, 390.00],
+            // [1659, 390.00, 390.00, 390.00],
+            // [1660, 390.00, 390.00, 390.00],
+            // [1661, 391.00, 391.00, 391.00],
+            // [1662, 391.00, 391.00, 391.00],
+            // [1663, 391.00, 391.00, 391.00],
+            // [1664, 392.00, 392.00, 392.00],
+            // [1665, 392.00, 392.00, 392.00],
+            // [1666, 392.00, 392.00, 392.00],
+            // [1667, 393.00, 393.00, 393.00],
+            // [1668, 393.00, 393.00, 393.00],
+            // [1669, 393.00, 393.00, 393.00],
+            // [1670, 394.00, 394.00, 394.00],
+            // [1671, 394.00, 394.00, 394.00],
+            // [1672, 394.00, 394.00, 394.00],
+            // [1673, 395.00, 395.00, 395.00],
+            // [1674, 395.00, 395.00, 395.00],
+            // [1675, 395.00, 395.00, 395.00],
+            // [1676, 396.00, 396.00, 396.00],
+            // [1677, 396.00, 396.00, 396.00],
+            // [1678, 397.00, 397.00, 397.00],
+            // [1679, 397.00, 397.00, 397.00],
+            // [1680, 397.00, 397.00, 397.00],
+            // [1681, 398.00, 398.00, 398.00],
+            // [1682, 398.00, 398.00, 398.00],
+            // [1683, 398.00, 398.00, 398.00],
+            // [1684, 399.00, 399.00, 399.00],
+            // [1685, 399.00, 399.00, 399.00],
+            // [1686, 399.00, 399.00, 399.00],
+            // [1687, 400.00, 400.00, 400.00],
+            // [1688, 400.00, 400.00, 400.00],
+            // [1689, 400.00, 400.00, 400.00],
+            // [1690, 401.00, 401.00, 401.00],
+            // [1691, 401.00, 401.00, 401.00],
+            // [1692, 401.00, 401.00, 401.00],
+            // [1693, 402.00, 402.00, 402.00],
+            // [1694, 402.00, 402.00, 402.00],
+            // [1695, 402.00, 402.00, 402.00],
+            // [1696, 403.00, 403.00, 403.00],
+            // [1697, 403.00, 403.00, 403.00],
+            // [1698, 403.00, 403.00, 403.00],
+            // [1699, 404.00, 404.00, 404.00],
+            // [1700, 404.00, 404.00, 404.00],
+            // [1701, 404.00, 404.00, 404.00],
+            // [1702, 405.00, 405.00, 405.00],
+            // [1703, 405.00, 405.00, 405.00],
+            // [1704, 405.00, 405.00, 405.00],
+            // [1705, 406.00, 406.00, 406.00],
+            // [1706, 406.00, 406.00, 406.00],
+            // [1707, 407.00, 407.00, 407.00],
+            // [1708, 407.00, 407.00, 407.00],
+            // [1709, 407.00, 407.00, 407.00],
+            // [1710, 408.00, 408.00, 408.00],
+            // [1711, 408.00, 408.00, 408.00],
+            // [1712, 408.00, 408.00, 408.00],
+            // [1713, 409.00, 409.00, 409.00],
+            // [1714, 409.00, 409.00, 409.00],
+            // [1715, 409.00, 409.00, 409.00],
+            // [1716, 410.00, 410.00, 410.00],
+            // [1717, 410.00, 410.00, 410.00],
+            // [1718, 410.00, 410.00, 410.00],
+            // [1719, 411.00, 411.00, 411.00],
+            // [1720, 411.00, 411.00, 411.00],
+            // [1721, 411.00, 411.00, 411.00],
+            // [1722, 412.00, 412.00, 412.00],
+            // [1723, 412.00, 412.00, 412.00],
+            // [1724, 412.00, 412.00, 412.00],
+            // [1725, 413.00, 413.00, 413.00],
+            // [1726, 413.00, 413.00, 413.00],
+            // [1727, 413.00, 413.00, 413.00],
+            // [1728, 414.00, 414.00, 414.00],
+            // [1729, 414.00, 414.00, 414.00],
+            // [1730, 414.00, 414.00, 414.00],
+            // [1731, 415.00, 415.00, 415.00],
+            // [1732, 415.00, 415.00, 415.00],
+            // [1733, 415.00, 415.00, 415.00],
+            // [1734, 416.00, 416.00, 416.00],
+            // [1735, 416.00, 416.00, 416.00],
+            // [1736, 417.00, 417.00, 417.00],
+            // [1737, 417.00, 417.00, 417.00],
+            // [1738, 417.00, 417.00, 417.00],
+            // [1739, 418.00, 418.00, 418.00],
+            // [1740, 418.00, 418.00, 418.00],
+            // [1741, 418.00, 418.00, 418.00],
+            // [1742, 419.00, 419.00, 419.00],
+            // [1743, 419.00, 419.00, 419.00],
+            // [1744, 419.00, 419.00, 419.00],
+            // [1745, 420.00, 420.00, 420.00],
+            // [1746, 420.00, 420.00, 420.00],
+            // [1747, 420.00, 420.00, 420.00],
+            // [1748, 421.00, 421.00, 421.00],
+            // [1749, 421.00, 421.00, 421.00],
+            // [1750, 421.00, 421.00, 421.00],
+            // [1751, 422.00, 422.00, 422.00],
+            // [1752, 422.00, 422.00, 422.00],
+            // [1753, 422.00, 422.00, 422.00],
+            // [1754, 423.00, 423.00, 423.00],
+            // [1755, 423.00, 423.00, 423.00],
+            // [1756, 423.00, 423.00, 423.00],
+            // [1757, 424.00, 424.00, 424.00],
+            // [1758, 424.00, 424.00, 424.00],
+            // [1759, 424.00, 424.00, 424.00],
+            // [1760, 425.00, 425.00, 425.00],
+            // [1761, 425.00, 425.00, 425.00],
+            // [1762, 425.00, 425.00, 425.00],
+            // [1763, 426.00, 426.00, 426.00],
+            // [1764, 426.00, 426.00, 426.00],
+            // [1765, 427.00, 427.00, 427.00],
+            // [1766, 427.00, 427.00, 427.00],
+            // [1767, 427.00, 427.00, 427.00],
+            // [1768, 428.00, 428.00, 428.00],
+            // [1769, 428.00, 428.00, 428.00],
+            // [1770, 428.00, 428.00, 428.00],
+            // [1771, 429.00, 429.00, 429.00],
+            // [1772, 429.00, 429.00, 429.00],
+            // [1773, 429.00, 429.00, 429.00],
+            // [1774, 430.00, 430.00, 430.00],
+            // [1775, 430.00, 430.00, 430.00],
+            // [1776, 430.00, 430.00, 430.00],
+            // [1777, 431.00, 431.00, 431.00],
+            // [1778, 431.00, 431.00, 431.00],
+            // [1779, 431.00, 431.00, 431.00],
+            // [1780, 432.00, 432.00, 432.00],
+            // [1781, 432.00, 432.00, 432.00],
+            // [1782, 432.00, 432.00, 432.00],
+            // [1783, 433.00, 433.00, 433.00],
+            // [1784, 433.00, 433.00, 433.00],
+            // [1785, 433.00, 433.00, 433.00],
+            // [1786, 434.00, 434.00, 434.00],
+            // [1787, 434.00, 434.00, 434.00],
+            // [1788, 434.00, 434.00, 434.00],
+            // [1789, 435.00, 435.00, 435.00],
+            // [1790, 435.00, 435.00, 435.00],
+            // [1791, 435.00, 435.00, 435.00],
+            // [1792, 436.00, 436.00, 436.00],
+            // [1793, 436.00, 436.00, 436.00],
+            // [1794, 437.00, 437.00, 437.00],
+            // [1795, 437.00, 437.00, 437.00],
+            // [1796, 437.00, 437.00, 437.00],
+            // [1797, 438.00, 438.00, 438.00],
+            // [1798, 438.00, 438.00, 438.00],
+            // [1799, 438.00, 438.00, 438.00],
+            // [1800, 439.00, 439.00, 439.00],
+            // [1801, 439.00, 439.00, 439.00],
+            // [1802, 439.00, 439.00, 439.00],
+            // [1803, 440.00, 440.00, 440.00],
+            // [1804, 440.00, 440.00, 440.00],
+            // [1805, 440.00, 440.00, 440.00],
+            // [1806, 441.00, 441.00, 441.00],
+            // [1807, 441.00, 441.00, 441.00],
+            // [1808, 441.00, 441.00, 441.00],
+            // [1809, 442.00, 442.00, 442.00],
+            // [1810, 442.00, 442.00, 442.00],
+            // [1811, 442.00, 442.00, 442.00],
+            // [1812, 443.00, 443.00, 443.00],
+            // [1813, 443.00, 443.00, 443.00],
+            // [1814, 443.00, 443.00, 443.00],
+            // [1815, 444.00, 444.00, 444.00],
+            // [1816, 444.00, 444.00, 444.00],
+            // [1817, 444.00, 444.00, 444.00],
+            // [1818, 445.00, 445.00, 445.00],
+            // [1819, 445.00, 445.00, 445.00],
+            // [1820, 445.00, 445.00, 445.00],
+            // [1821, 446.00, 446.00, 446.00],
+            // [1822, 446.00, 446.00, 446.00],
+            // [1823, 447.00, 447.00, 447.00],
+            // [1824, 447.00, 447.00, 447.00],
+            // [1825, 447.00, 447.00, 447.00],
+            // [1826, 448.00, 448.00, 448.00],
+            // [1827, 448.00, 448.00, 448.00],
+            // [1828, 448.00, 448.00, 448.00],
+            // [1829, 449.00, 449.00, 449.00],
+            // [1830, 449.00, 449.00, 449.00],
+            // [1831, 449.00, 449.00, 449.00],
+            // [1832, 450.00, 450.00, 450.00],
+            // [1833, 450.00, 450.00, 450.00],
+            // [1834, 450.00, 450.00, 450.00],
+            // [1835, 451.00, 451.00, 451.00],
+            // [1836, 451.00, 451.00, 451.00],
+            // [1837, 451.00, 451.00, 451.00],
+            // [1838, 452.00, 452.00, 452.00],
+            // [1839, 452.00, 452.00, 452.00],
+            // [1840, 452.00, 452.00, 452.00],
+            // [1841, 453.00, 453.00, 453.00],
+            // [1842, 453.00, 453.00, 453.00],
+            // [1843, 453.00, 453.00, 453.00],
+            // [1844, 454.00, 454.00, 454.00],
+            // [1845, 454.00, 454.00, 454.00],
+            // [1846, 454.00, 454.00, 454.00],
+            // [1847, 455.00, 455.00, 455.00],
+            // [1848, 455.00, 455.00, 455.00],
+            // [1849, 455.00, 455.00, 455.00],
+            // [1850, 456.00, 456.00, 456.00],
+            // [1851, 456.00, 456.00, 456.00],
+            // [1852, 457.00, 457.00, 457.00],
+            // [1853, 457.00, 457.00, 457.00],
+            // [1854, 457.00, 457.00, 457.00],
+            // [1855, 458.00, 458.00, 458.00],
+            // [1856, 458.00, 458.00, 458.00],
+            // [1857, 458.00, 458.00, 458.00],
+            // [1858, 459.00, 459.00, 459.00],
+            // [1859, 459.00, 459.00, 459.00],
+            // [1860, 459.00, 459.00, 459.00],
+            // [1861, 460.00, 460.00, 460.00],
+            // [1862, 460.00, 460.00, 460.00],
+            // [1863, 460.00, 460.00, 460.00],
+            // [1864, 461.00, 461.00, 461.00],
+            // [1865, 461.00, 461.00, 461.00],
+            // [1866, 461.00, 461.00, 461.00],
+            // [1867, 462.00, 462.00, 462.00],
+            // [1868, 462.00, 462.00, 462.00],
+            // [1869, 462.00, 462.00, 462.00],
+            // [1870, 463.00, 463.00, 463.00],
+            // [1871, 463.00, 463.00, 463.00],
+            // [1872, 463.00, 463.00, 463.00],
+            // [1873, 464.00, 464.00, 464.00],
+            // [1874, 464.00, 464.00, 464.00],
+            // [1875, 464.00, 464.00, 464.00],
+            // [1876, 465.00, 465.00, 465.00],
+            // [1877, 465.00, 465.00, 465.00],
+            // [1878, 466.00, 466.00, 466.00],
+            // [1879, 466.00, 466.00, 466.00],
+            // [1880, 466.00, 466.00, 466.00],
+            // [1881, 467.00, 467.00, 467.00],
+            // [1882, 467.00, 467.00, 467.00],
+            // [1883, 467.00, 467.00, 467.00],
+            // [1884, 468.00, 468.00, 468.00],
+            // [1885, 468.00, 468.00, 468.00],
+            // [1886, 468.00, 468.00, 468.00],
+            // [1887, 469.00, 469.00, 469.00],
+            // [1888, 469.00, 469.00, 469.00],
+            // [1889, 469.00, 469.00, 469.00],
+            // [1890, 470.00, 470.00, 470.00],
+            // [1891, 470.00, 470.00, 470.00],
+            // [1892, 470.00, 470.00, 470.00],
+            // [1893, 471.00, 471.00, 471.00],
+            // [1894, 471.00, 471.00, 471.00],
+            // [1895, 471.00, 471.00, 471.00],
+            // [1896, 472.00, 472.00, 472.00],
+            // [1897, 472.00, 472.00, 472.00],
+            // [1898, 472.00, 472.00, 472.00],
+            // [1899, 473.00, 473.00, 473.00],
+            // [1900, 473.00, 473.00, 473.00],
+            // [1901, 473.00, 473.00, 473.00],
+            // [1902, 474.00, 474.00, 474.00],
+            // [1903, 474.00, 474.00, 474.00],
+            // [1904, 474.00, 474.00, 474.00],
+            // [1905, 475.00, 475.00, 475.00],
+            // [1906, 475.00, 475.00, 475.00],
+            // [1907, 476.00, 476.00, 476.00],
+            // [1908, 476.00, 476.00, 476.00],
+            // [1909, 476.00, 476.00, 476.00],
+            // [1910, 477.00, 477.00, 477.00],
+            // [1911, 477.00, 477.00, 477.00],
+            // [1912, 477.00, 477.00, 477.00],
+            // [1913, 478.00, 478.00, 478.00],
+            // [1914, 478.00, 478.00, 478.00],
+            // [1915, 478.00, 478.00, 478.00],
+            // [1916, 479.00, 479.00, 479.00],
+            // [1917, 479.00, 479.00, 479.00],
+            // [1918, 479.00, 479.00, 479.00],
+            // [1919, 480.00, 480.00, 480.00],
+            // [1920, 480.00, 480.00, 480.00],
+            // [1921, 480.00, 480.00, 480.00],
+            // [1922, 481.00, 481.00, 481.00],
+            // [1923, 481.00, 481.00, 481.00],
+            // [1924, 481.00, 481.00, 481.00],
+            // [1925, 482.00, 482.00, 482.00],
+            // [1926, 482.00, 482.00, 482.00],
+            // [1927, 482.00, 482.00, 482.00],
+            // [1928, 483.00, 483.00, 483.00],
+            // [1929, 483.00, 483.00, 483.00],
+            // [1930, 483.00, 483.00, 483.00],
+            // [1931, 484.00, 484.00, 484.00],
+            // [1932, 484.00, 484.00, 484.00],
+            // [1933, 484.00, 484.00, 484.00],
+            // [1934, 485.00, 485.00, 485.00],
+            // [1935, 485.00, 485.00, 485.00],
+            // [1936, 486.00, 486.00, 486.00],
+            // [1937, 486.00, 486.00, 486.00],
+            // [1938, 486.00, 486.00, 486.00],
+            // [1939, 487.00, 487.00, 487.00],
+            // [1940, 487.00, 487.00, 487.00],
+            // [1941, 487.00, 487.00, 487.00],
+            // [1942, 488.00, 488.00, 488.00],
+            // [1943, 488.00, 488.00, 488.00],
+            // [1944, 488.00, 488.00, 488.00],
+            // [1945, 489.00, 489.00, 489.00],
+            // [1946, 489.00, 489.00, 489.00],
+            // [1947, 489.00, 489.00, 489.00],
+            // [1948, 490.00, 490.00, 490.00],
+            // [1949, 490.00, 490.00, 490.00],
+            // [1950, 490.00, 490.00, 490.00],
+            // [1951, 491.00, 491.00, 491.00],
+            // [1952, 491.00, 491.00, 491.00],
+            // [1953, 491.00, 491.00, 491.00],
+            // [1954, 492.00, 492.00, 492.00],
+            // [1955, 492.00, 492.00, 492.00],
+            // [1956, 492.00, 492.00, 492.00],
+            // [1957, 493.00, 493.00, 493.00],
+            // [1958, 493.00, 493.00, 493.00],
+            // [1959, 493.00, 493.00, 493.00],
+            // [1960, 494.00, 494.00, 494.00],
+            // [1961, 494.00, 494.00, 494.00],
+            // [1962, 494.00, 494.00, 494.00],
+            // [1963, 495.00, 495.00, 495.00],
+            // [1964, 495.00, 495.00, 495.00],
+            // [1965, 496.00, 496.00, 496.00],
+            // [1966, 496.00, 496.00, 496.00],
+            // [1967, 496.00, 496.00, 496.00],
+            // [1968, 497.00, 497.00, 497.00],
+            // [1969, 497.00, 497.00, 497.00],
+            // [1970, 497.00, 497.00, 497.00],
+            // [1971, 498.00, 498.00, 498.00],
+            // [1972, 498.00, 498.00, 498.00],
+            // [1973, 498.00, 498.00, 498.00],
+            // [1974, 499.00, 499.00, 499.00],
+            // [1975, 499.00, 499.00, 499.00],
+            // [1976, 499.00, 499.00, 499.00],
+            // [1977, 500.00, 500.00, 500.00],
+            // [1978, 500.00, 500.00, 500.00],
+            // [1979, 500.00, 500.00, 500.00],
+            // [1980, 501.00, 501.00, 501.00],
+            // [1981, 501.00, 501.00, 501.00],
+            // [1982, 501.00, 501.00, 501.00],
+            // [1983, 502.00, 502.00, 502.00],
+            // [1984, 502.00, 502.00, 502.00],
+            // [1985, 502.00, 502.00, 502.00],
+            // [1986, 503.00, 503.00, 503.00],
+            // [1987, 503.00, 503.00, 503.00],
+            // [1988, 503.00, 503.00, 503.00],
+            // [1989, 504.00, 504.00, 504.00],
+            // [1990, 504.00, 504.00, 504.00],
+            // [1991, 504.00, 504.00, 504.00],
+            // [1992, 505.00, 505.00, 505.00],
+            // [1993, 505.00, 505.00, 505.00],
+            // [1994, 506.00, 506.00, 506.00],
+            // [1995, 506.00, 506.00, 506.00],
+            // [1996, 506.00, 506.00, 506.00],
+            // [1997, 507.00, 507.00, 507.00],
+            // [1998, 507.00, 507.00, 507.00],
+            // [1999, 507.00, 507.00, 507.00],
+            // [2000, 508.00, 508.00, 508.00],
+            // [2001, 508.00, 508.00, 508.00],
+            // [2002, 508.00, 508.00, 508.00],
+            // [2003, 509.00, 509.00, 509.00],
+            // [2004, 509.00, 509.00, 509.00],
+            // [2005, 509.00, 509.00, 509.00],
+            // [2006, 510.00, 510.00, 510.00],
+            // [2007, 510.00, 510.00, 510.00],
+            // [2008, 510.00, 510.00, 510.00],
+            // [2009, 511.00, 511.00, 511.00],
+            // [2010, 511.00, 511.00, 511.00],
+            // [2011, 511.00, 511.00, 511.00],
+            // [2012, 512.00, 512.00, 512.00],
+            // [2013, 512.00, 512.00, 512.00],
+            // [2014, 512.00, 512.00, 512.00],
+            // [2015, 513.00, 513.00, 513.00],
+            // [2016, 513.00, 513.00, 513.00],
+            // [2017, 513.00, 513.00, 513.00],
+            // [2018, 514.00, 514.00, 514.00],
+            // [2019, 514.00, 514.00, 514.00],
+            // [2020, 514.00, 514.00, 514.00],
+            // [2021, 515.00, 515.00, 515.00],
+            // [2022, 515.00, 515.00, 515.00],
+            // [2023, 516.00, 516.00, 516.00],
+            // [2024, 516.00, 516.00, 516.00],
+            // [2025, 516.00, 516.00, 516.00],
+            // [2026, 517.00, 517.00, 517.00],
+            // [2027, 517.00, 517.00, 517.00],
+            // [2028, 517.00, 517.00, 517.00],
+            // [2029, 518.00, 518.00, 518.00],
+            // [2030, 518.00, 518.00, 518.00],
+            // [2031, 518.00, 518.00, 518.00],
+            // [2032, 519.00, 519.00, 519.00],
+            // [2033, 519.00, 519.00, 519.00],
+            // [2034, 519.00, 519.00, 519.00],
+            // [2035, 520.00, 520.00, 520.00],
+            // [2036, 520.00, 520.00, 520.00],
+            // [2037, 520.00, 520.00, 520.00],
+            // [2038, 521.00, 521.00, 521.00],
+            // [2039, 521.00, 521.00, 521.00],
+            // [2040, 521.00, 521.00, 521.00],
+            // [2041, 522.00, 522.00, 522.00],
+            // [2042, 522.00, 522.00, 522.00],
+            // [2043, 522.00, 522.00, 522.00],
+            // [2044, 523.00, 523.00, 523.00],
+            // [2045, 523.00, 523.00, 523.00],
+            // [2046, 523.00, 523.00, 523.00],
+            // [2047, 524.00, 524.00, 524.00],
+            // [2048, 524.00, 524.00, 524.00],
+            // [2049, 524.00, 524.00, 524.00],
+            // [2050, 525.00, 525.00, 525.00],
+            // [2051, 525.00, 525.00, 525.00],
+            // [2052, 526.00, 526.00, 526.00],
+            // [2053, 526.00, 526.00, 526.00],
+            // [2054, 526.00, 526.00, 526.00],
+            // [2055, 527.00, 527.00, 527.00],
+            // [2056, 527.00, 527.00, 527.00],
+            // [2057, 527.00, 527.00, 527.00],
+            // [2058, 528.00, 528.00, 528.00],
+            // [2059, 528.00, 528.00, 528.00],
+            // [2060, 528.00, 528.00, 528.00],
+            // [2061, 529.00, 529.00, 529.00],
+            // [2062, 529.00, 529.00, 529.00],
+            // [2063, 529.00, 529.00, 529.00],
+            // [2064, 530.00, 530.00, 530.00],
+            // [2065, 530.00, 530.00, 530.00],
+            // [2066, 530.00, 530.00, 530.00],
+            // [2067, 531.00, 531.00, 531.00],
+            // [2068, 531.00, 531.00, 531.00],
+            // [2069, 531.00, 531.00, 531.00],
+            // [2070, 532.00, 532.00, 532.00],
+            // [2071, 532.00, 532.00, 532.00],
+            // [2072, 532.00, 532.00, 532.00],
+            // [2073, 533.00, 533.00, 533.00],
+            // [2074, 533.00, 533.00, 533.00],
+            // [2075, 533.00, 533.00, 533.00],
+            // [2076, 534.00, 534.00, 534.00],
+            // [2077, 534.00, 534.00, 534.00],
+            // [2078, 535.00, 535.00, 535.00],
+            // [2079, 535.00, 535.00, 535.00],
+            // [2080, 535.00, 535.00, 535.00],
+            // [2081, 536.00, 536.00, 536.00],
+            // [2082, 536.00, 536.00, 536.00],
+            // [2083, 536.00, 536.00, 536.00],
+            // [2084, 537.00, 537.00, 537.00],
+            // [2085, 537.00, 537.00, 537.00],
+            // [2086, 537.00, 537.00, 537.00],
+            // [2087, 538.00, 538.00, 538.00],
+            // [2088, 538.00, 538.00, 538.00],
+            // [2089, 538.00, 538.00, 538.00],
+            // [2090, 539.00, 539.00, 539.00],
+            // [2091, 539.00, 539.00, 539.00],
+            // [2092, 539.00, 539.00, 539.00],
+            // [2093, 540.00, 540.00, 540.00],
+            // [2094, 540.00, 540.00, 540.00],
+            // [2095, 540.00, 540.00, 540.00],
+            // [2096, 541.00, 541.00, 541.00],
+            // [2097, 541.00, 541.00, 541.00],
+            // [2098, 541.00, 541.00, 541.00],
+            // [2099, 542.00, 542.00, 542.00],
+            // [2100, 542.00, 542.00, 542.00],
+            // [2101, 542.00, 542.00, 542.00],
+            // [2102, 543.00, 543.00, 543.00],
+            // [2103, 543.00, 543.00, 543.00],
+            // [2104, 543.00, 543.00, 543.00],
+            // [2105, 544.00, 544.00, 544.00],
+            // [2106, 544.00, 544.00, 544.00],
+            // [2107, 545.00, 545.00, 545.00],
+            // [2108, 545.00, 545.00, 545.00],
+            // [2109, 545.00, 545.00, 545.00],
+            // [2110, 546.00, 546.00, 546.00],
+            // [2111, 546.00, 546.00, 546.00],
+            // [2112, 546.00, 546.00, 546.00],
+            // [2113, 547.00, 547.00, 547.00],
+            // [2114, 547.00, 547.00, 547.00],
+            // [2115, 547.00, 547.00, 547.00],
+            // [2116, 548.00, 548.00, 548.00],
+            // [2117, 548.00, 548.00, 548.00],
+            // [2118, 548.00, 548.00, 548.00],
+            // [2119, 549.00, 549.00, 549.00],
+            // [2120, 549.00, 549.00, 549.00],
+            // [2121, 549.00, 549.00, 549.00],
+            // [2122, 550.00, 550.00, 550.00],
+            // [2123, 550.00, 550.00, 550.00],
+            // [2124, 550.00, 550.00, 550.00],
+            // [2125, 551.00, 551.00, 551.00],
+            // [2126, 551.00, 551.00, 551.00],
+            // [2127, 551.00, 551.00, 551.00],
+            // [2128, 552.00, 552.00, 552.00],
+            // [2129, 552.00, 552.00, 552.00],
+            // [2130, 552.00, 552.00, 552.00],
+            // [2131, 553.00, 553.00, 553.00],
+            // [2132, 553.00, 553.00, 553.00],
+            // [2133, 553.00, 553.00, 553.00],
+            // [2134, 554.00, 554.00, 554.00],
+            // [2135, 554.00, 554.00, 554.00],
+            // [2136, 555.00, 555.00, 555.00],
+            // [2137, 555.00, 555.00, 555.00],
+            // [2138, 555.00, 555.00, 555.00],
+            // [2139, 556.00, 556.00, 556.00],
+            // [2140, 556.00, 556.00, 556.00],
+            // [2141, 556.00, 556.00, 556.00],
+            // [2142, 557.00, 557.00, 557.00],
+            // [2143, 557.00, 557.00, 557.00],
+            // [2144, 557.00, 557.00, 557.00],
+            // [2145, 558.00, 558.00, 558.00],
+            // [2146, 558.00, 558.00, 558.00],
+            // [2147, 558.00, 558.00, 558.00],
+            // [2148, 559.00, 559.00, 559.00],
+            // [2149, 559.00, 559.00, 559.00],
+            // [2150, 559.00, 559.00, 559.00],
+            // [2151, 560.00, 560.00, 560.00],
+            // [2152, 560.00, 560.00, 560.00],
+            // [2153, 560.00, 560.00, 560.00],
+            // [2154, 561.00, 561.00, 561.00],
+            // [2155, 561.00, 561.00, 561.00],
+            // [2156, 561.00, 561.00, 561.00],
+            // [2157, 562.00, 562.00, 562.00],
+            // [2158, 562.00, 562.00, 562.00],
+            // [2159, 562.00, 562.00, 562.00],
+            // [2160, 563.00, 563.00, 563.00],
+            // [2161, 563.00, 563.00, 563.00],
+            // [2162, 563.00, 563.00, 563.00],
+            // [2163, 564.00, 564.00, 564.00],
+            // [2164, 564.00, 564.00, 564.00],
+            // [2165, 565.00, 565.00, 565.00],
+            // [2166, 565.00, 565.00, 565.00],
+            // [2167, 565.00, 565.00, 565.00],
+            // [2168, 566.00, 566.00, 566.00],
+            // [2169, 566.00, 566.00, 566.00],
+            // [2170, 566.00, 566.00, 566.00],
+            // [2171, 567.00, 567.00, 567.00],
+            // [2172, 567.00, 567.00, 567.00],
+            // [2173, 567.00, 567.00, 567.00],
+            // [2174, 568.00, 568.00, 568.00],
+            // [2175, 568.00, 568.00, 568.00],
+            // [2176, 568.00, 568.00, 568.00],
+            // [2177, 569.00, 569.00, 569.00],
+            // [2178, 569.00, 569.00, 569.00],
+            // [2179, 569.00, 569.00, 569.00],
+            // [2180, 570.00, 570.00, 570.00],
+            // [2181, 570.00, 570.00, 570.00],
+            // [2182, 570.00, 570.00, 570.00],
+            // [2183, 571.00, 571.00, 571.00],
+            // [2184, 571.00, 571.00, 571.00],
+            // [2185, 571.00, 571.00, 571.00],
+            // [2186, 572.00, 572.00, 572.00],
+            // [2187, 572.00, 572.00, 572.00],
+            // [2188, 572.00, 572.00, 572.00],
+            // [2189, 573.00, 573.00, 573.00],
+            // [2190, 573.00, 573.00, 573.00],
+            // [2191, 573.00, 573.00, 573.00],
+            // [2192, 574.00, 574.00, 574.00],
+            // [2193, 574.00, 574.00, 574.00],
+            // [2194, 575.00, 575.00, 575.00],
+            // [2195, 575.00, 575.00, 575.00],
+            // [2196, 575.00, 575.00, 575.00],
+            // [2197, 576.00, 576.00, 576.00],
+            // [2198, 576.00, 576.00, 576.00],
+            // [2199, 576.00, 576.00, 576.00],
+            // [2200, 577.00, 577.00, 577.00],
+            // [2201, 577.00, 577.00, 577.00],
+            // [2202, 577.00, 577.00, 577.00],
+            // [2203, 578.00, 578.00, 578.00],
+            // [2204, 578.00, 578.00, 578.00],
+            // [2205, 578.00, 578.00, 578.00],
+            // [2206, 579.00, 579.00, 579.00],
+            // [2207, 579.00, 579.00, 579.00],
+            // [2208, 579.00, 579.00, 579.00],
+            // [2209, 580.00, 580.00, 580.00],
+            // [2210, 580.00, 580.00, 580.00],
+            // [2211, 580.00, 580.00, 580.00],
+            // [2212, 581.00, 581.00, 581.00],
+            // [2213, 581.00, 581.00, 581.00],
+            // [2214, 581.00, 581.00, 581.00],
+            // [2215, 582.00, 582.00, 582.00],
+            // [2216, 582.00, 582.00, 582.00],
+            // [2217, 582.00, 582.00, 582.00],
+            // [2218, 583.00, 583.00, 583.00],
+            // [2219, 583.00, 583.00, 583.00],
+            // [2220, 583.00, 583.00, 583.00],
+            // [2221, 584.00, 584.00, 584.00],
+            // [2222, 584.00, 584.00, 584.00],
+            // [2223, 585.00, 585.00, 585.00],
+            // [2224, 585.00, 585.00, 585.00],
+            // [2225, 585.00, 585.00, 585.00],
+            // [2226, 586.00, 586.00, 586.00],
+            // [2227, 586.00, 586.00, 586.00],
+            // [2228, 586.00, 586.00, 586.00],
+            // [2229, 587.00, 587.00, 587.00],
+            // [2230, 587.00, 587.00, 587.00],
+            // [2231, 587.00, 587.00, 587.00],
+            // [2232, 588.00, 588.00, 588.00],
+            // [2233, 588.00, 588.00, 588.00],
+            // [2234, 588.00, 588.00, 588.00],
+            // [2235, 589.00, 589.00, 589.00],
+            // [2236, 589.00, 589.00, 589.00],
+            // [2237, 589.00, 589.00, 589.00],
+            // [2238, 590.00, 590.00, 590.00],
+            // [2239, 590.00, 590.00, 590.00],
+            // [2240, 590.00, 590.00, 590.00],
+            // [2241, 591.00, 591.00, 591.00],
+            // [2242, 591.00, 591.00, 591.00],
+            // [2243, 591.00, 591.00, 591.00],
+            // [2244, 592.00, 592.00, 592.00],
+            // [2245, 592.00, 592.00, 592.00],
+            // [2246, 592.00, 592.00, 592.00],
+            // [2247, 593.00, 593.00, 593.00],
+            // [2248, 593.00, 593.00, 593.00],
+            // [2249, 593.00, 593.00, 593.00],
+            // [2250, 594.00, 594.00, 594.00],
+            // [2251, 594.00, 594.00, 594.00],
+            // [2252, 595.00, 595.00, 595.00],
+            // [2253, 595.00, 595.00, 595.00],
+            // [2254, 595.00, 595.00, 595.00],
+            // [2255, 596.00, 596.00, 596.00],
+            // [2256, 596.00, 596.00, 596.00],
+            // [2257, 596.00, 596.00, 596.00],
+            // [2258, 597.00, 597.00, 597.00],
+            // [2259, 597.00, 597.00, 597.00],
+            // [2260, 597.00, 597.00, 597.00],
+            // [2261, 598.00, 598.00, 598.00],
+            // [2262, 598.00, 598.00, 598.00],
+            // [2263, 598.00, 598.00, 598.00],
+            // [2264, 599.00, 599.00, 599.00],
+            // [2265, 599.00, 599.00, 599.00],
+            // [2266, 599.00, 599.00, 599.00],
+            // [2267, 600.00, 600.00, 600.00],
+            // [2268, 600.00, 600.00, 600.00],
+            // [2269, 600.00, 600.00, 600.00],
+            // [2270, 601.00, 601.00, 601.00],
+            // [2271, 601.00, 601.00, 601.00],
+            // [2272, 601.00, 601.00, 601.00],
+            // [2273, 602.00, 602.00, 602.00],
+            // [2274, 602.00, 602.00, 602.00],
+            // [2275, 602.00, 602.00, 602.00],
+            // [2276, 603.00, 603.00, 603.00],
+            // [2277, 603.00, 603.00, 603.00],
+            // [2278, 604.00, 604.00, 604.00],
+            // [2279, 604.00, 604.00, 604.00],
+            // [2280, 604.00, 604.00, 604.00],
+            // [2281, 605.00, 605.00, 605.00],
+            // [2282, 605.00, 605.00, 605.00],
+            // [2283, 605.00, 605.00, 605.00],
+            // [2284, 606.00, 606.00, 606.00],
+            // [2285, 606.00, 606.00, 606.00],
+            // [2286, 606.00, 606.00, 606.00],
+            // [2287, 607.00, 607.00, 607.00],
+            // [2288, 607.00, 607.00, 607.00],
+            // [2289, 607.00, 607.00, 607.00],
+            // [2290, 608.00, 608.00, 608.00],
+            // [2291, 608.00, 608.00, 608.00],
+            // [2292, 608.00, 608.00, 608.00],
+            // [2293, 609.00, 609.00, 609.00],
+            // [2294, 609.00, 609.00, 609.00],
+            // [2295, 609.00, 609.00, 609.00],
+            // [2296, 610.00, 610.00, 610.00],
+            // [2297, 610.00, 610.00, 610.00],
+            // [2298, 610.00, 610.00, 610.00],
+            // [2299, 611.00, 611.00, 611.00],
+            // [2300, 611.00, 611.00, 611.00],
+            // [2301, 611.00, 611.00, 611.00],
+            // [2302, 612.00, 612.00, 612.00],
+            // [2303, 612.00, 612.00, 612.00],
+            // [2304, 612.00, 612.00, 612.00],
+            // [2305, 613.00, 613.00, 613.00],
+            // [2306, 613.00, 613.00, 613.00],
+            // [2307, 614.00, 614.00, 614.00],
+            // [2308, 614.00, 614.00, 614.00],
+            // [2309, 614.00, 614.00, 614.00],
+            // [2310, 615.00, 615.00, 615.00],
         ];
     }
 }
